### 2025-10-17 7b455ea
GIL released in CEC functions
