+++ minion/src/arrde.cpp
@@ -51,10 +51,9 @@ void ARRDE::initialize() {
     restartRelTol = reltol;
     refineRelTol = restartRelTol;
 
-    decrease = options.get<double>("refine_decrease_factor", 0.8);
-    startRefine = options.get<double>("restart-refine-duration", 0.8);
+    decrease = options.get<double>("refine_decrease_factor", 0.85);
+    startRefine = options.get<double>("restart-refine-duration", 0.85);
     maxRestart = options.get<int>("maximum_consecutive_restarts", 2);
-    Fw=1.25;
     useLatin = true;
     hasInitialized = true;
 }
@@ -80,10 +79,10 @@ void ARRDE::adaptParameters() {
 void ARRDE::adjustPopulationSize() {
     double nevalsEff = static_cast<double>(Nevals);
     double maxevalsEff = startRefine * static_cast<double>(maxevals);
-    double minSizeEff = std::max(4.0, static_cast<double>(bounds.size()));
+    double minSizeEff = std::max(10.0, static_cast<double>(bounds.size()));
     double maxSizeEff = static_cast<double>(populationSize);
 
-    reduction_strategy = "linear";
+    reduction_strategy = "agsk";
     if (final_refine) {
         nevalsEff = static_cast<double>(Nevals) - static_cast<double>(Neval_stratrefine);
         maxevalsEff = static_cast<double>(maxevals) - static_cast<double>(Neval_stratrefine);
@@ -131,6 +130,9 @@ void ARRDE::adjustArchiveSize() {
     while (archive.size() > targetSize) {
         const size_t randomIndex = rand_int(archive.size());
         archive.erase(archive.begin() + static_cast<std::ptrdiff_t>(randomIndex));
+        if (randomIndex < archive_fitness.size()) {
+            archive_fitness.erase(archive_fitness.begin() + static_cast<std::ptrdiff_t>(randomIndex));
+        }
     }
 }
 
@@ -151,12 +153,38 @@ void ARRDE::processRestartCycle() {
         fitness_records.push_back(value);
     }
 
+    if (!archive.empty()) {
+        if (archive_fitness.size() == archive.size()) {
+            const auto archiveIndices = argsort(archive_fitness, true);
+            for (size_t idx : archiveIndices) {
+                archive_records.push_back(archive[idx]);
+                archive_fitness_records.push_back(archive_fitness[idx]);
+            }
+        } else {
+            const size_t limit = std::min(archive.size(), archive_fitness.size());
+            for (size_t idx = 0; idx < limit; ++idx) {
+                archive_records.push_back(archive[idx]);
+                archive_fitness_records.push_back(archive_fitness[idx]);
+            }
+        }
+
+        if (archive_records.size() > archiveRecordMaxSize) {
+            const size_t excess = archive_records.size() - archiveRecordMaxSize;
+            archive_records.erase(archive_records.begin(),
+                                  archive_records.begin() + static_cast<std::ptrdiff_t>(excess));
+            archive_fitness_records.erase(
+                archive_fitness_records.begin(),
+                archive_fitness_records.begin() + static_cast<std::ptrdiff_t>(excess));
+        }
+    }
+
     update_locals();
 
     const size_t previousPopulationSize = population.size();
 
     fitness_before.clear();
     archive.clear();
+    archive_fitness.clear();
     population.clear();
     fitness.clear();
 
@@ -263,19 +291,45 @@ void ARRDE::executeFinalRefine(size_t /*targetSize*/) {
     const size_t refinedPopulationSize = maxPopSize_finalRefine;
     Neval_stratrefine = Nevals;
 
-    std::vector<size_t> indices;
-    if (fitness_records.size() < refinedPopulationSize) {
[…]
