+++ minion/include/default_options.h
@@ -33,9 +33,9 @@ class DefaultSettings{
         std::map<std::string, ConfigValue>  default_settings_ARRDE  = {
             {"population_size", 0},  
             {"minimum_population_size", 4}, 
-            {"archive_size_ratio", 2.0}, 
+            {"archive_size_ratio", 1.5}, 
             {"converge_reltol", 0.005}, 
-            {"refine_decrease_factor" , 0.75}, 
+            {"refine_decrease_factor" , 0.85}, 
             {"restart-refine-duration", 0.85}, 
             {"maximum_consecutive_restarts" , 2},
             {"bound_strategy" , std::string("reflect-random")} , 
