+++ minion/include/default_options.h
@@ -36,7 +36,7 @@ class DefaultSettings{
             {"archive_size_ratio", 1.5}, 
             {"converge_reltol", 0.005}, 
             {"refine_decrease_factor" , 0.85}, 
-            {"restart-refine-duration", 0.85}, 
+            {"restart-refine-duration", 0.9}, 
             {"maximum_consecutive_restarts" , 2},
             {"bound_strategy" , std::string("reflect-random")} , 
         };
