### 2025-10-26 1739155
Updated ARRDE
