### 2025-10-26 e9f79fe
Updated ARRDE
