### 2025-10-26 b09255d
Updated ARRDE
