+++ minion/src/j2020.cpp
@@ -1,291 +1,286 @@
 #include "j2020.h"
 #include "default_options.h"
 
-#include <algorithm>
-#include <cmath>
-#include <limits>
-#include <numeric>
-
 namespace minion {
 
-void j2020::initialize() {
-    auto defaults = DefaultSettings().getDefaultSettings("j2020");
-    for (const auto& entry : optionMap) {
-        defaults[entry.first] = entry.second;
-    }
-    Options options(defaults);
+void j2020::initialize  (){
+    auto defaultKey = DefaultSettings().getDefaultSettings("j2020");
+    for (auto el : optionMap) defaultKey[el.first] = el.second;
+    Options options(defaultKey);
 
-    boundStrategy = options.get<std::string>("bound_strategy", "reflect-random");
-    std::vector<std::string> supportedBounds = {"random", "reflect", "reflect-random", "clip", "periodic", "none"};
-    if (std::find(supportedBounds.begin(), supportedBounds.end(), boundStrategy) == supportedBounds.end()) {
-        std::cerr << "Bound stategy '" << boundStrategy << "' is not recognized. 'reflect-random' will be used.\n";
+    boundStrategy = options.get<std::string> ("bound_strategy", "reflect-random");
+    std::vector<std::string> all_boundStrategy = {"random", "reflect", "reflect-random", "clip", "periodic", "none"};
+    if (std::find(all_boundStrategy.begin(), all_boundStrategy.end(), boundStrategy)== all_boundStrategy.end()) {
+        std::cerr << "Bound stategy '"+ boundStrategy+"' is not recognized. 'Reflect-random' will be used.\n";
         boundStrategy = "reflect-random";
     }
 
-    desiredPopulation = options.get<int>("population_size", 0);
-    if (desiredPopulation == 0 || desiredPopulation < 32) {
-        size_t dim = bounds.size();
-        desiredPopulation = std::max<size_t>(std::min<size_t>(1000, 8 * dim), 32);
-    }
+    size_t populationSize = options.get<int> ("population_size", 0) ;
+    D = int(bounds.size());
+    populsize = populationSize;
+    if (populsize==0 || populsize<32) populsize = std::max(std::min(1000, 8*D), 32); 
 
-    bigPopulationSize = static_cast<size_t>(std::round(0.875 * static_cast<double>(desiredPopulation)));
-    if (bigPopulationSize == 0) {
-        bigPopulationSize = 1;
-    }
-    smallPopulationSize = desiredPopulation > bigPopulationSize ? desiredPopulation - bigPopulationSize : 1;
-    bigPopulationSize = desiredPopulation - smallPopulationSize;
-    populationSize = bigPopulationSize + smallPopulationSize;
+    tao1 = options.get<double>("tau1", 0.1);  
+    tao2 =  options.get<double>("tau2", 0.1);  
+    myEqs =  options.get<double>("myEqs", 0.25);  
 
-    tau1 = options.get<double>("tau1", 0.1);
-    tau2 = options.get<double>("tau2", 0.1);
-    myEqs = options.get<double>("myEqs", 0.25);
-
-    hasInitialized = true;
+    hasInitialized=true;
 }
 
-double j2020::distance_squared(const std::vector<double>& a, const std::vector<double>& b) const {
-    double acc = 0.0;
-    for (size_t d = 0; d < a.size(); ++d) {
-        double diff = a[d] - b[d];
-        acc += diff * diff;
-    }
-    return acc;
+
+double j2020::Dist(const std::vector<double>& A, const std::vector<double>& B) { 
+    double dist=0.0;
+    for (int j = 0; j<D; j++) 
+        dist += (A[j]-B[j])* (A[j]-B[j]);
[…]
