+++ minion/src/nlshadersp.cpp
@@ -1,468 +1,526 @@
 #include "nlshadersp.h"
 #include "default_options.h"
 
-#include <algorithm>
-#include <cmath>
-#include <cstddef>
-#include <numeric>
-#include <random>
-
 namespace minion {
 
-void NLSHADE_RSP::initialize() {
-    auto defaultKey = DefaultSettings().getDefaultSettings("NLSHADE_RSP");
-    for (const auto& el : optionMap) {
-        defaultKey[el.first] = el.second;
-    }
+void NLSHADE_RSP::initialize  (){
+     auto defaultKey = DefaultSettings().getDefaultSettings("NLSHADE_RSP");
+    for (auto el : optionMap) defaultKey[el.first] = el.second;
     Options options(defaultKey);
 
-    boundStrategy = options.get<std::string>("bound_strategy", "reflect-random");
+    boundStrategy = options.get<std::string> ("bound_strategy", "reflect-random");
     std::vector<std::string> all_boundStrategy = {"random", "reflect", "reflect-random", "clip", "periodic", "none"};
-    if (std::find(all_boundStrategy.begin(), all_boundStrategy.end(), boundStrategy) == all_boundStrategy.end()) {
-        std::cerr << "Bound stategy '" << boundStrategy << "' is not recognized. 'reflect-random' will be used.\n";
+    if (std::find(all_boundStrategy.begin(), all_boundStrategy.end(), boundStrategy)== all_boundStrategy.end()) {
+        std::cerr << "Bound stategy '"+ boundStrategy+"' is not recognized. 'Reflect-random' will be used.\n";
         boundStrategy = "reflect-random";
     }
+    int populsize = options.get<int> ("population_size", 0) ;
+    if (populsize==0) populsize= std::max(int(30*bounds.size()), 10);
 
-    int optPopulation = options.get<int>("population_size", 0);
-    if (optPopulation > 0) {
-        populationSize = static_cast<size_t>(optPopulation);
-    } else {
-        populationSize = std::max(static_cast<size_t>(30 * bounds.size()), static_cast<size_t>(10));
-    }
+    MaxFEval = int(maxevals); 
 
-    memorySize = options.get<int>("memory_size", 100);
-    if (memorySize <= 0) {
-        memorySize = 1;
-    }
-
-    archiveSizeRatio = options.get<double>("archive_size_ratio", 2.6);
-    if (archiveSizeRatio < 0.0) {
-        archiveSizeRatio = 2.6;
-    }
-
-    int optMinPopulation = options.get<int>("minimum_population_size", 4);
-    minPopulationSize = optMinPopulation > 0 ? static_cast<size_t>(optMinPopulation) : static_cast<size_t>(4);
-    minPopulationSize = std::max<size_t>(2, minPopulationSize);
-
-    maxPopulationSize = populationSize;
-    archiveCapacity = static_cast<size_t>(std::round(static_cast<double>(populationSize) * archiveSizeRatio));
-    archiveProbability = 0.5;
-    memoryIndex = 0;
-
-    parameterMemoryCR.assign(static_cast<size_t>(memorySize), 0.2);
-    parameterMemoryF.assign(static_cast<size_t>(memorySize), 0.2);
-    successfulCR.clear();
-    successfulF.clear();
-    successDelta.clear();
-
-    F = std::vector<double>(populationSize, 0.2);
-    CR = std::vector<double>(populationSize, 0.2);
-    p = std::vector<size_t>(populationSize, 1);
-
-    hasInitialized = true;
[…]

[…]
