### 2025-10-27 6390fad
Updated ARRDE
