+++ examples/main_cec.cpp
@@ -116,7 +116,8 @@ int main(int argc, char* argv[]) {
 
     std::vector<int> funcnums; 
     if (year==2017 || year == 2014) funcnums =  {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12,13, 14, 15, 16, 17, 18, 19, 20,  21, 22, 23, 24, 25, 26, 27, 28, 29, 30};
-    else if (year==2020 || year == 2019) funcnums = {1,2,3,4,5,6,7,8,9, 10}; 
+    else if (year==2020) funcnums = {1,2,3,4,5,6,7,8,9, 10}; 
+    else if (year == 2019) funcnums = {7,8,9, 10}; 
     else if (year==2022) funcnums =  {1,2,3,4,5,6,7,8,9, 10, 11, 12}; 
     else throw std::runtime_error("Year invalid.");
 
