+++ minion/include/default_options.h
@@ -32,10 +32,7 @@ class DefaultSettings{
 
         std::map<std::string, ConfigValue>  default_settings_ARRDE  = {
             {"population_size", 0},  
-            {"minimum_population_size", 4}, 
-            {"archive_size_ratio", 1.0}, 
-            {"converge_reltol", 0.0001}, 
-            {"refine_decrease_factor" , 1.0}, 
+            {"minimum_population_size", 4},  
             {"restart-refine-duration", 0.95}, 
             {"maximum_consecutive_restarts" , 2},
             {"bound_strategy" , std::string("reflect-random")} , 
