+++ minion/src/arrde.cpp
@@ -1,6 +1,7 @@
 #include "arrde.h" 
 
 #include <cstddef>
+#include <stdexcept>
 
 namespace minion {
 
@@ -21,7 +22,7 @@ void ARRDE::initialize() {
     const auto dimension = bounds.size();
     const double logComponent = std::pow(std::log10(maxevals), 2.0);
     const double eta = double(maxevals)/double(dimension);
-    const double defaultPopulation    = std::clamp(dimension*(1.0+std::pow(log10(eta), 2.3)), 4.0, 2000.0);
+    const double defaultPopulation    = std::clamp(dimension*(std::pow(log10(eta), 2.5)), 4.0, 2000.0);
     const int configuredPopulation = options.get<int>("population_size", 0);
     if (configuredPopulation > 0) {
         populationSize = static_cast<size_t>(configuredPopulation);
@@ -29,32 +30,22 @@ void ARRDE::initialize() {
         populationSize = static_cast<size_t>(defaultPopulation);
     }
     populationSize = std::max(populationSize, std::max(dimension, static_cast<size_t>(4)));
-    first_run_archive_max_size = populationSize;
-    first_run_archive.clear();
-    first_run_archive_fitness.clear();
-
-    maxPopSize_finalRefine =  std::max(size_t(0.5*populationSize), std::max(dimension, static_cast<size_t>(4)));
-    minPopSize = std::clamp(options.get<int>("minimum_population_size", 4), 4, static_cast<int>(maxPopSize_finalRefine));
+    minPopSize = std::max(options.get<int>("minimum_population_size", 4), 4);
 
     mutation_strategy = "current_to_pbest_AW_1bin";
-    Fw=1.0;
+    Fw=0.7;
     archive_size_ratio = 1.0;
-    if (archive_size_ratio < 0.0) {
-        archive_size_ratio = 2.0;
-    }
+    memorySize = 5;
 
-    memorySizeRatio = archive_size_ratio;
-    memorySize = 5; //std::max<size_t>(1, static_cast<size_t>(memorySizeRatio * static_cast<double>(populationSize)));
-
-    M_CR = rand_gen(0.5, 0.8, memorySize);
-    M_F = rand_gen(0.2, 0.5, memorySize);
+    M_CR = std::vector<double>(memorySize, 0.8) ;
+    M_F =  std::vector<double>(memorySize, 0.3) ;
 
     reltol = 1e-5;
     restartRelTol = reltol;
-    refineRelTol = restartRelTol;
+    refineRelTol = reltol;
 
-    decrease = 1.0; //options.get<double>("refine_decrease_factor", 0.85);
-    startRefine = options.get<double>("restart-refine-duration", 0.85);
+    reduction_strategy = "custom";
+    decrease = 1.0; //std::max(1.0, 1.0/log10(dimension)); //options.get<double>("refine_decrease_factor", 0.85);
     maxRestart = options.get<int>("maximum_consecutive_restarts", 2);
     useLatin = true;
     hasInitialized = true;
@@ -64,54 +55,62 @@ void ARRDE::initialize() {
 void ARRDE::adaptParameters() {
     adjustPopulationSize();
     adjustArchiveSize();
-
-    if (!final_refine && Nevals >= startRefine * maxevals) {
-        init_final_refine = true;
-    }
-
     const double spread = calcStdDev(fitness) / std::fabs(calcMean(fitness));
-    if (spread <= reltol || init_final_refine) {
+    if (spread <= reltol || do_refine) {
+        //std::cout << Nevals << " " << restart << " " << best_fitness << " " << spread << " " << reltol << " " << population.size()<<"\n";
         processRestartCycle();
     }
-
     updateParameterMemory();
[…]
