### 2025-11-04 a94d13d
Updated ARRDE
