### 2025-11-05 f573130
Updated lshade_cnepsin
