+++ minion/include/j2020.h
@@ -1,12 +1,8 @@
-#ifndef j2020_H
-#define j2020_H
+#ifndef J2020_H
+#define J2020_H
 
-#include <math.h>
-#include <iostream>
-#include <time.h>
-#include <fstream>
-#include <random>
-#include "minimizer_base.h"
+#include "de.h"
+#include <limits>
 
 namespace minion {
 
@@ -14,7 +10,7 @@ namespace minion {
  * @file j2020.h
  * @brief Header file for the j2020 class.
  *
- * This file was adapted from the original j2020 code obtained from Suganthan's GitHub repository.
+ * Implementation of the j2020 algorithm.
  * Reference : J. Brest, M. S. Maučec and B. Bošković, "Differential Evolution Algorithm for Single Objective Bound-Constrained Optimization: Algorithm j2020," 2020 IEEE Congress on Evolutionary Computation (CEC), Glasgow, UK, 2020, pp. 1-8, doi: 10.1109/CEC48606.2020.9185551.
  */
 
@@ -26,90 +22,49 @@ namespace minion {
  * for global optimization. It includes methods for initializing the population, 
  * computing distances, crowding mechanisms, and optimizing a given objective function.
  */
-class j2020 :public MinimizerBase{
-private : 
-    size_t populsize;
-    const int myPrec=11;    // setprecision(myPrec) for cout
-    long mySeed;            // 
-    const double eps=1e-12;
-    const double terminateErrorValue=1e-8;  // --> ZERO
-
-    std::vector<std::vector<double>> P;    // population NP x D; NP is population size, i.e., number of individuals
-
-    long nReset;                // reset counter
-    long sReset;                // reset counter
-
-    int D ;
-
-    /* --------- jDE constants ----------- */
-    const double Finit  = 0.5;  // F   INITIAL FACTOR VALUE
-    const double CRinit = 0.9;  // CR  INITIAL FACTOR VALUE
-
-    double Fl = 0.1;         // 
-    const double Fu = 1.1;   //
-
-    double CRl = 0.0;        // 
-    double CRu = 1.0;  //
-
-    double tao1 = 0.1;   // probability to adjust F 
-    double tao2 = 0.1;   // probability to adjust CR
-    double myEqs = 0.25; // for reset populations  CEC2019:0.25
-
-private : 
-    double Dist(const std::vector<double>& A, const std::vector<double>& B);
-    
-    int crowding(std::vector<std::vector<double>> myP, const std::vector<double>& U, const int NP);
-
-    // count how many individuals have similar fitness function value as the best one
-    int stEnakih(const double cost[], const int NP, const double cBest);
-
-    // Are there too many individuals that are equal (very close based on fitness) to the best one 
-    bool prevecEnakih(const std::vector<double>& cost, const int NP, const double cBest);
-
-    void swap(double &a, double &b);
-
-public : 
-
-    /**
-     * @brief Constructor for Differential_Evolution.
-     * @param func The objective function to minimize.
-     * @param bounds The bounds for the variables.
-     * @param x0 The initial guesses for the solution. Note that Minion assumes multiple initial guesses, thus, x0 is an std::vector<std::vector<double>> object. These guesses will be used for population initialization.
-     * @param data Additional data for the objective function.
-     * @param callback Callback function for intermediate results.
-     * @param tol The tolerance for stopping criteria.
-     * @param maxevals The maximum number of evaluations.
-     * @param seed The seed for random number generation.
-     * @param options Option map that specifies further configurational settings for the algorithm.
-     */
+class j2020 : public Differential_Evolution {
+public:
     j2020(
-            MinionFunction func, 
[…]

[…]
