+++ docs/source/Algorithms.rst
@@ -490,33 +490,13 @@ Parameters :
 
                 .. math::
 
-                    N = 2 \cdot D + \log(N_{maxevals})^2
+                    N = D \cdot  \log_{10}(N_{maxevals}/D)^{2.2}
 
                 where *D* is the dimensionality of the problem and :math:`N_{maxevals}` is the maximum number of function evaluations.
 
 - ``minimum_population_size``: 4
 
-  .. note:: final (minimum) population size during linear population size reduction.
-
-- ``archive_size_ratio``: 2.5  
-
-  .. note:: The ratio of archive size to the current population size .
-
-- ``converge_reltol``: 0.005  
-
-  .. note:: The value of std(f)/mean(f) below which a population is said to be converged.
-
-- ``refine_decrease_factor``: 0.8 
-
-  .. note:: The decrease factor of *converge_relTol* in the refinement phase. 
-
-- ``restart-refine-duration``: 0.8  
-
-  .. note:: A fraction of evaluation budget during which restart-refine phase occurs. The remaining fraction is dedicated to final refinement. 
-
-- ``maximum_consecutive_restarts``: 2  
-
-  .. note::  maximum number of consecutive restarts during restart-refine phase.
+  .. note:: final (minimum) population size after population size reduction.
 
 - ``bound_strategy``: ``reflect-random``  
 
@@ -756,7 +736,7 @@ Parameters :
 
                 .. math::
 
-                        N = 5 \cdot D
+                        N = 18 \cdot D
 
                 where *D* is the dimensionality of the problem.
 
