### 2025-11-07 a8d6702
Fixed memory leaks
