+++ minion/include/minimizer_base.h
@@ -249,7 +249,7 @@ class MinimizerBase {
         /**
          * @brief destructor
          */
-        ~MinimizerBase(){};
+        virtual ~MinimizerBase() = default;
 
         /**
          * @brief Virtual function to perform the optimization.
