### 2025-11-07 dad67c3
Fixed further memory leaks
