### 2025-11-07 46ff286
Updated main_cec
