+++ minionpy/__init__.py
@@ -7,7 +7,7 @@ try : from .cec_2011 import CEC2011
 except : pass
 
 # Define the package's version
-__version__ = "0.2.8"
+__version__ = "1.0.1"
 
 # Optionally, define the package name
 __name__ = "minionpy"
