+++ setup.py
@@ -2,7 +2,7 @@ from skbuild import setup
 
 setup(
     name="minionpy",
-    version="0.2.8",
+    version="1.0.1",
     description="MinionPy is the Python implementation of the Minion C++ library, designed for derivative-free optimization.",
     author="Khoirul Faiq Muzakka",
     license="MIT",
