### 2025-11-11 2b68453
Implemented rewrite of CEC2011 1-8
