+++ minion/cec/cec2011.cpp
@@ -1,0 +1,468 @@
+#include "cec2011.h"
+
+#include <algorithm>
+#include <array>
+#include <cmath>
+#include <limits>
+#include <numeric>
+#include <stdexcept>
+#include <vector>
+#include <Eigen/Dense>
+
+namespace minion {
+
+CEC2011Functions::CEC2011Functions(int function_number, int dimension)
+    : CECBase(function_number, dimension) {
+    testfunc = &CEC2011::evaluate;
+}
+
+namespace {
+constexpr double kTwoPi = 2.0 * PI;
+
+constexpr std::array<std::array<double, 4>, 10> kProblem3Coeff = {
+    {{{0.002918487, -0.008045787, 0.006749947, -0.001416647}},
+     {{9.509977, -35.00994, 42.83329, -17.33333}},
+     {{26.82093, -95.56079, 113.0398, -44.29997}},
+     {{208.7241, -719.8052, 827.7466, -316.6655}},
+     {{1.350005, -6.850027, 12.16671, -6.666689}},
+     {{0.01921995, -0.0794532, 0.110566, -0.05033333}},
+     {{0.1323596, -0.469255, 0.5539323, -0.2166664}},
+     {{7.339981, -25.27328, 29.93329, -11.99999}},
+     {{-0.3950534, 1.679353, -1.777829, 0.4974987}},
+     {{-2.504665e-05, 0.01005854, -0.01986696, 0.00983347}}}};
+
+template <size_t N, typename Deriv>
+void rk4Step(std::array<double, N> &state, double t, double dt, Deriv &&deriv) {
+    std::array<double, N> k1{};
+    deriv(t, state, k1);
+
+    std::array<double, N> temp{};
+    for (size_t j = 0; j < N; ++j) {
+        temp[j] = state[j] + 0.5 * dt * k1[j];
+    }
+
+    std::array<double, N> k2{};
+    deriv(t + 0.5 * dt, temp, k2);
+    for (size_t j = 0; j < N; ++j) {
+        temp[j] = state[j] + 0.5 * dt * k2[j];
+    }
+
+    std::array<double, N> k3{};
+    deriv(t + 0.5 * dt, temp, k3);
+    for (size_t j = 0; j < N; ++j) {
+        temp[j] = state[j] + dt * k3[j];
+    }
+
+    std::array<double, N> k4{};
+    deriv(t + dt, temp, k4);
+
+    for (size_t j = 0; j < N; ++j) {
+        state[j] += dt * (k1[j] + 2.0 * k2[j] + 2.0 * k3[j] + k4[j]) / 6.0;
+    }
+}
+
+template <size_t N, typename Deriv>
+std::array<double, N> rk4Integrate(std::array<double, N> state, double t0, double t1,
+                                   int steps, Deriv &&deriv) {
+    const double dt = (t1 - t0) / static_cast<double>(steps);
+    double t = t0;
+    for (int i = 0; i < steps; ++i) {
+        rk4Step(state, t, dt, deriv);
+        t += dt;
+    }
+    return state;
+}
+
+std::array<double, 10> buildProblem3Rates(double u) {
+    const double u2 = u * u;
+    const double u3 = u2 * u;
+    std::array<double, 10> k{};
+    for (size_t row = 0; row < kProblem3Coeff.size(); ++row) {
+        const auto &c = kProblem3Coeff[row];
+        k[row] = c[0] + c[1] * u + c[2] * u2 + c[3] * u3;
+    }
+    return k;
+}
+
+struct TersoffParams {
+    double R1;
+    double R2;
+    double A;
+    double B;
+    double lambda1;
+    double lambda2;
+    double lambda3;
+    double c;
+    double d;
+    double n;
+    double gamma;
[…]
