+++ minionpy/cec_2011.py
@@ -1,494 +1,749 @@
-#!/usr/bin/env python
-                     
 import numpy as np
-import matlab.engine
-import os
-import threading
-from concurrent.futures import ThreadPoolExecutor, as_completed
+
+from minionpycpp import CEC2011Functions as cppCEC2011Functions
 
 
-class Problem01:
-    def __init__(self, path="/cec_input_data/matlab_cec2011/pro_1_to_8"):
-        self.path_file = os.path.dirname(os.path.abspath(__file__)) + path
-        self.eng = matlab.engine.start_matlab()
-        self.eng.addpath(self.path_file, nargout=0)
-        self.ndim = 6
-        self.lb = [-6.4,]*6
-        self.ub = [6.35,]*6
+def _problem02_bounds(ndim: int):
+    lb = np.zeros(ndim)
+    for idx in range(3, ndim):
+        lb[idx] = -4 - 0.25 * int((idx - 4) / 3)
 
-    def evaluate(self, x):
-        return self.eng.bench_func(x, 1)
-
-
-class Problem02:
-    def __init__(self, path="/cec_input_data/matlab_cec2011/pro_1_to_8"):
-        self.path_file = os.path.dirname(os.path.abspath(__file__)) + path
-        self.eng = matlab.engine.start_matlab()
-        self.eng.addpath(self.path_file, nargout=0)
-        self.ndim = 30
-        self.lb = self.get_lb(self.ndim)
-        self.ub = self.get_ub(self.ndim)
-
-    def get_lb(self, ndim):
-        lb = np.zeros(ndim)
-        for idx in range(3, ndim):
-            lb[idx] = -4 - 1.0/4 * int((idx - 4.0) / 3)
-        return lb
-
-    def get_ub(self, ndim):
-        ub = np.zeros(ndim)
-        ub[0] = ub[1] = 4
-        ub[2] = np.pi
-        for idx in range(3, ndim):
-            ub[idx] = 4 + 1.0/4 * int((idx - 4.0) / 3)
-        return ub
-
-    def evaluate(self, x):
-        return self.eng.bench_func(x, 2)
+    ub = np.zeros(ndim)
+    ub[0] = ub[1] = 4.0
+    ub[2] = np.pi
+    for idx in range(3, ndim):
+        ub[idx] = 4 + 0.25 * int((idx - 4) / 3)
+    return lb, ub
 
 
-class Problem03:
-    def __init__(self, path="/cec_input_data/matlab_cec2011/pro_1_to_8"):
-        self.path_file = os.path.dirname(os.path.abspath(__file__)) + path
-        self.eng = matlab.engine.start_matlab()
-        self.eng.addpath(self.path_file, nargout=0)
-        self.ndim = 1
-        self.lb = [-0.6,]
-        self.ub = [0.9,]
-
-    def evaluate(self, x):
-        return self.eng.bench_func(x, 3)
+def _problem05_bounds(ndim: int):
+    lb = -1.0 * np.ones(ndim)
+    lb[:3] = 0.0
+    ub = np.zeros(ndim)
+    ub[0] = ub[1] = 4.0
+    ub[2] = np.pi
+    for idx in range(3, ndim):
+        ub[idx] = 4 + 0.25 * int((idx - 4) / 3)
+    return lb, ub
 
 
-class Problem04:
-    def __init__(self, path="/cec_input_data/matlab_cec2011/pro_1_to_8"):
-        self.path_file = os.path.dirname(os.path.abspath(__file__)) + path
-        self.eng = matlab.engine.start_matlab()
-        self.eng.addpath(self.path_file, nargout=0)
-        self.ndim = 1
-        self.lb = [0.,]
[…]
