+++ minion/cec/cec2011.cpp
@@ -157,8 +157,937 @@ constexpr std::array<LineSpec, 15> kCandidateLines = {{{1, 2, 0.4, 1.0, 40.0},
                                                        {4, 6, 0.30, 1.0, 30.0},
                                                        {5, 6, 0.61, 0.78, 61.0}}};
 
+struct Problem9Line {
+    int from;
+    int to;
+    double reactance;
+    double limit;
+};
+
+constexpr std::array<Problem9Line, 41> kProblem9Lines = {{
+    {1, 2, 0.0575, 130.0},
+    {1, 3, 0.1652, 130.0},
+    {2, 4, 0.1737, 65.0},
+    {3, 4, 0.0379, 130.0},
+    {2, 5, 0.1983, 130.0},
+    {2, 6, 0.1763, 65.0},
+    {4, 6, 0.0414, 90.0},
+    {7, 5, 0.1160, 70.0},
+    {6, 7, 0.0820, 130.0},
+    {6, 8, 0.0420, 32.0},
+    {6, 9, 0.2080, 65.0},
+    {6, 10, 0.5560, 32.0},
+    {11, 9, 0.2080, 65.0},
+    {9, 10, 0.1100, 65.0},
+    {4, 12, 0.2560, 65.0},
+    {13, 12, 0.1400, 65.0},
+    {12, 14, 0.2559, 32.0},
+    {12, 15, 0.1304, 32.0},
+    {12, 16, 0.1987, 32.0},
+    {14, 15, 0.1997, 16.0},
+    {16, 17, 0.1923, 16.0},
+    {15, 18, 0.2185, 16.0},
+    {18, 19, 0.1292, 16.0},
+    {20, 19, 0.0680, 32.0},
+    {10, 20, 0.2090, 32.0},
+    {10, 17, 0.0845, 32.0},
+    {10, 21, 0.0749, 32.0},
+    {10, 22, 0.1499, 32.0},
+    {22, 21, 0.0236, 32.0},
+    {15, 23, 0.2020, 16.0},
+    {22, 24, 0.1790, 16.0},
+    {23, 24, 0.2700, 16.0},
+    {25, 24, 0.3292, 16.0},
+    {25, 26, 0.3800, 16.0},
+    {27, 25, 0.2087, 16.0},
+    {28, 27, 0.3960, 65.0},
+    {27, 29, 0.4153, 16.0},
+    {27, 30, 0.6027, 16.0},
+    {29, 30, 0.4533, 16.0},
+    {8, 28, 0.2000, 32.0},
+    {6, 28, 0.0599, 32.0}
+}};
+
+constexpr std::array<int, 6> kProblem9GeneratorBuses = {1, 2, 5, 8, 11, 13};
+constexpr std::array<int, 21> kProblem9LoadBuses = {2, 3, 4, 5, 7, 8, 10, 12, 14, 15, 16,
+                                                    17, 18, 19, 20, 21, 23, 24, 26, 29, 30};
+constexpr std::array<double, 6> kProblem9Pg = {1.659, 0.491, 0.216, 0.228, 0.124, 0.116};
+constexpr std::array<double, 21> kProblem9Pd = {0.217, 0.024, 0.076, 0.942, 0.228, 0.300, 0.058,
+                                                0.112, 0.062, 0.082, 0.035, 0.090, 0.032, 0.095,
+                                                0.022, 0.175, 0.032, 0.087, 0.035, 0.024, 0.106};
+constexpr std::array<double, 6> kProblem9Rg = {32.7290, 32.1122, 30.3532, 33.6474, 64.1156, 66.8729};
+constexpr std::array<double, 21> kProblem9Rd = {
[…]
