+++ minion/cec/cec2011.cpp
@@ -9,6 +9,10 @@
 #include <vector>
 #include <Eigen/Dense>
 
+#ifdef E
+#undef E
+#endif
+
 namespace minion {
 
 CEC2011Functions::CEC2011Functions(int function_number, int dimension)
@@ -744,6 +748,15 @@ struct HydroThermalProblem {
     bool valvePoint;
 };
 
+enum class MGADSMObjectiveType { OrbitInsertion, TotalDVRendezvous };
+
+struct MGADSMProblem {
+    std::vector<int> sequence;
+    MGADSMObjectiveType objective;
+    double rpTarget;
+    double eTarget;
+};
+
 constexpr int kHydroUnits = 4;
 constexpr int kHydroHours = 24;
 constexpr std::array<double, kHydroHours> kHydroDemand = {1370, 1390, 1360, 1290, 1290, 1410, 1650, 2000, 2240, 2320, 2230, 2310,
@@ -1030,6 +1043,532 @@ double evaluateHydroThermal(const double *x, int nx, const HydroThermalProblem &
                                 weights.reservoir * reservoirPenalty;
     return totalCost + totalPenalty;
 }
+
+namespace mgadsm {
+
+constexpr double kMuSun = 1.32712428e11;
+constexpr double kSecondsPerDay = 86400.0;
+
+inline double clampCos(double value) {
+    return std::max(-1.0, std::min(1.0, value));
+}
+
+inline Eigen::Vector3d unitVector(const Eigen::Vector3d &v) {
+    const double n = v.norm();
+    if (n == 0.0) {
+        return Eigen::Vector3d::Zero();
+    }
+    return v / n;
+}
+
+double m2E(double M, double e);
+double e2M(double E, double e);
+std::array<double, 6> ic2par(const Eigen::Vector3d &r0, const Eigen::Vector3d &v0, double mu);
+double ni2E(double ni, double e);
+std::pair<Eigen::Vector3d, Eigen::Vector3d> par2IC(const std::array<double, 6> &elements, double mu);
+void plephAn(double mjd2000, int planet, Eigen::Vector3d &r, Eigen::Vector3d &v);
+
+Eigen::Vector3d propagatePosition(const Eigen::Vector3d &r0,
+                                  const Eigen::Vector3d &v0,
+                                  double dt,
+                                  double mu,
+                                  Eigen::Vector3d &vOut) {
+    Eigen::Matrix3d DD = Eigen::Matrix3d::Identity();
+    Eigen::Vector3d rAdj = r0;
+    Eigen::Vector3d vAdj = v0;
+    const Eigen::Vector3d h = r0.cross(v0);
+    Eigen::Vector3d ih = unitVector(h);
+    if (std::abs(std::abs(ih.z()) - 1.0) < 1e-3) {
+        DD << 1.0, 0.0, 0.0,
+              0.0, 0.0, 1.0,
+              0.0, -1.0, 0.0;
+        rAdj = DD * rAdj;
+        vAdj = DD * vAdj;
+    }
+    auto elements = ic2par(rAdj, vAdj, mu);
+    const double ecc = elements[1];
+    const double a = elements[0];
+    const double rate = (ecc < 1.0) ? std::sqrt(mu / (a * a * a)) : std::sqrt(-mu / (a * a * a));
+    const double M0 = e2M(elements[5], ecc);
+    const double M = M0 + rate * dt;
+    elements[5] = m2E(M, ecc);
+    auto state = par2IC(elements, mu);
+    Eigen::Vector3d rVec = DD.transpose() * state.first;
+    vOut = DD.transpose() * state.second;
[…]
