### 2025-11-12 74675ef
Removed old cec11 imlementation
