+++ minionpy/cec_2011.py
@@ -1,749 +1,0 @@
-import numpy as np
-
-from minionpycpp import CEC2011Functions as cppCEC2011Functions
-
-
-def _problem02_bounds(ndim: int):
-    lb = np.zeros(ndim)
-    for idx in range(3, ndim):
-        lb[idx] = -4 - 0.25 * int((idx - 4) / 3)
-
-    ub = np.zeros(ndim)
-    ub[0] = ub[1] = 4.0
-    ub[2] = np.pi
-    for idx in range(3, ndim):
-        ub[idx] = 4 + 0.25 * int((idx - 4) / 3)
-    return lb, ub
-
-
-def _problem05_bounds(ndim: int):
-    lb = -1.0 * np.ones(ndim)
-    lb[:3] = 0.0
-    ub = np.zeros(ndim)
-    ub[0] = ub[1] = 4.0
-    ub[2] = np.pi
-    for idx in range(3, ndim):
-        ub[idx] = 4 + 0.25 * int((idx - 4) / 3)
-    return lb, ub
-
-
-def _problem06_bounds(ndim: int):
-    lb = -1.0 * np.ones(ndim)
-    lb[:3] = 0.0
-    ub = np.zeros(ndim)
-    ub[0] = ub[1] = 4.0
-    ub[2] = np.pi
-    for idx in range(3, ndim, 3):
-        ub[idx] = 4 + 0.25 * int((1 - 4) / 3)
-        if idx + 1 < ndim:
-            ub[idx + 1] = 4 + 0.25 * int((2 - 4) / 3)
-        if idx + 2 < ndim:
-            ub[idx + 2] = 4 + 0.25 * int((3 - 4) / 3)
-    return lb, ub
-
-
-PROBLEM_METADATA = {
-    1: (6, np.full(6, -6.4), np.full(6, 6.35)),
-    2: (30, *_problem02_bounds(30)),
-    3: (1, np.array([-0.6]), np.array([0.9])),
-    4: (1, np.array([0.0]), np.array([5.0])),
-    5: (30, *_problem05_bounds(30)),
-    6: (30, *_problem06_bounds(30)),
-    7: (20, np.zeros(20), 2 * np.pi * np.ones(20)),
-    8: (7, np.zeros(7), 15.0 * np.ones(7)),
-    9: (
-        126,
-        np.zeros(126),
-        np.array(
-            [
-                0.217,
-                0.024,
-                0.076,
-                0.892,
-                0.128,
-                0.25,
-                0.058,
-                0.112,
-                0.062,
-                0.082,
-                0.035,
-                0.09,
-                0.032,
-                0.095,
-                0.022,
-                0.175,
-                0.032,
-                0.087,
-                0.035,
-                0.024,
-                0.106,
-                0.217,
-                0.024,
-                0.026,
-                0.491,
-                0.228,
-                0.3,
-                0.058,
-                0.112,
-                0.062,
-                0.082,
-                0.035,
-                0.09,
-                0.032,
-                0.095,
-                0.022,
-                0.175,
-                0.032,
-                0.087,
-                0.035,
-                0.024,
-                0.106,
-                0.216,
-                0.024,
-                0.076,
-                0.216,
-                0.216,
-                0.216,
-                0.058,
-                0.112,
-                0.062,
-                0.082,
-                0.035,
-                0.09,
-                0.032,
-                0.095,
-                0.022,
-                0.175,
-                0.032,
-                0.087,
-                0.035,
-                0.024,
-                0.081,
-                0.217,
-                0.024,
-                0.076,
-                0.228,
-                0.228,
-                0.228,
-                0.058,
-                0.112,
-                0.062,
-                0.082,
-                0.035,
-                0.09,
-                0.032,
-                0.095,
[…]
