+++ minionpy/__init__.py
@@ -2,7 +2,7 @@ from .algorithms import *
 from .test_functions import *
 from .cec import *
 from .parallel import *
-from .cec_2011 import CEC2011
+from .cec import CEC2011Functions as CEC2011
 
 # Define the package's version
 __version__ = "0.2.7"

[…]
