### 2025-11-13 129bcf7
Implemented IMODE and AGSK
