+++ minion/include/minimizer.h
@@ -9,6 +9,8 @@
 #include "lsrtde.h"
 #include "nelder_mead.h"
 #include "lshade.h"
+#include "agsk.h"
+#include "imode.h"
 #include "arrde.h"
 #include "de.h"
 #include "gwo_de.h"
@@ -50,7 +52,7 @@ class Minimizer {
          * @param x0 The initial guesses for the solution. Note that Minion assumes multiple initial guesses, thus, x0 is an std::vector<std::vector<double>> object. These guesses will be used for population initialization in the population-population based algorithms, or minion will pick teh best one in L-BFGS or NelderMead.
          * @param data Additional data to pass to the objective function.
          * @param callback A callback function to call after each iteration.
-         * @param algo Algorithm to use : "LSHADE", "DE", "JADE", "jSO", "DE", "NelderMead", "LSRTDE", "NLSHADE_RSP", "j2020", "GWO_DE", "PSO", "SPSO2011", "DMSPSO", "LSHADE_cnEpSin"
+         * @param algo Algorithm to use : "LSHADE", "AGSK", "DE", "JADE", "jSO", "IMODE", "NelderMead", "LSRTDE", "NLSHADE_RSP", "j2020", "GWO_DE", "PSO", "SPSO2011", "DMSPSO", "LSHADE_cnEpSin"
          * @param relTol The relative tolerance for convergence.
          * @param maxevals The maximum number of function evaluations.
          * @param seed global seed
@@ -72,11 +74,13 @@ class Minimizer {
 
             if (algoUpper == "DE") optimizer = new Differential_Evolution(func, bounds, x0, data, callback, tol, maxevals, seed, options);
             else if (algoUpper == "LSHADE") optimizer = new LSHADE(func, bounds, x0, data, callback, tol, maxevals, seed, options);
+            else if (algoUpper == "AGSK") optimizer = new AGSK(func, bounds, x0, data, callback, tol, maxevals, seed, options);
             else if (algoUpper == "JADE") optimizer = new JADE(func, bounds, x0, data, callback, tol, maxevals, seed, options);
             else if (algoUpper == "J2020") optimizer = new j2020(func, bounds, x0, data, callback, tol, maxevals, seed, options);
             else if (algoUpper == "NLSHADE_RSP") optimizer = new NLSHADE_RSP(func, bounds, x0, data, callback, tol, maxevals, seed, options);
             else if (algoUpper == "LSRTDE") optimizer = new LSRTDE(func, bounds, x0, data, callback, tol, maxevals, seed, options);
             else if (algoUpper == "JSO") optimizer = new jSO (func, bounds, x0, data, callback, tol, maxevals, seed, options);
+            else if (algoUpper == "IMODE") optimizer = new IMODE(func, bounds, x0, data, callback, tol, maxevals, seed, options);
             else if (algoUpper == "ARRDE") optimizer = new ARRDE (func, bounds, x0, data, callback, tol, maxevals, seed, options);
             else if (algoUpper == "GWO_DE") optimizer = new GWO_DE(func, bounds, x0, data, callback, tol, maxevals, seed, options);
             else if (algoUpper == "NELDERMEAD" || algoUpper == "NELDER_MEAD") optimizer = new NelderMead(func, bounds, x0, data, callback, tol, maxevals, seed, options);
@@ -123,11 +127,13 @@ class Minimizer {
 
             if (algoUpper == "DE") optimizer = new Differential_Evolution(func, bounds, {x0}, data, callback, tol, maxevals, seed, options);
             else if (algoUpper == "LSHADE") optimizer = new LSHADE(func, bounds, {x0}, data, callback, tol, maxevals, seed, options);
[…]
