+++ minion/include/minion.h
@@ -11,6 +11,7 @@
 #include "cec2019.h"
 #include "cec2014.h"
 #include "lshade.h"
+#include "agsk.h"
 #include "arrde.h"
 #include "nlshadersp.h"
 #include "cec.h"
@@ -21,6 +22,7 @@
 #include "nelder_mead.h"
 #include "default_options.h"
 #include "minimizer.h"
+#include "imode.h"
 #include "abc.h"
 #include "dual_annealing.h"
 #include "l_bfgs_b.h"

[…]
