### 2025-11-18 666b1ee
Added CEC comparison scripts
