+++ comparison_scripts/compare.py
@@ -1,0 +1,1638 @@
+import os
+import numpy as np
+import pandas as pd
+from scipy.stats import mannwhitneyu, rankdata
+import math
+import matplotlib.pyplot as plt
+
+
+try:
+    from IPython.display import display  # type: ignore
+except ImportError:  # pragma: no cover - CLI fallback
+    def display(obj):
+        print(obj)
+
+def MWUT(A, B, alpha=0.05, mode="ranksum"):
+    """
+    Perform the Mann-Whitney U Test (Wilcoxon rank-sum test) to compare two independent samples.
+
+    The function assumes that *smaller values are better* (e.g., errors),
+    and returns:
+        1  if A is significantly better (lower) than B,
+       -1  if A is significantly worse (higher) than B,
+        0  if there is no statistically significant difference.
+
+    Parameters
+    ----------
+    A : array-like
+        First independent sample.
+    B : array-like
+        Second independent sample.
+    alpha : float, optional (default=0.05)
+        Significance level.
+    mode : {"ranksum", "mean", "median"}, optional
+        Criterion to decide which side is "better" once significance is established:
+        - "ranksum": compare rank sums (nonparametric, consistent with MWU)
+        - "mean":    compare means
+        - "median":  compare medians
+
+    Returns
+    -------
+    int
+        1  if A is significantly lower than B,
+       -1  if A is significantly higher than B,
+        0  if no significant difference (p >= alpha) or exact tie by chosen mode.
+    """
+    A = np.asarray(A, dtype=float)
+    B = np.asarray(B, dtype=float)
+
+    if A.ndim != 1 or B.ndim != 1:
+        raise ValueError("A and B must be 1D arrays.")
+
+    # Mann-Whitney U test (two-sided)
+    # By SciPy's convention, 'statistic' is U for the first sample (A).
+    statistic, p_value = mannwhitneyu(A, B, alternative='two-sided')
+
+    # If not significant, return 0 (cannot reject H0)
+    if p_value >= alpha:
+        return 0
+
+    # At this point, we know the distributions differ significantly.
+    # Now decide direction using the requested 'mode'.
+    if mode == "ranksum":
+        nA = len(A)
+        nB = len(B)
+        U_A = statistic
+        assert (nA == nB)
+
+        # Rank-sum for A: R_A = U_A + nA(nA + 1)/2
+        rank_sum_A = U_A + nA * (nA + 1) / 2.0
+
+        # Total rank sum for all observations 1..(nA+nB)
+        total_rank_sum = (nA + nB) * (nA + nB + 1) / 2.0
+
+        # Rank-sum for B is what's left
+        rank_sum_B = total_rank_sum - rank_sum_A
+
+        if rank_sum_A < rank_sum_B:
+            return 1    # A has lower ranks → better
+        elif rank_sum_A > rank_sum_B:
+            return -1   # A has higher ranks → worse
+        else:
+            return 0    # extremely rare: exact tie in rank sums
+
+    elif mode == "mean":
+        meanA = np.mean(A)
+        meanB = np.mean(B)
+
+        if meanA < meanB:
+            return 1
+        elif meanA > meanB:
+            return -1
+        else:
+            return 0
+
+    elif mode == "median":
+        medianA = np.median(A)
+        medianB = np.median(B)
+
+        if medianA < medianB:
+            return 1
+        elif medianA > medianB:
+            return -1
+        else:
+            return 0
+
+    else:
+        raise Exception("Unknown mode: {}".format(mode))
+
+
+def calcRank(arrDict, ideal, mode="ranksum"):
+    """
+    Rank algorithms by aggregating run-level absolute errors against the ideal.
[…]
