### 2025-11-20 44adc77
Updated script
