+++ setup.py
@@ -2,7 +2,7 @@ from skbuild import setup
 
 setup(
     name="minionpy",
-    version="1.0.1",
+    version="1.1.0",
     description="MinionPy is the Python implementation of the Minion C++ library, designed for derivative-free optimization.",
     author="Khoirul Faiq Muzakka",
     license="MIT",
