+++ docs/source/changelog.rst
@@ -5,6 +5,14 @@ Changelog
    :local:
    :depth: 2
 
+Version 1.1.0 - 2025-11-23
+--------------------------
+
+**Added**
+
+- Added IMODE and AGSK
+- CEC2011 is now completely rewritten in C++. No MATLAB!
+
 Version 1.0.1 - 2025-11-07
 --------------------------
 
