### 2026-01-05 #9229 @khoirulmuzakka
@gkthiruvathukal I would suggest the following names: [Sander van Rijn](https://reviewers.joss.theoj.org/reviewers/345), [Francisco José Monaco](https://reviewers.joss.theoj.org/reviewers/4276), [Camilo Chacón Sartori](https://reviewers.joss.theoj.org/reviewers/5721), and [Alessandro Pierro](https://reviewers.joss.theoj.org/reviewers/3342).
