+++ minion/include/lsrtde.h
@@ -1,15 +1,20 @@
 #ifndef LSRTDE_H
 #define LSRTDE_H
 
-#include "de.h"
+#include <math.h>
+#include <iostream>
+#include <time.h>
+#include <fstream>
+#include <random>
+#include "minimizer_base.h"
 
 namespace minion {
 
 /**
- * @file lsrtde.h
+ * @file nlshader_sp.h
  * @brief Header file for the LSRTDE class, which implements a minimization algorithm.
  * 
- * Implementation of the LSRTDE algorithm.
+ * This code is adapted from the original LSRTDE code from Suganthan's GitHub repository.
  * Reference : V. Stanovov and E. Semenkin, "Success Rate-based Adaptive Differential Evolution L-SRTDE for CEC 2024 Competition," 2024 IEEE Congress on Evolutionary Computation (CEC), Yokohama, Japan, 2024, pp. 1-8, doi: 10.1109/CEC60901.2024.10611907.
  */
 
@@ -22,42 +27,180 @@ namespace minion {
  * optimization using a modified version of the L-SHADE algorithm with an archive and a
  * memory mechanism to store previous successes.
  */
-class LSRTDE : public Differential_Evolution {
-public:
+class LSRTDE : public MinimizerBase{
+private:
+    int MemorySize;
+    int MemoryIter;
+    int SuccessFilled;
+    int MemoryCurrentIndex;
+    int NVars;			    
+    int NIndsCurrent;
+    int NIndsFront;
+    int NIndsFrontMax;
+    int newNIndsFront;
+    int PopulSize;
+    int func_num;
+    int func_index;
+    int TheChosenOne;
+    int Generation;
+    int PFIndex;
+
+    double bestfit;
+    double SuccessRate;
+    double F;      
+    double Cr;
+
+    std::vector<std::vector<double>> Popul;	        
+    std::vector<std::vector<double>>  PopulFront;
+    std::vector<std::vector<double>>  PopulTemp;
+    std::vector<double> FitArr;	
+    std::vector<double> FitArrCopy;
+    std::vector<double> FitArrFront;
+    std::vector<double> Trial;
+    std::vector<double> tempSuccessCr;
+    std::vector<double> MemoryCr;
+    std::vector<double> FitDelta;
+    std::vector<double> Weights;
+
+    int* Indices;
+    int* Indices2;
+
+    int LastFEcount=0;
+    int NFEval = 0;
+    int MaxFEval = 0;
+    int GNVars;
+    double tempF[1];
+    double fopt;
+    char buffer[500];
+    double globalbest;
+    bool globalbestinit = false;
+    bool TimeComplexity = true;
+
+private : 
+    /**
+     * @brief Initialize the LSRTDE algorithm parameters.
+     * 
+     * This function initializes the population, memory, and other parameters
+     * required for the optimization process.
+     * 
+     * @param newNInds Number of individuals in the initial population.
+     * @param newNVars Number of variables in the optimization problem.
+     * @param NewMemSize Size of the memory for storing previous successes.
+     * @param NewArchSizeParam Parameter for determining the archive size.
+     */
+    void initialize_population(int newNInds, int newNVars);
+
+    /**
+     * @brief Clean up allocated memory.
+     * 
+     * This function deallocates any memory used by the class to prevent memory leaks.
+     */
+    void Clean();
+
+    /**
+     * @brief Main optimization cycle of the algorithm.
+     * 
+     * This function performs the main loop of the optimization algorithm, iteratively
+     * evolving the population and updating memory and archives.
+     */
+    void MainCycle();
+
+    /**
+     * @brief Find and save the best individual.
+     * 
+     * This function identifies the best individual in the current population and
+     * saves its information.
+     * 
[…]
