+++ minion/src/algotest.cpp
@@ -1,0 +1,218 @@
+#include "algotest.h"
+
+namespace minion {
+
+void NJADE::initialize() {
+    auto defaultKey = DefaultSettings().getDefaultSettings("NJADE");
+    for (auto el : optionMap) defaultKey[el.first] = el.second;
+    Options options(defaultKey);
+
+    boundStrategy = options.get<std::string>("bound_strategy", "reflect-random");
+    std::vector<std::string> all_boundStrategy = {"random", "reflect", "reflect-random", "clip", "periodic", "none"};
+    if (std::find(all_boundStrategy.begin(), all_boundStrategy.end(), boundStrategy) == all_boundStrategy.end()) {
+        std::cerr << "Bound stategy '" + boundStrategy + "' is not recognized. 'Reflect-random' will be used.\n";
+        boundStrategy = "reflect-random";
+    }
+
+    populationSize = options.get<int>("population_size", 0);
+    if (populationSize == 0) populationSize = 20 * bounds.size();
+    if (populationSize == minPopSize) popreduce = false;
+    else popreduce = true;
+
+    memorySize = static_cast<size_t>(options.get<int>("memory_size", 5));
+    if (memorySize < 1) memorySize = 1;
+    M_CR = std::vector<double>(memorySize, 0.8);
+    M_F = std::vector<double>(memorySize, 0.5);
+    memoryIndex = 0;
+    mutation_strategy = "current_to_pbest_A_1bin";
+    p = std::vector<size_t>(populationSize, 2);
+    hasInitialized = true;
+}
+
+void NJADE::adaptParameters() {
+    if (popreduce) {
+        const double progress = double(Nevals) / double(maxevals);
+        const double A = double(populationSize);
+        const double C = 4; //std::max(4.0, 0.5 * double(bounds.size()));
+        const double dim = double(bounds.size());
+        double pp = 1; //2 ;// 1.17 + 2.075 * exp(-0.0567 * dim);
+        double value;
+        const double t = progress ;
+        value = A - (A - C) * (1.0 - std::pow(1.0 - t, pp));
+        size_t new_population_size = static_cast<size_t>(std::round(value));
+        new_population_size = std::max(new_population_size, size_t(4));
+
+        if (population.size() > new_population_size) {
+            std::vector<size_t> sorted_index = argsort(fitness, true);
+            std::vector<std::vector<double>> new_population_subset(new_population_size);
+            std::vector<double> new_fitness_subset(new_population_size);
+            for (int i = 0; i < new_population_size; ++i) {
+                new_population_subset[i] = population[sorted_index[i]];
+                new_fitness_subset[i] = fitness[sorted_index[i]];
+            }
+            population = new_population_subset;
+            fitness = new_fitness_subset;
+        }
+    }
+
+    std::vector<double> S_CR, S_F, dif_fitness;
+    if (!fitness_before.empty()) {
+        for (size_t i = 0; i < population.size(); ++i) {
+            if (trial_fitness[i] < fitness_before[i]) {
+                double w = std::abs(fitness_before[i] - trial_fitness[i]);
+                S_CR.push_back(CR[i]);
+                S_F.push_back(F[i]);
+                dif_fitness.push_back(w);
+            }
+        }
+    }
+
+    if (!S_CR.empty()) {
+        double sum = 0.0;
+        for (size_t i = 0; i < dif_fitness.size(); ++i) {
+            sum += dif_fitness[i];
+        }
+
+        double temp_sum_sf = 0.0;
+        double temp_sum_cr = 0.0;
+        double meanF_lehmer = 0.0;
+        double meanCR_lehmer = 0.0;
+
+        for (size_t i = 0; i < S_F.size(); ++i) {
[…]
