### 2026-01-15 b8cb3b9
Updated compile.sh
