+++ compile.sh
@@ -6,6 +6,17 @@ set -e
 # Navigate to the project directory
 cd "$(dirname "$0")"
 
+# Build type selection
+build_type="Release"
+if [[ "$1" == "--debug" ]]; then
+  build_type="Debug"
+elif [[ "$1" == "--release" ]]; then
+  build_type="Release"
+elif [[ -n "$1" ]]; then
+  echo "Usage: $0 [--debug|--release]"
+  exit 1
+fi
+
 # Cleanup previous builds
 echo "Cleaning up old builds..."
 #rm -rf dist build *.egg-info
@@ -16,41 +27,18 @@ mkdir -p build && cd build
 
 # Configure with CMake
 echo "Configuring with CMake..."
-cmake -G "Unix Makefiles" ..
+if [[ "${build_type}" == "Debug" ]]; then
+  cmake -G "Unix Makefiles" \
+    -DCMAKE_BUILD_TYPE=${build_type} \
+    -DCMAKE_CXX_FLAGS_DEBUG="-O0 -g3 -fno-omit-frame-pointer" \
+    -DCMAKE_C_FLAGS_DEBUG="-O0 -g3 -fno-omit-frame-pointer" \
+    ..
+else
+  cmake -G "Unix Makefiles" -DCMAKE_BUILD_TYPE=${build_type} ..
+fi
 
 # Compile with optimal parallelization
 echo "Compiling Minion..."
-cmake --build .  --config Release -- -j$(nproc)
-
+cmake --build .  --config ${build_type} -- -j$(nproc)
 # Move back to root
 cd ..           
-
-# Run Doxygen if installed
-if command -v doxygen &> /dev/null; then
-    echo "Generating documentation with Doxygen..."
-    doxygen Doxyfile
-else
-    echo "Warning: Doxygen not found. Skipping documentation generation."
-fi
-
-# Build HTML documentation
-if [ -d "docs" ]; then
-    echo "Building HTML documentation..."
-    cd docs 
-    make clean 
-    make html
-    cd ..
-else
-    echo "Warning: 'docs' folder not found. Skipping HTML documentation."
-fi
-
-echo "Build complete!"
-
-
-#echo "building pip package..."
-#python -m build
-
-#echo "installing minionpy ... "
-#pip uninstall minionpy
-#pip install dist/minionpy-0.1-py3-none-any.whl
-#echo "done"
