+++ minion/include/default_options.h
@@ -194,6 +194,14 @@ class DefaultSettings{
             {"bound_strategy", std::string("reflect-random")}
         };
 
+        std::map<std::string, ConfigValue> default_settings_ACMAES = {
+            {"population_size", 0},
+            {"mu", 0},
+            {"max_iterations", 5000},
+            {"initial_step", 0.3},
+            {"bound_strategy", std::string("reflect-random")}
+        };
+
         std::map<std::string, ConfigValue>  default_settings_DA= {
             {"acceptance_par", -5.0},  
             {"visit_par", 2.67},  
@@ -258,6 +266,7 @@ class DefaultSettings{
                 {"LSHADE_cnEpSin", default_settings_LSHADE_cnEpSin}, 
                 {"CMAES", default_settings_CMAES}, 
                 {"BIPOP_aCMAES", default_settings_BIPOP_aCMAES}, 
+                {"ACMAES", default_settings_ACMAES},
                 {"DA", default_settings_DA}, 
                 {"L_BFGS_B", default_settings_LBFGSB}, 
                 {"L_BFGS", default_settings_LBFGS}
