+++ minion/include/acmaes.h
@@ -71,7 +71,14 @@ private:
         bool h_sig = false;
 
         void reserve(size_t n_offsprings_reserve_, size_t n_parents_reserve_, size_t n_params_);
-        void reinit(size_t n_offsprings_, size_t n_parents_, size_t n_params_, const Eigen::VectorXd& x_mean_, double sigma_);
+        void reinit(
+            size_t n_offsprings_,
+            size_t n_parents_,
+            size_t n_params_,
+            const Eigen::VectorXd& x_mean_,
+            double sigma_,
+            size_t nevals,
+            double best_fitness);
     };
 
     std::vector<double> applyBounds(const std::vector<double>& candidate) const;
@@ -115,6 +122,10 @@ private:
 
     bool support_tol = true;
     bool should_stop = false;
+
+    size_t no_improve_generations = 0;
+    size_t restart_no_improve_tol = 1000;
+    size_t no_improve_restarts = 0;
 };
 
 }
