+++ minion/src/acmaes.cpp
@@ -5,7 +5,6 @@
 
 #include <algorithm>
 #include <cmath>
-#include <iostream>
 #include <limits>
 #include <numeric>
 
@@ -49,14 +48,7 @@ void ACMAES::Parameter::reserve(size_t n_offsprings_reserve_, size_t n_parents_r
     keys_offsprings.resize(n_offsprings_reserve);
 }
 
-void ACMAES::Parameter::reinit(
-    size_t n_offsprings_,
-    size_t n_parents_,
-    size_t n_params_,
-    const Eigen::VectorXd& x_mean_,
-    double sigma_,
-    size_t nevals,
-    double best_fitness) {
+void ACMAES::Parameter::reinit(size_t n_offsprings_, size_t n_parents_, size_t n_params_, const Eigen::VectorXd& x_mean_, double sigma_) {
     n_params = n_params_;
     n_offsprings = std::min(n_offsprings_, n_offsprings_reserve);
     n_parents = std::min(n_parents_, n_parents_reserve);
@@ -125,63 +117,6 @@ void ACMAES::Parameter::reinit(
     D.setIdentity();
     eigvals_C.setOnes();
     f_offsprings.head(static_cast<Eigen::Index>(n_offsprings)) = Eigen::VectorXd::Constant(static_cast<Eigen::Index>(n_offsprings), std::numeric_limits<double>::infinity());
-    //std::cout << "Nevals : " << nevals << "  ACMAES initialized with " << n_offsprings << "Best fitness: " << best_fitness << "\n";
-}
-
-void ACMAES::Parameter::resize(size_t n_offsprings_, size_t n_parents_, size_t n_params_) {
-    n_params = n_params_;
-    n_offsprings = std::min(n_offsprings_, n_offsprings_reserve);
-    n_parents = std::min(n_parents_, n_parents_reserve);
-
-    double w_neg_sum = 0.0;
-    double w_pos_sum = 0.0;
-    for (size_t i = 0; i < n_offsprings; ++i) {
-        double wi = std::log((static_cast<double>(n_offsprings) + 1.0) / 2.0) - std::log(static_cast<double>(i) + 1.0);
-        w(static_cast<Eigen::Index>(i)) = wi;
-        if (wi >= 0.0) {
-            w_pos_sum += wi;
-        } else {
-            w_neg_sum += wi;
-        }
-    }
-
-    double w_sum_parent = 0.0;
-    double w_sq_sum_parent = 0.0;
-    for (size_t i = 0; i < n_parents; ++i) {
-        double wi = w(static_cast<Eigen::Index>(i));
-        w_sum_parent += wi;
-        w_sq_sum_parent += wi * wi;
-    }
-    double denom = w_sq_sum_parent <= 0.0 ? 1e-12 : w_sq_sum_parent;
-    n_mu_eff = (w_sum_parent * w_sum_parent) / denom;
-
-    double dim = static_cast<double>(n_params);
-    c_s = (n_mu_eff + 2.0) / (dim + n_mu_eff + 5.0);
-    c_c = (4.0 + n_mu_eff / dim) / (dim + 4.0 + 2.0 * n_mu_eff / dim);
-    c_1 = 2.0 / (std::pow(dim + 1.3, 2.0) + n_mu_eff);
-    c_mu = 2.0 * (n_mu_eff - 2.0 + 1.0 / n_mu_eff) / (std::pow(dim + 2.0, 2.0) + n_mu_eff);
-    c_mu = std::min(1.0 - c_1, c_mu);
[…]
