+++ minion/src/acmaes.cpp
@@ -119,6 +119,62 @@ void ACMAES::Parameter::reinit(size_t n_offsprings_, size_t n_parents_, size_t n
     f_offsprings.head(static_cast<Eigen::Index>(n_offsprings)) = Eigen::VectorXd::Constant(static_cast<Eigen::Index>(n_offsprings), std::numeric_limits<double>::infinity());
 }
 
+void ACMAES::Parameter::resize(size_t n_offsprings_, size_t n_parents_, size_t n_params_) {
+    n_params = n_params_;
+    n_offsprings = std::min(n_offsprings_, n_offsprings_reserve);
+    n_parents = std::min(n_parents_, n_parents_reserve);
+
+    double w_neg_sum = 0.0;
+    double w_pos_sum = 0.0;
+    for (size_t i = 0; i < n_offsprings; ++i) {
+        double wi = std::log((static_cast<double>(n_offsprings) + 1.0) / 2.0) - std::log(static_cast<double>(i) + 1.0);
+        w(static_cast<Eigen::Index>(i)) = wi;
+        if (wi >= 0.0) {
+            w_pos_sum += wi;
+        } else {
+            w_neg_sum += wi;
+        }
+    }
+
+    double w_sum_parent = 0.0;
+    double w_sq_sum_parent = 0.0;
+    for (size_t i = 0; i < n_parents; ++i) {
+        double wi = w(static_cast<Eigen::Index>(i));
+        w_sum_parent += wi;
+        w_sq_sum_parent += wi * wi;
+    }
+    double denom = w_sq_sum_parent <= 0.0 ? 1e-12 : w_sq_sum_parent;
+    n_mu_eff = (w_sum_parent * w_sum_parent) / denom;
+
+    double dim = static_cast<double>(n_params);
+    c_s = (n_mu_eff + 2.0) / (dim + n_mu_eff + 5.0);
+    c_c = (4.0 + n_mu_eff / dim) / (dim + 4.0 + 2.0 * n_mu_eff / dim);
+    c_1 = 2.0 / (std::pow(dim + 1.3, 2.0) + n_mu_eff);
+    c_mu = 2.0 * (n_mu_eff - 2.0 + 1.0 / n_mu_eff) / (std::pow(dim + 2.0, 2.0) + n_mu_eff);
+    c_mu = std::min(1.0 - c_1, c_mu);
+    d_s = 1.0 + c_s + 2.0 * std::max(0.0, std::sqrt((n_mu_eff - 1.0) / (dim + 1.0)) - 1.0);
+    chi = std::sqrt(dim) * (1.0 - 1.0 / (4.0 * dim) + 1.0 / (21.0 * dim * dim));
+    p_s_fact = std::sqrt(c_s * (2.0 - c_s) * n_mu_eff);
+    p_c_fact = std::sqrt(c_c * (2.0 - c_c) * n_mu_eff);
+
+    double a_mu = 1.0 + c_1 / std::max(1e-12, c_mu);
+    double a_mueff = 1.0 + 2.0 * n_mu_eff;
+    double a_posdef = (1.0 - c_1 - c_mu) / (dim * std::max(1e-12, c_mu));
+    double a_min = std::min({a_mu, a_mueff, a_posdef});
+    for (size_t i = 0; i < n_offsprings; ++i) {
+        double wi = w(static_cast<Eigen::Index>(i));
+        if (wi >= 0.0) {
+            w(static_cast<Eigen::Index>(i)) = wi / std::max(1e-12, w_pos_sum);
+        } else {
+            w(static_cast<Eigen::Index>(i)) = a_min * wi / std::max(1e-12, std::abs(w_neg_sum));
+        }
+    }
+
[…]
