+++ minion/include/default_options.h
@@ -196,7 +196,6 @@ class DefaultSettings{
 
         std::map<std::string, ConfigValue> default_settings_ACMAES = {
             {"population_size", 0},
-            {"mu", 0},
             {"max_iterations", 5000},
             {"initial_step", 0.3},
             {"bound_strategy", std::string("reflect-random")}
