+++ minion/src/rcmaes.cpp
@@ -1,4 +1,4 @@
-#include "acmaes.h"
+#include "rcmaes.h"
 
 #include "default_options.h"
 #include "utility.h"
@@ -10,7 +10,7 @@
 
 namespace minion {
 
-ACMAES::ACMAES(
+RCMAES::RCMAES(
     MinionFunction func,
     const std::vector<std::pair<double, double>>& bounds,
     const std::vector<std::vector<double>>& x0,
@@ -22,7 +22,7 @@ ACMAES::ACMAES(
     std::map<std::string, ConfigValue> options)
     : MinimizerBase(func, bounds, x0, data, callback, tol, maxevals, seed, options) {}
 
-void ACMAES::Parameter::reserve(size_t n_offsprings_reserve_, size_t n_parents_reserve_, size_t n_params_) {
+void RCMAES::Parameter::reserve(size_t n_offsprings_reserve_, size_t n_parents_reserve_, size_t n_params_) {
     n_params = n_params_;
     n_offsprings_reserve = n_offsprings_reserve_;
     n_parents_reserve = n_parents_reserve_;
@@ -48,7 +48,7 @@ void ACMAES::Parameter::reserve(size_t n_offsprings_reserve_, size_t n_parents_r
     keys_offsprings.resize(n_offsprings_reserve);
 }
 
-void ACMAES::Parameter::reinit(
+void RCMAES::Parameter::reinit(
     size_t n_offsprings_,
     size_t n_parents_,
     size_t n_params_,
@@ -127,7 +127,7 @@ void ACMAES::Parameter::reinit(
     //std::cout << "Nevals " << nevals << " Population " << n_offsprings << " Best fitness " << best_fitness << "\n";
 }
 
-void ACMAES::Parameter::resize(size_t n_offsprings_, size_t n_parents_, size_t n_params_) {
+void RCMAES::Parameter::resize(size_t n_offsprings_, size_t n_parents_, size_t n_params_) {
     n_params = n_params_;
     n_offsprings = std::min(n_offsprings_, n_offsprings_reserve);
     n_parents = std::min(n_parents_, n_parents_reserve);
@@ -183,7 +183,7 @@ void ACMAES::Parameter::resize(size_t n_offsprings_, size_t n_parents_, size_t n
     f_offsprings.head(static_cast<Eigen::Index>(n_offsprings)) = Eigen::VectorXd::Constant(static_cast<Eigen::Index>(n_offsprings), std::numeric_limits<double>::infinity());
 }
 
-std::vector<double> ACMAES::applyBounds(const std::vector<double>& candidate) const {
+std::vector<double> RCMAES::applyBounds(const std::vector<double>& candidate) const {
     if (!useBounds) {
         return candidate;
     }
@@ -192,11 +192,11 @@ std::vector<double> ACMAES::applyBounds(const std::vector<double>& candidate) co
     return wrapper.front();
 }
 
-std::vector<double> ACMAES::eigenToStd(const Eigen::VectorXd& vec) const {
+std::vector<double> RCMAES::eigenToStd(const Eigen::VectorXd& vec) const {
     return std::vector<double>(vec.data(), vec.data() + vec.size());
 }
 
-void ACMAES::applyCovarianceScale() {
+void RCMAES::applyCovarianceScale() {
     if (cov_scale.empty()) {
         return;
     }
@@ -213,8 +213,8 @@ void ACMAES::applyCovarianceScale() {
     era.eigvals_C = diag.cwiseProduct(diag);
 }
 
-void ACMAES::initialize() {
-    auto defaults = DefaultSettings().getDefaultSettings("ACMAES");
[…]
