+++ minion/src/bindings.cpp
@@ -30,7 +30,7 @@
 #include "dmspso.h"
 #include "lshadecnepsin.h"
 #include "cmaes.h"
-#include "acmaes.h"
+#include "rcmaes.h"
 #include "dual_annealing.h"
 #include "l_bfgs_b.h"
 #include "minimizer.h"
@@ -374,7 +374,7 @@ PYBIND11_MODULE(minionpycpp, m) {
             py::arg("options") = std::map<std::string, ConfigValue>())
         .def("optimize", &CMAES::optimize, py::call_guard<py::gil_scoped_release>());
 
-    py::class_<ACMAES, MinimizerBase>(m, "ACMAES")
+    py::class_<RCMAES, MinimizerBase>(m, "RCMAES")
         .def(py::init<MinionFunction, const std::vector<std::pair<double, double>>&,
                       const std::vector<std::vector<double>>&, void*, std::function<void(MinionResult*)>,
                       double, size_t, int, std::map<std::string, ConfigValue> >(),
@@ -387,7 +387,7 @@ PYBIND11_MODULE(minionpycpp, m) {
             py::arg("maxevals") = 100000,
             py::arg("seed") = -1,
             py::arg("options") = std::map<std::string, ConfigValue>())
-        .def("optimize", &ACMAES::optimize, py::call_guard<py::gil_scoped_release>());
+        .def("optimize", &RCMAES::optimize, py::call_guard<py::gil_scoped_release>());
 
     py::class_<BIPOP_aCMAES, MinimizerBase>(m, "BIPOP_aCMAES")
         .def(py::init<MinionFunction, const std::vector<std::pair<double, double>>&,
