+++ minion/include/default_options.h
@@ -189,13 +189,14 @@ class DefaultSettings{
         std::map<std::string, ConfigValue> default_settings_BIPOP_aCMAES = {
             {"population_size", 0},
             {"max_restarts", 8},
+            {"max_iterations", 100000},
             {"initial_step", 0.3},
             {"bound_strategy", std::string("reflect-random")}
         };
 
         std::map<std::string, ConfigValue> default_settings_RCMAES = {
             {"population_size", 0},
-            {"initial_step", 0.25},
+            {"initial_step", 0.3},
             {"bound_strategy", std::string("reflect-random")}
         };
 
