+++ minion/src/bipop_acmaes.cpp
@@ -132,6 +132,23 @@ std::vector<double> BIPOP_aCMAES::eigenToStd(const Eigen::VectorXd& vec) const {
     return std::vector<double>(vec.data(), vec.data() + vec.size());
 }
 
+void BIPOP_aCMAES::applyCovarianceScale() {
+    if (cov_scale.empty()) {
+        return;
+    }
+    const size_t dim = cov_scale.size();
+    Eigen::VectorXd diag(dim);
+    for (size_t i = 0; i < dim; ++i) {
+        diag(static_cast<Eigen::Index>(i)) = cov_scale[i];
+    }
+    era.B.setIdentity();
+    era.D = diag.asDiagonal();
+    era.C = era.D * era.D;
+    Eigen::VectorXd inv = diag.cwiseInverse();
+    era.C_invsqrt = inv.asDiagonal();
+    era.eigvals_C = diag.cwiseProduct(diag);
+}
+
 void BIPOP_aCMAES::initialize() {
     auto defaults = DefaultSettings().getDefaultSettings("BIPOP_aCMAES");
     for (const auto& item : optionMap) {
@@ -155,7 +172,6 @@ void BIPOP_aCMAES::initialize() {
     lambda0 = std::max<size_t>(lambda0, 4);
     mu0 = std::max<size_t>(lambda0 / 2, 1);
 
-    maxRestarts = static_cast<size_t>(options.get<int>("max_restarts", 8));
     maxIterations = static_cast<size_t>(options.get<int>("max_iterations", 5000));
     support_tol = true;
 
@@ -165,11 +181,21 @@ void BIPOP_aCMAES::initialize() {
     }
 
     double avgRange = 0.0;
+    cov_scale.clear();
+    cov_scale.reserve(dimension);
     for (const auto& b : bounds) {
-        avgRange += (b.second - b.first);
+        double range = b.second - b.first;
+        avgRange += range;
+        cov_scale.push_back(range);
     }
-    avgRange = dimension > 0 ? avgRange / static_cast<double>(dimension) : 1.0;
-    sigma0 *= avgRange;
+    avg_range = dimension > 0 ? avgRange / static_cast<double>(dimension) : 1.0;
+    sigma0 *= avg_range;
+    if (avg_range > 0.0) {
+        for (double& v : cov_scale) {
+            v /= avg_range;
+            if (v <= 0.0) v = 1.0;
+        }
+    }
 
     initialMean = Eigen::VectorXd::Zero(static_cast<Eigen::Index>(dimension));
     if (!x0.empty()) {
@@ -192,14 +218,7 @@ void BIPOP_aCMAES::initialize() {
         }
     }
 
-    size_t lambda_max = lambda0;
-    if (maxRestarts > 0) {
-        size_t limit = std::min<size_t>(maxRestarts, 20);
-        lambda_max = lambda0 * (1ull << limit);
-    }
-    lambda_max = std::max<size_t>(lambda_max, lambda0);
-    size_t mu_max = std::max<size_t>(lambda_max / 2, 1);
-    era.reserve(lambda_max, mu_max, dimension);
+    era.reserve(lambda0, mu0, dimension);
 
     best = eigenToStd(initialMean);
     best_fitness = std::numeric_limits<double>::infinity();
@@ -401,12 +420,6 @@ void BIPOP_aCMAES::checkStoppingCriteria() {
         return;
     }
 
-    if (best_fitness < 1e-10) {
-        should_stop_run = true;
-        should_stop_optimization = true;
-        return;
-    }
-
     double sigma_fac = sigma0 > 0.0 ? era.sigma / sigma0 : era.sigma;
[…]
