+++ examples/main_cec_multi.cpp
@@ -1,0 +1,444 @@
+#include <iostream>
+#include <vector>
+#include <chrono>
+#include <array>
+#include <unordered_map>
+#include "minion.h"
+#include "utility.h"
+#include "cec2011.h"
+#include <fstream>
+#include <cmath>
+#include <thread>
+#include <algorithm>
+
+namespace {
+
+constexpr double kPi = 3.14159265358979323846;
+constexpr std::array<double, 126> kProblem9UpperBoundsExample = {
+   0.217, 0.024, 0.076, 0.892, 0.128, 0.25, 0.058, 0.112, 0.062, 0.082, 0.035, 0.09, 0.032, 0.095, 0.022, 0.175, 0.032, 0.087, 0.035, 0.024, 0.106,
+    0.217, 0.024, 0.026, 0.491, 0.228, 0.3, 0.058, 0.112, 0.062, 0.082, 0.035, 0.09, 0.032, 0.095, 0.022, 0.175, 0.032, 0.087, 0.035, 0.024, 0.106,
+    0.216, 0.024, 0.076, 0.216, 0.216, 0.216, 0.058, 0.112, 0.062, 0.082, 0.035, 0.09, 0.032, 0.095, 0.022, 0.175, 0.032, 0.087, 0.035, 0.024, 0.081,
+    0.217, 0.024, 0.076, 0.228, 0.228, 0.228, 0.058, 0.112, 0.062, 0.082, 0.035, 0.09, 0.032, 0.095, 0.022, 0.025, 0.032, 0.087, 0.035, 0.024, 0.081,
+    0.124, 0.024, 0.076, 0.124, 0.124, 0.124, 0.058, 0.112, 0.062, 0.082, 0.035, 0.065, 0.032, 0.095, 0.022, 0.124, 0.032, 0.087, 0.035, 0.024, 0.106,
+    0.116, 0.024, 0.076, 0.116, 0.116, 0.116, 0.058, 0.087, 0.062, 0.082, 0.035, 0.09, 0.032, 0.095, 0.022, 0.116, 0.032, 0.087, 0.035, 0.024, 0.106
+};
+
+static constexpr double kELD6Pmin[6] = {100.000000, 50.000000, 80.000000, 50.000000, 50.000000, 50.000000};
+static constexpr double kELD6Pmax[6] = {500.000000, 200.000000, 300.000000, 150.000000, 200.000000, 120.000000};
+
+static constexpr double kELD13Pmin[13] = {0.000000, 0.000000, 0.000000, 60.000000, 60.000000, 60.000000, 60.000000, 60.000000, 60.000000, 40.000000, 40.000000, 55.000000, 55.000000};
+static constexpr double kELD13Pmax[13] = {680.000000, 360.000000, 360.000000, 180.000000, 180.000000, 180.000000, 180.000000, 180.000000, 180.000000, 120.000000, 120.000000, 120.000000, 120.000000};
+
+static constexpr double kELD15Pmin[15] = {150.000000, 150.000000, 20.000000, 20.000000, 150.000000, 135.000000, 135.000000, 60.000000, 25.000000, 25.000000, 20.000000, 20.000000, 25.000000, 15.000000, 15.000000};
+static constexpr double kELD15Pmax[15] = {455.000000, 455.000000, 130.000000, 130.000000, 470.000000, 460.000000, 465.000000, 300.000000, 162.000000, 160.000000, 80.000000, 80.000000, 85.000000, 55.000000, 55.000000};
+
[…]
