+++ CMakeLists.txt
@@ -112,6 +112,15 @@ if(NOT SKBUILD)
         set_target_properties(main_cec_2 PROPERTIES RUNTIME_OUTPUT_DIRECTORY ${CMAKE_SOURCE_DIR}/bin)
     endif()
 
+    #------------------------
+    add_executable(main_cec_multi examples/main_cec_multi.cpp)
+    target_link_libraries(main_cec_multi PRIVATE minion)
+    if (MSVC)
+        set_target_properties(main_cec_multi PROPERTIES RUNTIME_OUTPUT_DIRECTORY ${CMAKE_SOURCE_DIR}/bin/$<0:>)
+    else()
+        set_target_properties(main_cec_multi PROPERTIES RUNTIME_OUTPUT_DIRECTORY ${CMAKE_SOURCE_DIR}/bin)
+    endif()
+
 
     # 2. example/main_minimizer.cpp
     #------------------------
