+++ minionpy.md
@@ -36,6 +36,7 @@ MinionPy is the Python implementation of the Minion C++ library, designed for de
       - Canonical PSO, SPSO-2011, Dynamic Multi-Swarm PSO (DMS-PSO)  
       - CMA-ES *(Covariance Matrix Adaptation Evolution Strategy)*  
       - BIPOP-aCMAES
+      - RCMAES
     - **Classical optimization algorithms:**
       - Nelder-Mead  
       - Generalized Simulated Annealing (Dual Annealing)  
