+++ minion/src/rcmaes.cpp
@@ -11,6 +11,77 @@
 
 namespace {
 
+std::vector<std::pair<double, double>> unit_bounds_for(const std::vector<std::pair<double, double>>& bounds) {
+    if (bounds.empty()) {
+        return bounds;
+    }
+    return std::vector<std::pair<double, double>>(bounds.size(), {0.0, 1.0});
+}
+
+std::vector<double> normalize_point(const std::vector<double>& point,
+                                    const std::vector<std::pair<double, double>>& bounds) {
+    if (bounds.empty()) {
+        return point;
+    }
+    std::vector<double> normalized(point.size(), 0.0);
+    for (size_t i = 0; i < point.size(); ++i) {
+        const double low = bounds[i].first;
+        const double high = bounds[i].second;
+        const double range = high - low;
+        if (range <= 0.0) {
+            normalized[i] = 0.0;
+        } else {
+            normalized[i] = std::clamp((point[i] - low) / range, 0.0, 1.0);
+        }
+    }
+    return normalized;
+}
+
+std::vector<std::vector<double>> normalize_points(const std::vector<std::vector<double>>& points,
+                                                  const std::vector<std::pair<double, double>>& bounds) {
+    if (bounds.empty()) {
+        return points;
+    }
+    std::vector<std::vector<double>> normalized;
+    normalized.reserve(points.size());
+    for (const auto& point : points) {
+        normalized.push_back(normalize_point(point, bounds));
+    }
+    return normalized;
+}
+
+std::vector<double> denormalize_point(const std::vector<double>& point,
+                                      const std::vector<std::pair<double, double>>& bounds) {
+    if (bounds.empty()) {
+        return point;
+    }
+    std::vector<double> actual(point.size(), 0.0);
+    for (size_t i = 0; i < point.size(); ++i) {
+        const double low = bounds[i].first;
+        const double high = bounds[i].second;
+        const double range = high - low;
+        if (range <= 0.0) {
+            actual[i] = low;
+        } else {
+            actual[i] = low + point[i] * range;
+        }
+    }
+    return actual;
+}
+
+std::vector<std::vector<double>> denormalize_points(const std::vector<std::vector<double>>& points,
+                                                    const std::vector<std::pair<double, double>>& bounds) {
+    if (bounds.empty()) {
+        return points;
+    }
+    std::vector<std::vector<double>> actual;
+    actual.reserve(points.size());
+    for (const auto& point : points) {
+        actual.push_back(denormalize_point(point, bounds));
+    }
+    return actual;
+}
+
 std::vector<std::pair<double, double>> merge_intervals_1d(std::vector<std::pair<double, double>> intervals) {
     if (intervals.empty()) {
         return intervals;
@@ -82,7 +153,19 @@ RCMAES::RCMAES(
     size_t maxevals,
     int seed,
     std::map<std::string, ConfigValue> options)
-    : MinimizerBase(func, bounds, x0, data, callback, tol, maxevals, seed, options) {}
+    : MinimizerBase(
+          [func, bounds](const std::vector<std::vector<double>>& candidates, void* user_data) {
+              return func(denormalize_points(candidates, bounds), user_data);
+          },
+          unit_bounds_for(bounds),
+          normalize_points(x0, bounds),
+          data,
+          callback,
+          tol,
+          maxevals,
+          seed,
+          options),
+      original_bounds(bounds) {}
 
 void RCMAES::Parameter::reserve(size_t n_offsprings_reserve_, size_t n_parents_reserve_, size_t n_params_) {
[…]
