+++ minion/src/rcmaes.cpp
@@ -725,7 +725,7 @@ MinionResult RCMAES::optimize() {
             double dim = double(bounds.size());
             const double A = double(lambda);
             const double C = std::max(double(lambda_min), double(dim));
-            double pp =  1.0+ 1.2*exp(-0.034*dim);
+            double pp =  1.0 + exp(-0.03*dim);
             const double t = progress;
             double value = A - (A - C) * (1.0 - std::pow(1.0 - t, pp));
             size_t lambda_target = static_cast<size_t>(std::round(value));
