+++ examples/main_cec_multi.cpp
@@ -10,6 +10,8 @@
 #include <cmath>
 #include <thread>
 #include <algorithm>
+#include <limits>
+#include <filesystem>
 
 namespace {
 
@@ -272,19 +274,83 @@ std::vector <double> objective_function (const std::vector<std::vector<double>> 
     return func->operator()(x); // Call the operator with a single vector
 }
 
+struct MinEvLogger {
+    size_t step = 0;
+    size_t evals = 0;
+    size_t next_index = 0;
+    double best = std::numeric_limits<double>::infinity();
+    std::vector<double> samples;
+
+    MinEvLogger(size_t dimension, size_t max_evals) {
+        step = std::max<size_t>(1, 10 * dimension);
+        size_t count = max_evals / step + 1;
+        samples.assign(count, std::numeric_limits<double>::infinity());
+    }
+
+    void update(const std::vector<double>& values) {
+        for (double v : values) {
+            if (v < best) {
+                best = v;
+            }
+            ++evals;
+            while (next_index < samples.size() && evals >= next_index * step) {
+                samples[next_index] = best;
+                ++next_index;
+            }
+        }
+    }
+
+    void finalize() {
+        while (next_index < samples.size()) {
+            samples[next_index] = best;
+            ++next_index;
+        }
+    }
+};
+
+struct LogContext {
+    minion::CECBase* func = nullptr;
+    MinEvLogger* logger = nullptr;
+};
+
+std::vector<double> objective_function_logged(const std::vector<std::vector<double>>& x, void* data) {
+    auto* ctx = static_cast<LogContext*>(data);
+    auto values = ctx->func->operator()(x);
+    if (ctx->logger) {
+        ctx->logger->update(values);
+    }
+    return values;
+}
+
 void callBack(minion::MinionResult* res) {
     //std::cout << "Best fitness " << res->fun << "\n";
 };
 
-double minimize_cec_functions(int function_number, int dimension, int population_size, int max_evals, int year=2022, std::string algo="ARRDE", int seed=-1) {
+int get_effective_dimension(int function_number, int dimension, int year) {
+    if (year == 2019) { 
+        if (function_number == 1) return 9;
+        if (function_number == 2) return 16;
+        if (function_number == 3) return 18;
+        return 10;
+    }
+    if (year == 2011) {
+        return get_cec2011_problem(function_number).dimension;
+    }
+    return dimension;
+}
+
+double minimize_cec_functions(int function_number,
+                              int dimension,
+                              int population_size,
+                              int max_evals,
+                              int year = 2022,
+                              std::string algo = "ARRDE",
+                              int seed = -1,
+                              MinEvLogger* logger = nullptr) {
     minion::CECBase* cecfunc;
     std::vector<std::pair<double, double>> bounds;
-    int effective_dimension = dimension;
+    int effective_dimension = get_effective_dimension(function_number, dimension, year);
     if (year==2019) { 
-        if (function_number ==1) effective_dimension =9; 
-        else if (function_number==2) effective_dimension =16; 
-        else if (function_number==3) effective_dimension =18;
-        else effective_dimension=10;
         for (int i=0; i<effective_dimension; i++) {
             if (function_number ==1) bounds.push_back(std::make_pair(-8192, 8192)); 
             else if (function_number==2) bounds.push_back(std::make_pair(-16384, 16384)); 
@@ -316,7 +382,17 @@ double minimize_cec_functions(int function_number, int dimension, int population
         x0 = {x00};
     };
 
-    minion::Minimizer optimizer (objective_function,  bounds, x0, cecfunc, callBack, algo, 0.0, max_evals,  seed, settings);
[…]
