### 2026-02-01 f85a349
Switched back to the paper parameter set
