### 2026-02-17 #9229 @gkthiruvathukal
👋 @sjvrijn, @monacofj, @camilochs, @AlessandroPierro - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html.

I need two reviewers to accept before I can commence review. Thank you for your consideration!
