### 2026-02-25 #9229 @monacofj
Hello.

Count me in! I'll be glad help.

On Wed, Feb 25, 2026 at 2:14 AM George K. Thiruvathukal <
***@***.***> wrote:

> *gkthiruvathukal* left a comment (openjournals/joss-reviews#9229)
> <https://github.com/openjournals/joss-reviews/issues/9229#issuecomment-3956878931>
>
> @khoirulmuzakka <https://github.com/khoirulmuzakka> Thus far, none of
> your suggestions has agreed. Please suggest two others. I may have to pause
> this submission.
>
