### 2026-03-04 #9229 @gkthiruvathukal
👋 @anugrahjo, @camilochs - Would either of you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html.

Thanks to @monacofj for accepting already!
