### 2026-03-04 8cf4b12
Improve the CMakeList, make the library header only
