+++ minion/include/minion_cec.h
@@ -1,0 +1,12 @@
+#ifndef MINION_CEC_H
+#define MINION_CEC_H
+
+#include "cec.h"
+#include "cec2011.h"
+#include "cec2014.h"
+#include "cec2017.h"
+#include "cec2019.h"
+#include "cec2020.h"
+#include "cec2022.h"
+
+#endif // MINION_CEC_H
