+++ compile.sh
@@ -30,11 +30,19 @@ echo "Configuring with CMake..."
 if [[ "${build_type}" == "Debug" ]]; then
   cmake -G "Unix Makefiles" \
     -DCMAKE_BUILD_TYPE=${build_type} \
+    -DMINION_BUILD_CEC=ON \
+    -DMINION_BUILD_EXAMPLES=ON \
+    -DMINION_BUILD_PYTHON=ON \
     -DCMAKE_CXX_FLAGS_DEBUG="-O0 -g3 -fno-omit-frame-pointer" \
     -DCMAKE_C_FLAGS_DEBUG="-O0 -g3 -fno-omit-frame-pointer" \
     ..
 else
-  cmake -G "Unix Makefiles" -DCMAKE_BUILD_TYPE=${build_type} ..
+  cmake -G "Unix Makefiles" \
+    -DCMAKE_BUILD_TYPE=${build_type} \
+    -DMINION_BUILD_CEC=ON \
+    -DMINION_BUILD_EXAMPLES=ON \
+    -DMINION_BUILD_PYTHON=ON \
+    ..
 fi
 
 # Compile with optimal parallelization
