+++ minion/include/minion.h
@@ -3,24 +3,18 @@
 
 #include "utility.h"
 #include "minimizer_base.h"
+#include "default_options.h"
 #include "de.h"
 #include "gwo_de.h"
-#include "cec2020.h"
-#include "cec2022.h"
-#include "cec2017.h"
-#include "cec2019.h"
-#include "cec2014.h"
 #include "lshade.h"
 #include "agsk.h"
 #include "arrde.h"
 #include "nlshadersp.h"
-#include "cec.h"
 #include "j2020.h"
 #include "jade.h"
 #include "jso.h"
 #include "lsrtde.h"
 #include "nelder_mead.h"
-#include "default_options.h"
 #include "minimizer.h"
 #include "imode.h"
 #include "abc.h"
@@ -33,5 +27,35 @@
 #include "lshadecnepsin.h"
 #include "bipop_acmaes.h"
 #include "rcmaes.h"
+#include "cmaes.h"
+#include "test_functions.h"
+
+#ifdef MINION_ALGORITHMS_IMPLEMENTATION
+#include "utility.cpp"
+#include "test_functions.cpp"
+#include "de.cpp"
+#include "gwo_de.cpp"
+#include "lshade.cpp"
+#include "agsk.cpp"
+#include "arrde.cpp"
+#include "nlshadersp.cpp"
+#include "j2020.cpp"
+#include "jade.cpp"
+#include "jso.cpp"
+#include "lsrtde.cpp"
+#include "imode.cpp"
+#include "abc.cpp"
+#include "dual_annealing.cpp"
+#include "l_bfgs_b.cpp"
+#include "l_bfgs.cpp"
+#include "nelder_mead.cpp"
+#include "pso.cpp"
+#include "spso2011.cpp"
+#include "dmspso.cpp"
+#include "lshadecnepsin.cpp"
+#include "cmaes.cpp"
+#include "bipop_acmaes.cpp"
+#include "rcmaes.cpp"
+#endif
 
 #endif // MINION_H
