+++ minion/src/bindings.cpp
@@ -3,38 +3,9 @@
 #include <pybind11/functional.h>
 #include <pybind11/chrono.h>
 #include <pybind11/complex.h>
-#include "minimizer_base.h" // Include your MinionResult struct definition
-#include "gwo_de.h"
-#include "nelder_mead.h"
-#include "utility.h"
-#include "cec2011.h"
-#include "cec2017.h"
-#include "cec2014.h"
-#include "cec2019.h"
-#include "cec2020.h"
-#include "cec2022.h"
-#include "lshade.h"
-#include "agsk.h"
-#include "jade.h"
-#include "de.h"
-#include "arrde.h"
-#include "nlshadersp.h"
-#include "j2020.h"
-#include "lsrtde.h"
-#include "imode.h"
-#include "jso.h"
-#include "abc.h"
-#include "pso.h"
-#include "spso2011.h"
-#include "dmspso.h"
-#include "lshadecnepsin.h"
-#include "cmaes.h"
-#include "rcmaes.h"
-#include "dual_annealing.h"
-#include "l_bfgs_b.h"
-#include "minimizer.h"
-#include "l_bfgs.h"
-#include "bipop_acmaes.h"
+#define MINION_ALGORITHMS_IMPLEMENTATION
+#include "minion.h"
+#include "minion_cec.h"
 #include <exception>
 #include <pybind11/stl_bind.h>
 #include <any>
