+++ cmake/MinionConfig.cmake.in
@@ -1,0 +1,6 @@
+@PACKAGE_INIT@
+
+include(CMakeFindDependencyMacro)
+find_dependency(Eigen3 CONFIG)
+
+include("${CMAKE_CURRENT_LIST_DIR}/MinionTargets.cmake")

[…]
