### 2026-03-04 c2375d0
Updated compile script
