+++ minion/include/jade.h
@@ -31,7 +31,6 @@ class JADE : public Differential_Evolution {
          * 
          * @param func The objective function to minimize.
          * @param bounds The bounds for the variables.
-         * @param options A map of configuration options.
          * @param x0 The initial guesses for the solution. Note that Minion assumes multiple initial guesses, thus, x0 is an std::vector<std::vector<double>> object. These guesses will be used for population initialization.
          * @param data Additional data for the objective function.
          * @param callback Callback function for intermediate results.
