+++ minion/include/l_bfgs_b.h
@@ -50,7 +50,7 @@ public:
     /**
      * @brief Constructor for L-BFGS-B.
      * @param func The objective function to minimize.
-     * @param bounds The bounds for the variables.
+     * @param bounds_ The bounds for the variables.
      * @param x0 The initial guesses. Note that Minion assumes multiple initial guesses, thus, x0 is an std::vector<std::vector<double>> object. Minion will pick the best guess from x0 as the the actual initialization.
      * @param data Additional data for the objective function.
      * @param callback Callback function for intermediate results.
