+++ docs/source/changelog.rst
@@ -5,6 +5,15 @@ Changelog
    :local:
    :depth: 2
 
+
+Version 1.2.0 - 2026-03-05
+--------------------------
+
+**Added**
+
+- Added RCMAES
+- Streamlined CMake usage.
+
 Version 1.1.0 - 2025-11-23
 --------------------------
 
