+++ docs/source/installation.rst
@@ -1,114 +1,156 @@
 Compilation and Installation
 =============================
 
-MinionPy Installation
-----------------------
+MinionPy (Python)
+-----------------
 
-MinionPy can be installed directly via PIP:
+Install from PyPI:
 
 .. code-block:: shell
 
    pip install --upgrade minionpy
 
-To verify the installation, run:
+Quick check:
 
 .. code-block:: python
 
    import minionpy as mpy
 
-If MinionPy is not installed via PIP and you want to use the locally compiled version, ensure that the correct path is added to `sys.path`:
+
+Build Minion From Source
+------------------------
+
+Build requirements:
+
+- CMake >= 3.18
+- C++17 compiler
+- pybind11 (only when building the Python extension)
+
+On Windows, use Visual C++ Build Tools.
+
+Build with helper scripts:
+
+- Linux/macOS: ``./compile.sh``
+- Windows: ``compile.bat``
+
+Or build manually with CMake.
+
+Minimal C++ setup (headers + interface target):
+
+.. code-block:: shell
+
+   cmake -S . -B build \
+     -DMINION_BUILD_CEC=OFF \
+     -DMINION_BUILD_PYTHON=OFF \
+     -DMINION_BUILD_EXAMPLES=OFF
+   cmake --build build --config Release
+
+Build all optional components:
+
+.. code-block:: shell
+
+   cmake -S . -B build \
+     -DMINION_BUILD_CEC=ON \
+     -DMINION_BUILD_PYTHON=ON \
+     -DMINION_BUILD_EXAMPLES=ON
+   cmake --build build --config Release
+
+
+Install as a CMake Package
+--------------------------
+
+Install Minion so downstream projects can use ``find_package(Minion CONFIG REQUIRED)``:
+
+.. code-block:: shell
+
+   cmake -S . -B build \
+     -DMINION_BUILD_CEC=ON \
+     -DMINION_BUILD_PYTHON=OFF \
+     -DMINION_BUILD_EXAMPLES=OFF
+   cmake --build build --config Release
+   cmake --install build --prefix /path/to/minion-install
+
+This installs:
+
+- headers under ``include/``
+- CMake package files under ``lib/cmake/Minion``
+- optional ``Minion_cec`` static library when ``MINION_BUILD_CEC=ON``
+
+
+Use Minion in Your CMake Project
+--------------------------------
+
+1. Point CMake to your Minion install prefix:
+
+.. code-block:: shell
+
+   cmake -S . -B build -DCMAKE_PREFIX_PATH=/path/to/minion-install
+
+2. In your project ``CMakeLists.txt``:
+
+.. code-block:: cmake
+
+   cmake_minimum_required(VERSION 3.18)
+   project(MyApp LANGUAGES CXX)
+
+   find_package(Minion CONFIG REQUIRED)
+
+   add_executable(my_app src/main.cpp src/solver.cpp)
+   target_link_libraries(my_app PRIVATE minion)
+
+   # Optional: only if you use CEC benchmarks
+   # target_link_libraries(my_app PRIVATE Minion_cec)
+
+   target_compile_features(my_app PRIVATE cxx_std_17)
+
+3. In exactly one translation unit, define the implementation macro:
+
+.. code-block:: cpp
+
+   #define MINION_ALGORITHMS_IMPLEMENTATION
+   #include <minion/minion.h>
+
+In all other translation units, include the header without the macro:
+
+.. code-block:: cpp
+
+   #include <minion/minion.h>
+
+
+Alternative: FetchContent (No System Install)
+---------------------------------------------
+
+If you do not want to pre-install Minion:
+
+.. code-block:: cmake
+
+   include(FetchContent)
+
[…]
