+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "minionpy"
-version = "1.1.0"
+version = "1.2.0"
 description = "MinionPy is the Python implementation of the Minion C++ library, designed for derivative-free optimization."
 readme = "minionpy.md"
 license = { file = "LICENSE.txt" }
