+++ .github/workflows/ci.yml
@@ -3,11 +3,11 @@ name: CI
 on:
   push:
     branches:
-      - main  # or any branch you want
-      - 'feature/*'  # Optional: trigger on feature branches
+      - main
+      - "feature/*"
   pull_request:
     branches:
-      - main  # Optional: trigger on PRs
+      - main
 
 jobs:
   build:
@@ -15,46 +15,27 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: ["3.8", "3.13"]
-        
+        python-version: ["3.9", "3.13"]
+
     steps:
-    # Checkout the repository
-    - name: Checkout code
-      uses: actions/checkout@v4
+      - name: Checkout code
+        uses: actions/checkout@v4
 
-    # Set up Python
-    - name: Setup Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v4
-      with:
-        python-version: ${{ matrix.python-version }}
+      - name: Setup Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
 
-    # Install dependencies
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install numpy scipy build pybind11
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install numpy scipy build pybind11
 
-    # Install Eigen
-    - name: Install Eigen (Linux/macOS)
-      if: runner.os != 'Windows'
-      run: |
-        mkdir -p eigen
-        curl -L https://gitlab.com/libeigen/eigen/-/archive/3.4.0/eigen-3.4.0.tar.gz | tar xz --strip-components=1 -C eigen
+      - name: Configure CMake
+        run: cmake -B build -S . -DCMAKE_BUILD_TYPE=Release -DMINION_BUILD_PYTHON=ON -DMINION_BUILD_CEC=ON -DMINION_BUILD_EXAMPLES=ON
 
-    - name: Install Eigen (Windows)
-      if: runner.os == 'Windows'
-      run: |
-        mkdir eigen
-        Invoke-WebRequest -Uri "https://gitlab.com/libeigen/eigen/-/archive/3.4.0/eigen-3.4.0.zip" -OutFile "eigen.zip"
-        Expand-Archive -Path "eigen.zip" -DestinationPath "eigen"
+      - name: Build C++ extension
+        run: cmake --build build --target minionpycpp --config Release
 
-    # Build the library
-    - name: Build the library
-      run: |
-        cmake -B build -S . -DCMAKE_BUILD_TYPE=Release -DEIGEN3_INCLUDE_DIR=$EIGEN3_INCLUDE_DIR
-        cmake --build build --target minionpycpp
-        cmake --build build --target minion
-
-    # Verify build and run Python build
-    - name: Python build
-      run: python -m build
+      - name: Python build
+        run: python -m build
