### 2026-03-05 5ca51f7
Reverted back to compiled lib style
