+++ CMakeLists.txt
@@ -30,6 +30,7 @@ option(MINION_BUILD_EXAMPLES "Build C++ examples" OFF)
 
 file(GLOB MINION_ALGO_IMPL_SOURCES minion/src/*.cpp)
 list(REMOVE_ITEM MINION_ALGO_IMPL_SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/minion/src/bindings.cpp)
+file(GLOB MINION_CEC_SOURCE_FILES minion/cec/*.cpp)
 
 include(FetchContent)
 option(MINION_FETCH_EIGEN "Fetch Eigen headers when Eigen3 package is not found" ON)
@@ -63,48 +64,24 @@ if(NOT Eigen3_FOUND)
 endif()
 
 # + ----------------------------------------- +
-# | Header-only algorithms public target      |
+# | Compiled algorithms core target           |
 # + ----------------------------------------- +
-add_library(minion INTERFACE)
-target_include_directories(minion INTERFACE
+set(MINION_SOURCE_FILES
+    ${MINION_ALGO_IMPL_SOURCES}
+    ${MINION_CEC_SOURCE_FILES}
+)
+
+add_library(minion SHARED ${MINION_SOURCE_FILES})
+
+target_include_directories(minion PUBLIC
     $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/minion/include>
-    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/minion/src>
+    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/minion/cec>
     $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/external/lbfgspp/include>
     $<INSTALL_INTERFACE:include>
     $<INSTALL_INTERFACE:include/minion>
-    $<INSTALL_INTERFACE:include/minion/src>
 )
-target_link_libraries(minion INTERFACE
-    Eigen3::Eigen
-)
-
-# + -------------------------- +
-# | Optional compiled CEC lib  |
-# + -------------------------- +
-if(MINION_BUILD_CEC OR MINION_BUILD_PYTHON OR MINION_BUILD_EXAMPLES)
-    file(GLOB MINION_CEC_SOURCE_FILES minion/cec/*.cpp)
-    add_library(Minion_cec STATIC ${MINION_CEC_SOURCE_FILES})
-    target_include_directories(Minion_cec PUBLIC
-        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/minion/include>
-        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/minion/cec>
-        $<INSTALL_INTERFACE:include>
-    )
-    target_link_libraries(Minion_cec PUBLIC Eigen3::Eigen)
-
-    if (MSVC)
-        set_target_properties(Minion_cec PROPERTIES
-            LIBRARY_OUTPUT_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/lib/$<0:>
-            RUNTIME_OUTPUT_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/lib/$<0:>
-            ARCHIVE_OUTPUT_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/lib/$<0:>
-            WINDOWS_EXPORT_ALL_SYMBOLS TRUE
-        )
-    else()
-        set_target_properties(Minion_cec PROPERTIES
[…]
