+++ CMakeLists.txt
@@ -28,6 +28,11 @@ option(MINION_BUILD_CEC "Build compiled CEC library" OFF)
 option(MINION_BUILD_PYTHON "Build Python extension module" ${_default_python})
 option(MINION_BUILD_EXAMPLES "Build C++ examples" OFF)
 
+if(MINION_BUILD_PYTHON AND NOT MINION_BUILD_CEC)
+    message(STATUS "MINION_BUILD_PYTHON=ON requires CEC symbols from bindings; enabling MINION_BUILD_CEC.")
+    set(MINION_BUILD_CEC ON CACHE BOOL "Build compiled CEC library" FORCE)
+endif()
+
 file(GLOB MINION_ALGO_IMPL_SOURCES minion/src/*.cpp)
 list(REMOVE_ITEM MINION_ALGO_IMPL_SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/minion/src/bindings.cpp)
 file(GLOB MINION_CEC_SOURCE_FILES minion/cec/*.cpp)
@@ -66,22 +71,32 @@ endif()
 # + ----------------------------------------- +
 # | Compiled algorithms core target           |
 # + ----------------------------------------- +
-set(MINION_SOURCE_FILES
-    ${MINION_ALGO_IMPL_SOURCES}
-    ${MINION_CEC_SOURCE_FILES}
-)
-
-add_library(minion SHARED ${MINION_SOURCE_FILES})
+set(MINION_CORE_SOURCE_FILES ${MINION_ALGO_IMPL_SOURCES})
+add_library(minion SHARED ${MINION_CORE_SOURCE_FILES})
 
 target_include_directories(minion PUBLIC
     $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/minion/include>
-    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/minion/cec>
     $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/external/lbfgspp/include>
     $<INSTALL_INTERFACE:include>
     $<INSTALL_INTERFACE:include/minion>
 )
 target_link_libraries(minion PUBLIC Eigen3::Eigen)
 set_target_properties(minion PROPERTIES POSITION_INDEPENDENT_CODE ON)
+
+if(MINION_BUILD_CEC)
+    add_library(minion_cec SHARED ${MINION_CEC_SOURCE_FILES})
+    target_include_directories(minion_cec PUBLIC
+        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/minion/cec>
+        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/minion/include>
+        $<INSTALL_INTERFACE:include>
+        $<INSTALL_INTERFACE:include/minion>
+    )
+    target_link_libraries(minion_cec PUBLIC minion Eigen3::Eigen)
+    set_target_properties(minion_cec PROPERTIES POSITION_INDEPENDENT_CODE ON)
+
+    # Keep CEC input data available next to build-tree binaries/libraries.
+    file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/cec_input_data DESTINATION ${CMAKE_BINARY_DIR})
+endif()
 
 # + -------------------------- +
 # | Optional C++ examples      |
@@ -90,26 +105,38 @@ if(MINION_BUILD_EXAMPLES)
     add_executable(main_mini examples/main_minimizer.cpp)
     target_link_libraries(main_mini PRIVATE minion)
 
-    add_executable(main_cec examples/main_cec.cpp)
-    target_link_libraries(main_cec PRIVATE minion)
[…]
