+++ CMakeLists.txt
@@ -8,12 +8,12 @@ if (POLICY CMP0148)
   cmake_policy(SET CMP0148 NEW)
 endif()
 
-if (APPLE)
-    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -O3 -Wno-unused-result")
-elseif (WIN32)
+if (MSVC)
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} /O2")
     add_definitions(-D_CRT_SECURE_NO_WARNINGS)
-elseif (UNIX AND CMAKE_COMPILER_IS_GNUCXX)
+elseif (MINGW OR CMAKE_CXX_COMPILER_ID STREQUAL "GNU" OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
+    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -O3 -Wno-unused-result")
+elseif (APPLE)
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -O3 -Wno-unused-result")
 else()
     message(WARNING "Unknown platform. Please check compiler settings.")
