### 2026-03-06 #10145 @editorialbot
[REVIEW]: Minion: A C++ and Python Library for Single-Objective Optimization Algorithms
**Submitting author:** @Muzakka (<a href="http://orcid.org/0000-0002-3888-1697">Khoirul Faiq  Muzakka</a>)
**Repository:** https://github.com/khoirulmuzakka/Minion
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.2.5
