+++ tests/main.cpp
@@ -1,0 +1,149 @@
+#include <iomanip>
+#include <iostream>
+#include <cmath>
+#include <limits>
+#include <map>
+#include <string>
+#include <vector>
+
+#include <minion.h>
+#include <minion_cec.h>
+
+namespace {
+
+std::vector<double> rosenbrock_batch(const std::vector<std::vector<double>>& X, void*) {
+    std::vector<double> out(X.size(), 0.0);
+    for (size_t i = 0; i < X.size(); ++i) {
+        const auto& x = X[i];
+        double f = 0.0;
+        for (size_t j = 0; j + 1 < x.size(); ++j) {
+            const double a = x[j + 1] - x[j] * x[j];
+            const double b = 1.0 - x[j];
+            f += 100.0 * a * a + b * b;
+        }
+        out[i] = f;
+    }
+    return out;
+}
+
+std::vector<double> cec2017_batch(const std::vector<std::vector<double>>& X, void* data) {
+    auto* cec = static_cast<minion::CECBase*>(data);
+    return (*cec)(X);
+}
+
+}  // namespace
+
+int main() {
+    minion::MinionFunction func = rosenbrock_batch;
+
+    const size_t dim = 5;
+    std::vector<std::pair<double, double>> bounds(dim, {-5.0, 5.0});
+    std::vector<double> x0(dim, 0.5);
+
+    const std::vector<std::string> algorithms = {
+        "DE", "LSHADE", "AGSK", "JADE", "j2020", "NLSHADE_RSP", "LSRTDE", "jSO",
+        "IMODE", "ARRDE", "GWO_DE", "NelderMead", "ABC", "PSO", "SPSO2011", "DMSPSO",
+        "LSHADE_cnEpSin", "CMAES", "RCMAES", "BIPOP_aCMAES", "DA", "L_BFGS_B", "L_BFGS"
+    };
+
+    const size_t maxevals = 4000;
+    const double tol = 1e-8;
+    const size_t eval_slack = 100;
+
+    int failed_checks = 0;
+    int total_checks = 0;
+
+    std::cout << "Rosenbrock minimization using all Minion algorithms\n";
+    std::cout << "dimension=" << dim << ", maxevals=" << maxevals << "\n\n";
+    std::cout << std::left << std::setw(18) << "Algorithm" << std::right << std::setw(16) << "best_f" << std::setw(12)
+              << "nfev" << '\n';
+    std::cout << std::string(46, '-') << '\n';
+
+    const std::map<std::string, double> rosenbrock_upper = {
+        {"DE", 10.0}, {"LSHADE", 10.0}, {"AGSK", 10.0}, {"JADE", 10.0}, {"j2020", 10.0},
+        {"NLSHADE_RSP", 10.0}, {"LSRTDE", 10.0}, {"jSO", 10.0}, {"IMODE", 10.0}, {"ARRDE", 10.0},
+        {"GWO_DE", 10.0}, {"NelderMead", 10.0}, {"ABC", 10.0}, {"PSO", 10.0}, {"SPSO2011", 10.0},
+        {"DMSPSO", 10.0}, {"LSHADE_cnEpSin", 10.0}, {"CMAES", 10.0}, {"RCMAES", 10.0},
+        {"BIPOP_aCMAES", 10.0}, {"DA", 10.0}, {"L_BFGS_B", 10.0}, {"L_BFGS", 10.0},
+    };
+
+    for (const auto& algo : algorithms) {
+        try {
+            minion::Minimizer opt(func, bounds, x0, nullptr, nullptr, algo, tol, maxevals, 42);
+            minion::MinionResult res = opt.optimize();
+            std::cout << std::left << std::setw(18) << algo << std::right << std::setw(16) << std::setprecision(8)
[…]
