+++ docs/source/minion/examples.rst
@@ -1,152 +1,272 @@
-This section provides examples of how to use the **Minion** library to solve bound-constrained optimization problems using various algorithms.
+Examples
+========
 
-Basic Usage
-===========
+This page shows practical C++ usage patterns for Minion.
 
-This example demonstrates how to use **Minion** to optimize the Rosenbrock and Rastrigin functions using different optimization algorithms.
+Reference files in this repository:
 
-## Example: Optimizing Rosenbrock Function with Minion
-
-The following C++ example showcases how to use Minion to optimize the Rosenbrock function using multiple evolutionary algorithms.
-
-.. code-block:: cpp
-
-    #include <iostream>
-    #include <vector>
-    #include "minion.h"
-
-    // Define the Rosenbrock function
-    double rosenbrock(const std::vector<double>& x) {
-        double sum = 0.0;
-        for (size_t i = 0; i < x.size() - 1; i++) {
-            sum += 100 * pow(x[i + 1] - x[i] * x[i], 2) + pow(1 - x[i], 2);
-        }
-        return sum;
-    }
-
-    // Vectorized version for Minion
-    std::vector<double> rosenbrock_vect(const std::vector<std::vector<double>>& X, void* data) {
-        std::vector<double> ret;
-        for (const auto& x : X) ret.push_back(rosenbrock(x));
-        return ret;
-    }
-
-    int main() {
-        // List of optimization algorithms to test
-        std::vector<std::string> algoList = {"ARRDE", "AGSK", "LSHADE", "LSRTDE"};
-        
-        // Define the problem dimensions and bounds
-        size_t dimension = 10;
-        std::vector<std::pair<double, double>> bounds(dimension, {-5.0, 5.0});
-        size_t max_evals = 100000;
-
-        // Iterate over each algorithm and perform optimization
-        for (const auto& algo : algoList) {
-            auto settings = minion::DefaultSettings().getDefaultSettings(algo);
-            minion::MinionResult res = minion::Minimizer(rosenbrock_vect, bounds, {}, nullptr, nullptr, algo, 0.0, max_evals, -1, settings).optimize();
-            std::cout << algo << " : " << res.fun << "\n";
-        }
-        return 0;
-    }
+- ``examples/main_minimizer.cpp``
+- ``examples/main_cec.cpp``
+- ``tests/main.cpp``
 
 
-Minion supports the following optimization algorithms:
+Basic Pattern
+=============
 
-- **Differential Evolution (DE) Variants:**
-
-    - ``"LSHADE"``  (Success History Adaptive DE with Linear Population Size Reduction)
-    - ``"DE"``      (The original DE)
-    - ``"JADE"``     (An adaptive DE)
-    - ``"jSO"``      (A variant of LSHADE with some improvements)
-    - ``"NLSHADE_RSP"`` (A variant of LSHADE with some improvements)
-    - ``"LSHADE_cnEpSin"`` (A variant of LSHADE with some improvements)
-    - ``"IMODE"`` (Improved Multi-operator Differential Evolution)
-    - ``"j2020"``  (A variant of jDE algorithm)
-    - ``"GWO_DE"``  (Grey Wolf-DE optimization)
-    - ``"ARRDE"``   (Adaptive restart-refine DE)
-    - ``"AGSK"``    (Adaptive gaining-sharing knowledge-based DE)
-    - ``"LSRTDE"``  (A variant of LSHADE with some improvements)
-
-- **Swarm Intelligence Algorithms:**
-
-    - ``"ABC"`` (Artificial Bee Colony)
-    - ``"PSO"`` (Particle Swarm Optimization)
[…]
