### 2026-03-06 7fb8eee
Version bump v1.3.0
