### 2026-03-06 e571592
Fixed python extension crash
