+++ pyproject.toml
@@ -4,8 +4,8 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "minionpy"
-version = "1.3.0"
-description = "MinionPy is the Python implementation of the Minion C++ library, designed for derivative-free optimization."
+version = "1.4.0"
+description = "MinionPy is the Python implementation of the Minion (C++) ibrary."
 readme = "minionpy.md"
 license = { file = "LICENSE.txt" }
 authors = [
