+++ docs/source/installation.rst
@@ -1,7 +1,7 @@
 Compilation and Installation
-=============================
+============================
 
-MinionPy (Python)
+Python (MinionPy)
 -----------------
 
 Install from PyPI:
@@ -17,106 +17,78 @@ Quick check:
    import minionpy as mpy
 
 
-Build Minion From Source
+Native Installation by Platform
+-------------------------------
+
+- **Windows**: compile Minion from source with CMake.
+- **Linux**: install the ``.deb`` package from GitHub Release assets.
+- **macOS**: install from release archive (``.tgz`` / ``tgz.zip`` asset).
+
+Linux ``.deb`` install:
+
+.. code-block:: shell
+
+   sudo dpkg -i minion_<version>_<arch>.deb
+   sudo apt-get install -f
+
+macOS ``.tgz`` / ``tgz.zip`` install:
+
+.. code-block:: shell
+
+   unzip minion-<version>-macos.tgz.zip
+   tar -xzf minion-<version>-macos.tgz -C /tmp/minion_pkg
+   sudo rsync -a /tmp/minion_pkg/ /usr/local/
+
+
+Compilation From Scratch
 ------------------------
 
 Build requirements:
 
 - CMake >= 3.18
-- C++17 compiler
-- pybind11 (only when building the Python extension)
-
-On Windows, use Visual C++ Build Tools.
+- C++17 compiler (GCC/Clang/MSVC)
+- Eigen3 (or automatic fetch through CMake)
+- Python 3 + pybind11 (only when building Python bindings)
 
 Build with helper scripts:
 
 - Linux/macOS: ``./compile.sh``
 - Windows: ``compile.bat``
 
-Or build manually with CMake.
-
-Minimal C++ setup (compiled library):
-
-.. code-block:: shell
-
-   cmake -S . -B build \
-     -DMINION_BUILD_CEC=OFF \
-     -DMINION_BUILD_PYTHON=OFF \
-     -DMINION_BUILD_EXAMPLES=OFF
-   cmake --build build --config Release
-
-Build all optional components:
+Manual build with CMake:
 
 .. code-block:: shell
 
    cmake -S . -B build \
      -DMINION_BUILD_CEC=ON \
-     -DMINION_BUILD_PYTHON=ON \
+     -DMINION_BUILD_PYTHON=OFF \
      -DMINION_BUILD_EXAMPLES=ON
    cmake --build build --config Release
 
-
-Install as a CMake Package
---------------------------
-
-Install Minion so downstream projects can use ``find_package(Minion CONFIG REQUIRED)``:
+Install:
 
 .. code-block:: shell
 
-   cmake -S . -B build \
-     -DMINION_BUILD_CEC=ON \
-     -DMINION_BUILD_PYTHON=OFF \
-     -DMINION_BUILD_EXAMPLES=OFF
-   cmake --build build --config Release
    cmake --install build --prefix /path/to/minion-install
 
-This installs:
 
-- headers under ``include/``
-- compiled Minion library under ``lib/`` (for example ``libminion.so`` on Linux)
-- CMake package files under ``lib/cmake/Minion``
-- optional compiled CEC library ``libminion_cec`` when ``MINION_BUILD_CEC=ON``
-- CEC input data under ``cec_input_data/`` when ``MINION_BUILD_CEC=ON``
+Using Minion in a C++ Project
+-----------------------------
 
+Include header:
 
-Use Minion in Your CMake Project
---------------------------------
+.. code-block:: cpp
 
-1. Point CMake to your Minion install prefix:
+   #include <minion/minion.h>
 
-.. code-block:: shell
-
[…]
