+++ README.md
@@ -53,56 +53,77 @@ Minion also serves as a **research platform** for developing and testing new opt
   - The benchmark problems are directly adapted from the original C++ implementations.  
   - The CEC 2011 suite, containing 22 real-world optimization problems, has been completely rewritten into C++ from the MATLAB version, making it much faster than the original.
 
-## 🚀 Installation
-The Python wrapper (**minionpy**) is available on PyPI:
+## 🚀 Installation and Usage
+
+### Python
+Install from PyPI:
 
 ```sh
 pip install --upgrade minionpy
 ```
 
-For compiling the C++ version, please refer to the official Minion documentation.
+### C++ / Native package
+- Windows: compile from source with CMake (see "Compilation From Scratch" and "Using Minion in a C++ Project" below).
+- Linux: install the `.deb` package from GitHub Release assets.
+- macOS: use the release archive (`.tgz` / `tgz.zip` asset).
 
-### C++ Compilation (Quick)
+#### Linux `.deb` install
+```sh
+sudo dpkg -i minion_<version>_<arch>.deb
+sudo apt-get install -f
+```
+
+#### macOS `.tgz` / `tgz.zip` install
+If the asset is `tgz.zip`, unzip first:
+```sh
+unzip minion-<version>-macos.tgz.zip
+```
+
+Extract package content:
+```sh
+tar -xzf minion-<version>-macos.tgz -C /tmp/minion_pkg
+```
+
+Install manually (example to `/usr/local`):
+```sh
+sudo rsync -a /tmp/minion_pkg/ /usr/local/
+```
+
+### Compilation From Scratch
+Dependencies:
+- CMake >= 3.18
+- C++17 compiler (GCC/Clang/MSVC)
+- Eigen3 (or allow automatic fetch via CMake)
+- Optional for Python bindings: Python 3 + `pybind11`
+
 Build with helper scripts:
-
 - Windows: `compile.bat`
 - Linux/macOS: `compile.sh`
 
-Or with CMake manually:
-
+Manual CMake build:
 ```sh
-cmake -S . -B build -DMINION_BUILD_CEC=OFF -DMINION_BUILD_PYTHON=OFF -DMINION_BUILD_EXAMPLES=OFF
+cmake -S . -B build \
+  -DMINION_BUILD_CEC=ON \
+  -DMINION_BUILD_PYTHON=OFF \
+  -DMINION_BUILD_EXAMPLES=ON
 cmake --build build --config Release
+cmake --install build --prefix /usr/local
 ```
 
-Enable all components (CEC, Python extension, examples):
-
-```sh
-cmake -S . -B build -DMINION_BUILD_CEC=ON -DMINION_BUILD_PYTHON=ON -DMINION_BUILD_EXAMPLES=ON
-cmake --build build --config Release
+### Using Minion in a C++ Project
+Header include:
+```cpp
+#include <minion/minion.h>
 ```
 
-### Use Minion in Another C++ Project
-Install Minion first:
-
-```sh
-cmake -S . -B build
-cmake --install build --prefix /path/to/minion/install
-```
-
-Then in your project CMake:
+CMake with `find_package` first, then `FetchContent` fallback:
 
 ```cmake
-find_package(Minion CONFIG REQUIRED)
-add_executable(my_app main.cpp)
-target_link_libraries(my_app PRIVATE minion)
-# Optional: CEC benchmark suite
-# target_link_libraries(my_app PRIVATE minion_cec)
-```
+cmake_minimum_required(VERSION 3.18)
+project(my_app LANGUAGES CXX)
+set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
[…]
