### 2026-03-06 86459fc
Fixed problem with minmax on some C++ project in Windows
