### 2026-03-06 c76d2cd
Fixed CEC data can not found issue
