### 2026-03-10 68701c8
Improved LBFGS and LBFGSB, where they are now standalone of the LbFGSpp library
