+++ external/lbfgspp/include/LBFGSB.h
@@ -1,290 +1,0 @@
-// Copyright (C) 2020-2023 Yixuan Qiu <yixuan.qiu@cos.name>
-// Under MIT license
-
-#ifndef LBFGSPP_LBFGSB_H
-#define LBFGSPP_LBFGSB_H
-
-#include <stdexcept>  // std::invalid_argument
-#include <vector>
-#include <Eigen/Core>
-#include "LBFGSpp/Param.h"
-#include "LBFGSpp/BFGSMat.h"
-#include "LBFGSpp/Cauchy.h"
-#include "LBFGSpp/SubspaceMin.h"
-#include "LBFGSpp/LineSearchMoreThuente.h"
-
-namespace LBFGSpp {
-
-///
-/// L-BFGS-B solver for box-constrained numerical optimization
-///
-template <typename Scalar,
-          template <class> class LineSearch = LineSearchMoreThuente>
-class LBFGSBSolver
-{
-private:
-    using Vector = Eigen::Matrix<Scalar, Eigen::Dynamic, 1>;
-    using Matrix = Eigen::Matrix<Scalar, Eigen::Dynamic, Eigen::Dynamic>;
-    using MapVec = Eigen::Map<Vector>;
-    using IndexSet = std::vector<int>;
-    BFGSMat<Scalar, true> m_bfgs;        // Approximation to the Hessian matrix
-
-    const LBFGSBParam<Scalar>& m_param;  // Parameters to control the LBFGS algorithm
-    
-    Vector m_fx;                         // History of the objective function values
-    Vector m_xp;                         // Old x
-    Vector m_grad;                       // New gradient
-    Scalar m_projgnorm;                  // Projected gradient norm
-    Vector m_gradp;                      // Old gradient
-    Vector m_drt;                        // Moving direction
-
-    // Reset internal variables
-    // n: dimension of the vector to be optimized
-    inline void reset(int n)
-    {
-        const int m = m_param.m;
-        m_bfgs.reset(n, m);
-        m_xp.resize(n);
-        m_grad.resize(n);
-        m_gradp.resize(n);
-        m_drt.resize(n);
-        if (m_param.past > 0)
-            m_fx.resize(m_param.past);
-    }
-
-    // Project the vector x to the bound constraint set
-    static void force_bounds(Vector& x, const Vector& lb, const Vector& ub)
-    {
-        x.noalias() = x.cwiseMax(lb).cwiseMin(ub);
-    }
-
-    // Norm of the projected gradient
-    // ||P(x-g, l, u) - x||_inf
-    static Scalar proj_grad_norm(const Vector& x, const Vector& g, const Vector& lb, const Vector& ub)
-    {
-        return ((x - g).cwiseMax(lb).cwiseMin(ub) - x).cwiseAbs().maxCoeff();
-    }
-
-    // The maximum step size alpha such that x0 + alpha * d stays within the bounds
-    static Scalar max_step_size(const Vector& x0, const Vector& drt, const Vector& lb, const Vector& ub)
-    {
-        const int n = x0.size();
-        Scalar step = std::numeric_limits<Scalar>::infinity();
-
-        for (int i = 0; i < n; i++)
-        {
-            if (drt[i] > Scalar(0))
-            {
-                step = std::min(step, (ub[i] - x0[i]) / drt[i]);
-            }
-            else if (drt[i] < Scalar(0))
-            {
-                step = std::min(step, (lb[i] - x0[i]) / drt[i]);
-            }
-        }
-        //std::cout << n << x0 << drt<< lb << ub  << step << "\n";
-        return step;
-    }
-
-public:
-    ///
-    /// Constructor for the L-BFGS-B solver.
-    ///
[…]
