+++ external/lbfgspp/include/LBFGSpp/Param.h
@@ -1,399 +1,0 @@
-// Copyright (C) 2016-2023 Yixuan Qiu <yixuan.qiu@cos.name>
-// Under MIT license
-
-#ifndef LBFGSPP_PARAM_H
-#define LBFGSPP_PARAM_H
-
-#include <Eigen/Core>
-#include <stdexcept>  // std::invalid_argument
-
-namespace LBFGSpp {
-
-///
-/// \defgroup Enumerations
-///
-/// Enumeration types for line search.
-///
-
-///
-/// \ingroup Enumerations
-///
-/// The enumeration of line search termination conditions.
-///
-enum LINE_SEARCH_TERMINATION_CONDITION
-{
-    ///
-    /// Backtracking method with the Armijo condition.
-    /// The backtracking method finds the step length such that it satisfies
-    /// the sufficient decrease (Armijo) condition,
-    /// \f$f(x + a \cdot d) \le f(x) + \beta' \cdot a \cdot g(x)^T d\f$,
-    /// where \f$x\f$ is the current point, \f$d\f$ is the current search direction,
-    /// \f$a\f$ is the step length, and \f$\beta'\f$ is the value specified by
-    /// \ref LBFGSParam::ftol. \f$f\f$ and \f$g\f$ are the function
-    /// and gradient values respectively.
-    ///
-    LBFGS_LINESEARCH_BACKTRACKING_ARMIJO = 1,
-
-    ///
-    /// The backtracking method with the defualt (regular Wolfe) condition.
-    /// An alias of `LBFGS_LINESEARCH_BACKTRACKING_WOLFE`.
-    ///
-    LBFGS_LINESEARCH_BACKTRACKING = 2,
-
-    ///
-    /// Backtracking method with regular Wolfe condition.
-    /// The backtracking method finds the step length such that it satisfies
-    /// both the Armijo condition (`LBFGS_LINESEARCH_BACKTRACKING_ARMIJO`)
-    /// and the curvature condition,
-    /// \f$g(x + a \cdot d)^T d \ge \beta \cdot g(x)^T d\f$, where \f$\beta\f$
-    /// is the value specified by \ref LBFGSParam::wolfe.
-    ///
-    LBFGS_LINESEARCH_BACKTRACKING_WOLFE = 2,
-
-    ///
-    /// Backtracking method with strong Wolfe condition.
-    /// The backtracking method finds the step length such that it satisfies
-    /// both the Armijo condition (`LBFGS_LINESEARCH_BACKTRACKING_ARMIJO`)
-    /// and the following condition,
-    /// \f$\vert g(x + a \cdot d)^T d\vert \le \beta \cdot \vert g(x)^T d\vert\f$,
-    /// where \f$\beta\f$ is the value specified by \ref LBFGSParam::wolfe.
-    ///
-    LBFGS_LINESEARCH_BACKTRACKING_STRONG_WOLFE = 3
-};
-
-///
-/// Parameters to control the L-BFGS algorithm.
-///
-template <typename Scalar = double>
-class LBFGSParam
-{
-public:
-    ///
-    /// The number of corrections to approximate the inverse Hessian matrix.
-    /// The L-BFGS routine stores the computation results of previous \ref m
-    /// iterations to approximate the inverse Hessian matrix of the current
-    /// iteration. This parameter controls the size of the limited memories
-    /// (corrections). The default value is \c 6. Values less than \c 3 are
-    /// not recommended. Large values will result in excessive computing time.
-    ///
-    int m;
-    ///
-    /// Absolute tolerance for convergence test.
-    /// This parameter determines the absolute accuracy \f$\epsilon_{abs}\f$
[…]

[…]
