### 2026-03-11 d75da4f
Nelder mead will not throw error anymore when x0 is empty, instead, it will sample ranodmly from the given bound
