### 2026-03-11 c17411f
Now l-bfgs-b does not throw error when x0 is empty
