### 2026-03-18 #10145 @monacofj
Copy that, affirmative.
I'm getting used to the review workflow...

On Wed, Mar 18, 2026 at 2:08 AM George K. Thiruvathukal <
***@***.***> wrote:

> *gkthiruvathukal* left a comment (openjournals/joss-reviews#10145)
> <https://github.com/openjournals/joss-reviews/issues/10145#issuecomment-4079725026>
>
> Hi @khoirulmuzakka <https://github.com/khoirulmuzakka> They should be
> here but I will nudge them! Thank you for your patience.
>
> @monacofj <https://github.com/monacofj>, @anugrahjo
> <https://github.com/anugrahjo> Can you please confirm that you are here?
>
