### 2026-03-22 #10145 @khoirulmuzakka
Hi @gkthiruvathukal  @monacofj @anugrahjo State of the field,  Software Design, Software Design, and AI usage disclosure  are not required when I submit the paper months ago. Should I update the paper first?
