### 2026-03-28 8efb7a7
Updated compile scripts
