### 2026-04-08 #10145 @anugrahjo
> Hi [@gkthiruvathukal](https://github.com/gkthiruvathukal) [@monacofj](https://github.com/monacofj) [@anugrahjo](https://github.com/anugrahjo) State of the field, Software Design, Software Design, and AI usage disclosure are not required when I submit the paper months ago. Should I update the paper first?

@khoirulmuzakka Apologies for the delayed response. The journal requirements have been updated since your original submission, and these sections are now required. Could you update the manuscript to include all required components? Without these, we won’t be able to complete the review checklist, so the review would remain incomplete.
