### 2026-04-12 #10145 @anugrahjo
@khoirulmuzakka @gkthiruvathukal

The software appears well-written and has the potential to support the development of new differential evolution algorithms. However, several issues need to be addressed before it is suitable for publication.

**Major comments (blocking):**
1. The installation instructions are incomplete for macOS and Python users (see issues [3](https://github.com/khoirulmuzakka/Minion/issues/3) and [4](https://github.com/khoirulmuzakka/Minion/issues/4)).
2. The contributing guidelines don't seem to be verified. **The instructions reference GitLab workflows, but this repository is hosted on GitHub. If AI was used to generate any part of the documentation or paper, all such use must be disclosed in an “AI usage disclosure” statement as per JOSS guidelines.**
3. There are no instructions for running tests locally (see issue [5](https://github.com/khoirulmuzakka/Minion/issues/5)). Please confirm if the CI pipeline includes tests for Python functionality or only covers the core C++ backend.
4. Add citations in the paper for the code used in the CEC benchmark suite.
5. The manuscript is missing four required sections per JOSS submission guidelines.
6. The repository history shows contributions only from the first author, while the paper lists three authors. Please clarify the contributions of each co-author to the software.
7. I encountered issues when running the `examples/minimizer.ipynb` notebook. Specifically, the multiprocessing sections fail to execute successfully on my system (see issue [6](https://github.com/khoirulmuzakka/Minion/issues/6)).
8. The README states that all problems from the CEC 2011 suite have been fully rewritten in C++ from the original MATLAB implementation. However, the [documentation](https://minion-py.readthedocs.io/en/latest/minionpy/minimizer.html#Algorithm-Comparisons-Using-CEC-Benchmark-Problems) still indicates that MATLAB is required and references `examples/cec_11.py`, which does not appear to exist in the repository.

**Minor comments:**

9. Pages 4-7 in the documentation reference four different sections within the same page. Please split them into 4 separate pages to avoid confusion.
10. CI and tests badges in the README should be consolidated, as it looks like the CI badge only checks the build workflows that do not perform any tests. CI should include some functional testing.
11. The paper title should mention "gradient-free" for specificity.
12. Include specific real-world application examples in the paper beyond CEC benchmarks to better motivate the work.

@gkthiruvathukal Based on the current state of the submission, there are still several blocking issues in the documentation, testing, and manuscript that should be addressed before further evaluation. Most importantly, there is no evidence of community engagement or demonstrated research impact (which seems like a necessary criterion for acceptance now).
