+++ paper/paper.md
@@ -26,33 +26,49 @@ bibliography: paper.bib
 
 # Summary
 
-Minion is a derivative-free optimization library for single-objective problems in which gradient-based methods are impractical. It provides a C++ backend with a Python interface (MinionPy), supporting applications in engineering, machine learning, and scientific computing.
+Minion is a derivative-free optimization library for single-objective blackbox problems where gradient information is unavailable or unreliable. It provides a C++ backend with a Python interface (MinionPy), supporting applications in engineering, machine learning, and scientific computing.
 
-The library offers a centralized implementation of state-of-the-art Differential Evolution (DE) algorithms that have performed strongly in IEEE CEC competitions. Alongside these research-grade solvers, Minion ships widely used optimizers such as Nelder–Mead, Dual Annealing (generalized simulated annealing), Covariance Matrix Adaptation Evolution Strategy (CMA-ES), and several Particle Swarm Optimization (PSO) variants, allowing practitioners to combine established baselines with advanced heuristics within a single API. Many existing optimization libraries include only elementary DE variants and lack standardized benchmark problems. Minion addresses this by integrating multiple CEC benchmark suites (2011, 2014, 2017, 2019, 2020, and 2022) to facilitate algorithm evaluation and comparison.
+The library offers a centralized implementation of advanced derivative-free optimization algorithms, including state-of-the-art Differential Evolution (DE) methods that have performed strongly in IEEE CEC competitions, CMA-ES variants, and several Particle Swarm Optimization (PSO) variants. Alongside these research-grade solvers, Minion ships widely used optimizers such as Nelder–Mead, Dual Annealing (generalized simulated annealing), L-BFGS, and L-BFGS-B, allowing practitioners to combine established baselines with advanced heuristics within a single API. Many existing optimization libraries include only elementary variants of these methods and lack standardized benchmark problems. Minion addresses this by integrating multiple CEC benchmark suites (2011, 2014, 2017, 2019, 2020, and 2022) to facilitate algorithm evaluation and comparison.
 
-Compared with widely adopted toolkits such as SciPy, NLopt, and pagmo2/pygmo, Minion emphasises a unified interface for batch-evaluated objective functions, provides native support for modern CEC-winning DE variants, and bundles curated benchmark suites for reproducible experimentation. These design choices serve researchers developing bespoke algorithms as well as practitioners seeking robust defaults for black-box optimisation.
+Compared with widely adopted toolkits such as SciPy, NLopt, and pagmo2/pygmo, Minion emphasizes a unified interface for batch-evaluated objective functions, provides native support for modern population-based methods and finite-difference quasi-Newton solvers, and bundles curated CEC benchmark suites for reproducible experimentation. These design choices serve researchers developing bespoke algorithms as well as practitioners seeking robust defaults for blackbox optimization.
 
-# Review of existing optimization libraries
+# State of the field
 
-SciPy [@2020SciPy-NMeth] underpins a large fraction of scientific computing in Python and offers a stable interface to classical optimisation routines, including gradient-based methods and a handful of derivative-free heuristics such as Nelder–Mead and Powell’s method. Its design prioritises broad accessibility and numerical reliability; consequently, coverage of recent population-based metaheuristics or fully vectorised objective evaluations is therefore limited.
+SciPy [@2020SciPy-NMeth] underpins a large fraction of scientific computing in Python and offers a stable interface to classical optimization routines, including gradient-based methods and a handful of derivative-free heuristics such as Nelder–Mead and Powell’s method. Its design prioritizes broad accessibility and numerical reliability; consequently, coverage of recent population-based metaheuristics or fully vectorized objective evaluations is limited.
 
[…]
