+++ docs/source/installation.rst
@@ -1,87 +1,270 @@
 Compilation and Installation
 ============================
 
-Python (MinionPy)
------------------
+Minion can be used in two ways:
 
-Install from PyPI:
+- as the Python package ``minionpy``;
+- as a native C++ library linked from another C++ project.
+
+For Python users, start with the PyPI package. If a wheel is not available for
+your platform or Python version, build the Python interface from source. For
+C++ users, use a release package when available, or build the native library
+with CMake.
+
+
+Python Package
+--------------
+
+Install MinionPy from PyPI:
 
 .. code-block:: shell
 
-   pip install --upgrade minionpy
+   python -m pip install --upgrade minionpy
 
-Quick check:
+Check the installation:
 
-.. code-block:: python
+.. code-block:: shell
 
-   import minionpy as mpy
+   python -c "import minionpy; print(minionpy.__version__)"
+
+PyPI wheels are not available for every operating system, Python version, or
+CPU architecture. If ``pip`` cannot find a compatible wheel, use
+`Build MinionPy From Source`_.
 
 
-Native Installation by Platform
--------------------------------
+Released Native Packages
+------------------------
 
-- **Windows**: compile Minion from source with CMake.
-- **Linux**: install the ``.deb`` package from GitHub Release assets.
-- **macOS**: install from release archive (``.tgz`` / ``tgz.zip`` asset).
-
-Linux ``.deb`` install:
+Linux users can install the ``.deb`` package from the GitHub release assets:
 
 .. code-block:: shell
 
    sudo dpkg -i minion_<version>_<arch>.deb
    sudo apt-get install -f
 
-macOS ``.tgz`` / ``tgz.zip`` install:
+macOS users can install from the release archive. If the downloaded asset is a
+``tgz.zip`` file, unzip it first:
 
 .. code-block:: shell
 
    unzip minion-<version>-macos.tgz.zip
+   mkdir -p /tmp/minion_pkg
    tar -xzf minion-<version>-macos.tgz -C /tmp/minion_pkg
    sudo rsync -a /tmp/minion_pkg/ /usr/local/
 
+Windows users should build from source with CMake.
 
-Compilation From Scratch
-------------------------
 
-Build requirements:
+Source Build Requirements
+-------------------------
 
-- CMake >= 3.18
-- C++17 compiler (GCC/Clang/MSVC)
-- Eigen3 (or automatic fetch through CMake)
-- Python 3 + pybind11 (only when building Python bindings)
+Required for native C++ builds:
 
-Build with helper scripts:
+- Git, when cloning the repository;
+- CMake >= 3.18;
+- a C++17 compiler, such as GCC, Clang, or MSVC;
+- Eigen3, or network access so CMake can fetch Eigen automatically.
+
+Required for Python builds, including the default helper-script builds:
+
+- Python 3.9 or newer;
+- Python development headers;
+- ``pybind11``.
+
+Optional for documentation builds:
+
+- Doxygen;
+- Pandoc;
+- Sphinx;
+- sphinx-rtd-theme;
+- nbsphinx;
+- Breathe.
+
+On Ubuntu 24.04, install the native build tools with:
+
+.. code-block:: shell
+
+   sudo apt update
+   sudo apt install git cmake build-essential
+
+For Python builds on Ubuntu 24.04, also install:
+
+.. code-block:: shell
+
+   sudo apt install python3-dev python3-pip python3-pybind11
+
+For documentation builds on Ubuntu 24.04, also install:
+
+.. code-block:: shell
+
+   sudo apt install doxygen pandoc python3-sphinx python3-sphinx-rtd-theme \
[…]
