+++ tests/test_minionpy.py
@@ -1,0 +1,308 @@
+from __future__ import annotations
+
+import math
+import os
+import sys
+import numpy as np
+
+REPO_ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
+if REPO_ROOT not in sys.path:
+    sys.path.insert(0, REPO_ROOT)
+
+import minionpy
+
+
+ALGORITHMS = (
+    "DE",
+    "LSHADE",
+    "AGSK",
+    "JADE",
+    "j2020",
+    "NLSHADE_RSP",
+    "LSRTDE",
+    "jSO",
+    "IMODE",
+    "ARRDE",
+    "GWO_DE",
+    "NelderMead",
+    "ABC",
+    "PSO",
+    "SPSO2011",
+    "DMSPSO",
+    "LSHADE_cnEpSin",
+    "CMAES",
+    "RCMAES",
+    "BIPOP_aCMAES",
+    "DA",
+    "L_BFGS_B",
+    "L_BFGS",
+)
+
+SPHERE_UPPER = {algorithm: 1e-3 for algorithm in ALGORITHMS}
+ROSENBROCK_UPPER = {algorithm: 10.0 for algorithm in ALGORITHMS}
+EVAL_SLACK = 100
+
+
+def sphere_batch(x_batch):
+    return minionpy.sphere(np.asarray(x_batch, dtype=float)).tolist()
+
+
+def rosenbrock_batch(x_batch):
+    return minionpy.rosenbrock(np.asarray(x_batch, dtype=float)).tolist()
+
+
+def make_cec2017_batch(function_number: int, dimension: int):
+    cec2017 = minionpy.CEC2017Functions(function_number, dimension)
+
+    def objective(x_batch):
+        return list(cec2017(x_batch))
+
+    return objective
+
+
+def _run_case(
+    func,
+    bounds,
+    x0,
+    algorithm: str,
+    maxevals: int,
+    seed: int,
+    quality_upper: float | None = None,
+):
+    result = minionpy.Minimizer(
+        func=func,
+        bounds=bounds,
+        x0=x0,
+        algo=algorithm,
+        relTol=1e-8,
+        maxevals=maxevals,
+        seed=seed,
+    ).optimize()
+
+    finite_ok = math.isfinite(result.fun)
+    eval_ok = result.nfev <= maxevals + EVAL_SLACK
+    quality_ok = True if quality_upper is None else result.fun <= quality_upper
+    passed = finite_ok and eval_ok and quality_ok
+
+    return {
+        "algorithm": algorithm,
+        "fun": float(result.fun),
+        "nfev": int(result.nfev),
+        "passed": passed,
+        "message": (
+            f"finite={finite_ok} nfev={result.nfev} "
+            f"limit={maxevals + EVAL_SLACK} best_f={result.fun}"
+            + ("" if quality_upper is None else f" quality_limit={quality_upper}")
+        ),
+    }
+
+
+def run_sphere_suite(
+    dimension: int = 5,
+    maxevals: int = 4000,
+    seed: int = 42,
+):
+    bounds = [(-5.0, 5.0)] * dimension
+    x0 = [[0.5] * dimension]
+    records = []
+
+    for algorithm in ALGORITHMS:
+        record = _run_case(
+            func=sphere_batch,
+            bounds=bounds,
+            x0=x0,
+            algorithm=algorithm,
+            maxevals=maxevals,
+            seed=seed,
+            quality_upper=SPHERE_UPPER[algorithm],
+        )
+        records.append(record)
+
+    return records
+
+
+def run_rosenbrock_suite(
+    dimension: int = 5,
+    maxevals: int = 4000,
+    seed: int = 42,
+):
[…]
