+++ docs/source/installation.rst
@@ -35,22 +35,64 @@ CPU architecture. If ``pip`` cannot find a compatible wheel, use
 Released Native Packages
 ------------------------
 
-Linux users can install the ``.deb`` package from the GitHub release assets:
+Linux users can install the native ``.deb`` package from the GitHub release
+assets:
 
 .. code-block:: shell
 
    sudo dpkg -i minion_<version>_<arch>.deb
    sudo apt-get install -f
 
-macOS users can install from the release archive. If the downloaded asset is a
-``tgz.zip`` file, unzip it first:
+macOS users can install the native C++ package from the ``.tar.gz`` archive
+published in the GitHub release assets. Choose the archive that matches your
+machine:
+
+- Apple Silicon: ``minion-arm64.tar.gz``
+- Intel: ``minion-x64.tar.gz``
+
+These archives are generated by CPack in CI.
+
+On macOS, Safari may automatically extract a downloaded ``.tar.gz`` archive.
+If that happens, use the extracted folder directly instead of running ``tar``
+again, or disable Safari's ``Open "safe" files after downloading`` setting.
+
+For a user-local installation, first extract the archive into a temporary
+directory:
 
 .. code-block:: shell
 
-   unzip minion-<version>-macos.tgz.zip
+   mkdir -p /tmp/minion_pkg "$HOME/.local/minion"
+   tar -xzf minion-<arch>.tar.gz -C /tmp/minion_pkg
+   rsync -a /tmp/minion_pkg/minion-*/ "$HOME/.local/minion/"
+
+For a system-wide installation, extract into a temporary directory and then
+copy into the system prefix:
+
+.. code-block:: shell
+
    mkdir -p /tmp/minion_pkg
-   tar -xzf minion-<version>-macos.tgz -C /tmp/minion_pkg
-   sudo rsync -a /tmp/minion_pkg/ /usr/local/
+   tar -xzf minion-<arch>.tar.gz -C /tmp/minion_pkg
+   sudo rsync -a /tmp/minion_pkg/minion-*/ /usr/local/
+
+This native package installs headers, shared libraries, CMake package files,
+and the CEC input-data directory. It does not install the Python package
+``minionpy``.
+
+If you install into a non-default prefix such as ``$HOME/.local/minion``, add
+that prefix to ``CMAKE_PREFIX_PATH`` when configuring your project.
+
+For example:
+
+.. code-block:: shell
+
+   cmake -S . -B build -DCMAKE_PREFIX_PATH="$HOME/.local/minion"
+
+If you use this installation often, you can also make the setting persistent in
+your shell profile, for example in ``~/.zshrc`` or ``~/.bashrc``:
+
+.. code-block:: shell
+
+   export CMAKE_PREFIX_PATH="$HOME/.local/minion:$CMAKE_PREFIX_PATH"
 
 Windows users should build from source with CMake.
 
