+++ docs/source/installation.rst
@@ -47,8 +47,8 @@ macOS users can install the native C++ package from the ``.tar.gz`` archive
 published in the GitHub release assets. Choose the archive that matches your
 machine:
 
-- Apple Silicon: ``minion-arm64.tar.gz``
-- Intel: ``minion-x64.tar.gz``
+- Apple Silicon: ``minion-<version>-macos-arm64.tar.gz``
+- Intel: ``minion-<version>-macos-x64.tar.gz``
 
 These archives are generated by CPack in CI.
 
@@ -62,7 +62,7 @@ directory:
 .. code-block:: shell
 
    mkdir -p /tmp/minion_pkg "$HOME/.local/minion"
-   tar -xzf minion-<arch>.tar.gz -C /tmp/minion_pkg
+   tar -xzf minion-<version>-macos-<arch>.tar.gz -C /tmp/minion_pkg
    rsync -a /tmp/minion_pkg/minion-*/ "$HOME/.local/minion/"
 
 For a system-wide installation, extract into a temporary directory and then
@@ -71,7 +71,7 @@ copy into the system prefix:
 .. code-block:: shell
 
    mkdir -p /tmp/minion_pkg
-   tar -xzf minion-<arch>.tar.gz -C /tmp/minion_pkg
+   tar -xzf minion-<version>-macos-<arch>.tar.gz -C /tmp/minion_pkg
    sudo rsync -a /tmp/minion_pkg/minion-*/ /usr/local/
 
 This native package installs headers, shared libraries, CMake package files,
