+++ CMakeLists.txt
@@ -291,6 +291,11 @@ set(CPACK_PACKAGE_NAME "minion")
 set(CPACK_PACKAGE_VENDOR "Minion")
 set(CPACK_PACKAGE_VERSION ${PROJECT_VERSION})
 set(CPACK_PACKAGE_DESCRIPTION_SUMMARY "Minion optimization library")
+set(MINION_PACKAGE_SUFFIX "" CACHE STRING "Optional suffix appended to the CPack package file name")
+set(CPACK_PACKAGE_FILE_NAME "${CPACK_PACKAGE_NAME}-${CPACK_PACKAGE_VERSION}")
+if(MINION_PACKAGE_SUFFIX)
+    string(APPEND CPACK_PACKAGE_FILE_NAME "-${MINION_PACKAGE_SUFFIX}")
+endif()
 if(UNIX AND NOT APPLE)
     set(CPACK_GENERATOR "TGZ;DEB")
     set(CPACK_DEBIAN_FILE_NAME DEB-DEFAULT)
