+++ docs/source/installation.rst
@@ -35,6 +35,16 @@ CPU architecture. If ``pip`` cannot find a compatible wheel, use
 Released Native Packages
 ------------------------
 
+To use an installed native C++ package from your own CMake project, make sure
+``Eigen3`` is also installed and discoverable by CMake.
+
+On Ubuntu:
+
+.. code-block:: shell
+
+   sudo apt update
+   sudo apt install libeigen3-dev
+
 Linux users can install the native ``.deb`` package from the GitHub release
 assets:
 
@@ -52,40 +62,51 @@ machine:
 
 These archives are generated by CPack in CI.
 
+On macOS with Homebrew:
+
+.. code-block:: shell
+
+   brew install eigen
+
 On macOS, Safari may automatically extract a downloaded ``.tar.gz`` archive.
 If that happens, use the extracted folder directly instead of running ``tar``
 again, or disable Safari's ``Open "safe" files after downloading`` setting.
 
-For a user-local installation, first extract the archive into a temporary
-directory:
+Install into a user-local prefix such as ``$HOME/.local/minion``:
 
 .. code-block:: shell
 
    mkdir -p /tmp/minion_pkg "$HOME/.local/minion"
    tar -xzf minion-<version>-macos-<arch>.tar.gz -C /tmp/minion_pkg
-   rsync -a /tmp/minion_pkg/minion-<version>-macos-<arch>/ "$HOME/.local/minion/"
-
-For a system-wide installation, extract into a temporary directory and then
-copy into the system prefix:
-
-.. code-block:: shell
-
-   mkdir -p /tmp/minion_pkg
-   tar -xzf minion-<version>-macos-<arch>.tar.gz -C /tmp/minion_pkg
-   sudo rsync -a /tmp/minion_pkg/minion-<version>-macos-<arch>/ /usr/local/
+   ditto /tmp/minion_pkg/minion-<version>-macos-<arch> "$HOME/.local/minion"
 
 This native package installs headers, shared libraries, CMake package files,
 and the CEC input-data directory. It does not install the Python package
 ``minionpy``.
 
-If you install into a non-default prefix such as ``$HOME/.local/minion``, add
-that prefix to ``CMAKE_PREFIX_PATH`` when configuring your project.
+Because the macOS package is unsigned, Gatekeeper may block the installed
+libraries the first time you use them. If that happens, remove the quarantine
+attribute from the local install:
+
+.. code-block:: shell
+
+   xattr -dr com.apple.quarantine "$HOME/.local/minion"
+
+To let CMake find this local installation, add the prefix to
+``CMAKE_PREFIX_PATH`` when configuring your project.
 
 For example:
 
 .. code-block:: shell
 
    cmake -S . -B build -DCMAKE_PREFIX_PATH="$HOME/.local/minion"
+
+On Apple Silicon with Homebrew ``Eigen3``, you may need to include the
+Homebrew prefix as well:
+
+.. code-block:: shell
+
+   cmake -S . -B build -DCMAKE_PREFIX_PATH="$HOME/.local/minion;/opt/homebrew"
 
 If you use this installation often, you can also make the setting persistent in
 your shell profile, for example in ``~/.zshrc`` or ``~/.bashrc``:
