+++ paper/paper.bib
@@ -308,6 +308,20 @@ URL = {https://doi.org/10.1214/aoms/1177730491}
 }
 
 
+@article{MUZAKKA2026166076,
+title = {AutoNRA: A high-throughput parallel tool for self-consistent ion beam analysis},
+journal = {Nuclear Instruments and Methods in Physics Research Section B: Beam Interactions with Materials and Atoms},
+volume = {574},
+pages = {166076},
+year = {2026},
+issn = {0168-583X},
+doi = {10.1016/j.nimb.2026.166076},
+url = {https://www.sciencedirect.com/science/article/pii/S0168583X26000777},
+author = {K.F. Muzakka and S. Möller and S.A. Afzal and M. Bram and M. Finsterbusch},
+keywords = {Ion beam analysis, Rutherford backscattering spectrometry, Nuclear reaction analysis},
+abstract = {We present AutoNRA, a high-throughput framework for self-consistent ion beam analysis (IBA). Developed to meet the growing demand for automated processing of large-scale datasets, AutoNRA enables parallel analysis of spectra from thousands of sample points, each potentially containing multiple IBA measurements such as Rutherford Backscattering Spectrometry (RBS) and Nuclear Reaction Analysis (NRA). By exploiting multiprocessing, it fully utilizes available CPU cores and achieves speed-ups of up to two orders of magnitude compared to conventional sequential SIMNRA workflows. This acceleration makes it particularly suited for spatially resolved applications, including three-dimensional compositional tomography. AutoNRA ensures model consistency by maintaining shared parameters across spectra from the same sample point while performing independent, parallel evaluations across different points. The framework also provides a model-agnostic machine learning interface, allowing seamless integration of surrogate models for accelerated spectrum prediction and adaptive optimization. Its performance and scalability are demonstrated on BZCY721(BaZr0.7Ce0.2Y0.1O3−δ) sample fabricated by cold sintering, where AutoNRA enabled quantitative mapping of elemental distributions and detection of residual heavy water within the structure.}
+}
+
 @techreport{cec2014,
   author = {J.J.Liang and and B.Y.Qu and P.N.Suganthan},
   title = {{Problem Definitions and Evaluation Criteria for the CEC 2014 Special Session and Competition on Single Objective Real-Parameter Numerical Optimization}},
@@ -462,4 +476,51 @@ numpages = {8},
 keywords = {evolutionary computation, black-box optimization, benchmarking, CMA-ES},
 location = {Montreal, Qu\'{e}bec, Canada},
 series = {GECCO '09}
+}
+
+
+@techreport{cec2017,
+  author = {N.H.Awad and M.Z.Ali and P.N.Suganthan and J.J.Liang and and B.Y.Qu},
+  title = {{Problem Definitions and Evaluation Criteria for the CEC 2017 Special Session and Competition on Single Objective Real-Parameter Numerical Optimization}},
+  institution = {Nanyang Technological University, Jordan University of Science and Technology, and Zhengzhou University},
+  year = {2016},
+  note = {Available at \url{https://github.com/P-N-Suganthan/CEC2017-BoundContrained}}
+}
+
+
+@techreport{cec2019,
+	author = {K.V.Price and N.H. Awad and M.Z.Ali and P.N.Suganthan},
+	title = {{Problem Definitions and Evaluation Criteria for the 100-Digit
+	Challenge Special Session and Competition on Single Objective
+	Numerical Optimization}},
+	institution = {Nanyang Technological University, and Jordan University of Science and Technology},
+	year = {2016},
[…]
