+++ paper/paper.md
@@ -1,11 +1,13 @@
 ---
-title: 'Minion: A C++ and Python Library for Single-Objective Optimization Algorithms'
+title: 'Minion: A C++ and Python Library for Single-Objective Optimization'
 tags:
-  - C++ 
-  - Python 
-  - Optimization
+  - Blackbox Optimization
+  - Derivative-Free Optimization
+  - Evolutionary Algorithms
   - Differential Evolution
   - CEC Benchmark Problems
+  - Python
+  - C++
 authors:
   - name: Khoirul Faiq Muzakka
     orcid: 0000-0002-3888-1697
@@ -26,7 +28,7 @@ bibliography: paper.bib
 
 # Summary
 
-Minion is a derivative-free optimization library for single-objective blackbox problems where gradient information is unavailable or unreliable. It provides a C++ backend with a Python interface (MinionPy), supporting applications in engineering, machine learning, and scientific computing.
+Minion is a derivative-free optimization library for unconstrained and bound-constrained single-objective blackbox problems, where users provide objective-function evaluations rather than analytic gradients. It provides a C++ backend with a Python interface (MinionPy), supporting applications in engineering, machine learning, and scientific computing.
 
 The library offers a centralized implementation of advanced derivative-free optimization algorithms, including state-of-the-art Differential Evolution (DE) methods that have performed strongly in IEEE Congress on Evolutionary Computation (CEC) competitions, Covariance Matrix Adaptation Evolution Strategy (CMA-ES) variants, and several Particle Swarm Optimization (PSO) variants. Alongside these research-grade solvers, Minion ships widely used optimizers such as Nelder–Mead, Dual Annealing (generalized simulated annealing), L-BFGS, and L-BFGS-B, allowing practitioners to combine established baselines with advanced heuristics within a single API. Many existing optimization libraries include only elementary variants of these methods and lack standardized benchmark problems. Minion addresses this by integrating multiple CEC benchmark suites (2011, 2014, 2017, 2019, 2020, and 2022) to facilitate algorithm evaluation and comparison.
 
@@ -93,9 +95,9 @@ Minion currently implements the following optimization algorithms:
 - L-BFGS-B [@doi:10.1137/0916069]
 - L-BFGS [@Liu1989]
 
-Additional algorithms are planned for future releases. Minion also ships benchmark suites from CEC competitions spanning 2011, 2014, 2017, 2019, 2020, and 2022. The library further bundles classic analytic test functions—such as sphere, Rosenbrock, and Rastrigin—for quick experimentation and unit testing.
+Additional algorithms are planned for future releases. Minion also ships benchmark suites from CEC competitions spanning 2011 [@cec2011], 2014 [@cec2014], 2017 [@cec2017], 2019 [@cec2019], 2020 [@cec2020], and 2022 [@cec2022]. The library further bundles classic analytic test functions—such as sphere, Rosenbrock, and Rastrigin—for quick experimentation and unit testing.
 
[…]
