+++ paper/paper.bib
@@ -90,7 +90,7 @@ abstract = {Since its inception in 1995, Differential Evolution (DE) has emerged
 
 @misc{ARRDE,
       title={Robust Differential Evolution via Nonlinear Population Size Reduction and Adaptive Restart: The ARRDE Algorithm}, 
-      author={Khoirul Faiq Muzakka and Ahsani Hafizhu Shali and Haris Suhendar and Sören Möller and Martin Finsterbusch},
+      author={Muzakka, K. F. and Shali, A. H. and Suhendar, H. and Möller, S. and Finsterbusch, M. },
       year={2026},
       eprint={2511.18429},
       archivePrefix={arXiv},
@@ -273,9 +273,7 @@ abstract = {Since its inception in 1995, Differential Evolution (DE) has emerged
 
 
 @software{muzakka_2025_14893994,
-  author       = {Muzakka, Khoirul Faiq and
-                  Möller, Sören and
-                  Finsterbusch, Martin},
+  author       = {Muzakka, K. F. and Möller, S. and Finsterbusch, M.},
   title        = {Minion : a high-performance derivative-free
                    optimization library designed for solving complex
                    optimization problems.
@@ -317,7 +315,7 @@ year = {2026},
 issn = {0168-583X},
 doi = {10.1016/j.nimb.2026.166076},
 url = {https://www.sciencedirect.com/science/article/pii/S0168583X26000777},
-author = {K.F. Muzakka and S. Möller and S.A. Afzal and M. Bram and M. Finsterbusch},
+author = {Muzakka, K. F. and Möller, S. and Afzal, S. A. and Bram, M. and Finsterbusch, M.},
 keywords = {Ion beam analysis, Rutherford backscattering spectrometry, Nuclear reaction analysis},
 abstract = {We present AutoNRA, a high-throughput framework for self-consistent ion beam analysis (IBA). Developed to meet the growing demand for automated processing of large-scale datasets, AutoNRA enables parallel analysis of spectra from thousands of sample points, each potentially containing multiple IBA measurements such as Rutherford Backscattering Spectrometry (RBS) and Nuclear Reaction Analysis (NRA). By exploiting multiprocessing, it fully utilizes available CPU cores and achieves speed-ups of up to two orders of magnitude compared to conventional sequential SIMNRA workflows. This acceleration makes it particularly suited for spatially resolved applications, including three-dimensional compositional tomography. AutoNRA ensures model consistency by maintaining shared parameters across spectra from the same sample point while performing independent, parallel evaluations across different points. The framework also provides a model-agnostic machine learning interface, allowing seamless integration of surrogate models for accelerated spectrum prediction and adaptive optimization. Its performance and scalability are demonstrated on BZCY721(BaZr0.7Ce0.2Y0.1O3−δ) sample fabricated by cold sintering, where AutoNRA enabled quantitative mapping of elemental distributions and detection of residual heavy water within the structure.}
 }
