+++ paper/paper.bib
@@ -477,48 +477,3 @@ series = {GECCO '09}
 }
 
 
-@techreport{cec2017,
-  author = {N.H.Awad and M.Z.Ali and P.N.Suganthan and J.J.Liang and and B.Y.Qu},
-  title = {{Problem Definitions and Evaluation Criteria for the CEC 2017 Special Session and Competition on Single Objective Real-Parameter Numerical Optimization}},
-  institution = {Nanyang Technological University, Jordan University of Science and Technology, and Zhengzhou University},
-  year = {2016},
-  note = {Available at \url{https://github.com/P-N-Suganthan/CEC2017-BoundContrained}}
-}
-
-
-@techreport{cec2019,
-	author = {K.V.Price and N.H. Awad and M.Z.Ali and P.N.Suganthan},
-	title = {{Problem Definitions and Evaluation Criteria for the 100-Digit
-	Challenge Special Session and Competition on Single Objective
-	Numerical Optimization}},
-	institution = {Nanyang Technological University, and Jordan University of Science and Technology},
-	year = {2016},
-	note = {Available at \url{https://github.com/P-N-Suganthan/CEC2017-BoundContrained}}
-}
-
-@techreport{cec2020,
-  author = {C. T. Yue and K. V. Price and P. N. Suganthan and J. J. Liang and M. Z. Ali and B. Y. Qu and N. H. Awad and P. P. Biswas},
-  title = {{Problem Definitions and Evaluation Criteria for the CEC 2020
-Special Session and Competition on Single Objective Bound
-Constrained Numerical Optimization}},
-  institution = {Zhengzhou University and Nanyang Technological University},
-  year = {2019},
-  note = {Available at \url{https://github.com/P-N-Suganthan/2020-Bound-Constrained-Opt-Benchmark}}
-}
-
-
-@techreport{cec2022,
-  author = {A. Kumar and K. V. Price and A. W. Mohamed and A. A. Hadi and P. N. Suganthan},
-  title = {{Problem Definitions and Evaluation Criteria for the CEC 2022 Special Session and Competition on Single Objective Bound Constrained Numerical Optimization}},
-  institution = {},
-  year = {2021},
-  note = {Available at \url{https://github.com/P-N-Suganthan/2022-SO-BO}}
-}
-
-@inproceedings{cec2011,
-  title={{Problem Definitions and Evaluation Criteria for CEC 2011 Competition on Testing Evolutionary Algorithms on Real World Optimization Problems}},
-  author={Swagatam Das and Ponnuthurai Nagaratnam Suganthan},
-  booktitle={Proceedings of the IEEE Congress on Evolutionary Computation (CEC)},
-  year={2011},
-  url={https://api.semanticscholar.org/CorpusID:14174334}
-}
