+++ README.md
@@ -54,169 +54,20 @@ Minion also serves as a **research platform** for developing and testing new opt
 
 ## 🚀 Installation and Usage
 
-### Python
-Install from PyPI:
+### Quick install
 
 ```sh
 pip install --upgrade minionpy
 ```
 
-### C++ / Native package
-- Windows: compile from source with CMake (see "Compilation From Scratch" and "Using Minion in a C++ Project" below).
-- Linux: install the `.deb` package from GitHub Release assets.
-- macOS: use the release archive (`.tgz` / `tgz.zip` asset).
-
-#### Linux `.deb` install
-```sh
-sudo dpkg -i minion_<version>_<arch>.deb
-sudo apt-get install -f
-```
-
-#### macOS `.tgz` / `tgz.zip` install
-If the asset is `tgz.zip`, unzip first:
-```sh
-unzip minion-<version>-macos.tgz.zip
-```
-
-Extract package content:
-```sh
-tar -xzf minion-<version>-macos.tgz -C /tmp/minion_pkg
-```
-
-Install manually (example to `/usr/local`):
-```sh
-sudo rsync -a /tmp/minion_pkg/ /usr/local/
-```
-
-### Compilation From Scratch
-Dependencies:
+### Build dependencies
 - CMake >= 3.18
 - C++17 compiler (GCC/Clang/MSVC)
 - Eigen3 (or allow automatic fetch via CMake)
 - Optional for Python bindings: Python 3 + `pybind11`
 - Optional for documentation: `doxygen`, `pandoc`, Python 3, `sphinx`, `sphinx-rtd-theme`, `nbsphinx`, `breathe`
 
-Build with helper scripts:
-- Windows: `compile.bat`
-- Linux/macOS: `compile.sh`
-
-The helper scripts are intended for native C++ builds and configure:
-- `MINION_BUILD_CEC=ON`
-- `MINION_BUILD_EXAMPLES=ON`
-- `MINION_BUILD_PYTHON=OFF`
-
-If the optional documentation toolchain is already installed, the scripts also generate Doxygen and HTML docs. If any docs dependency is missing, the scripts print a warning and skip documentation generation.
-
-Manual CMake build:
-```sh
-cmake -S . -B build \
-  -DMINION_BUILD_CEC=ON \
-  -DMINION_BUILD_PYTHON=OFF \
-  -DMINION_BUILD_EXAMPLES=ON
-cmake --build build --config Release
-cmake --install build --prefix /usr/local
-```
-
-Manual CMake build with Python bindings:
-```sh
-cmake -S . -B build \
-  -DMINION_BUILD_CEC=ON \
-  -DMINION_BUILD_PYTHON=ON \
-  -DMINION_BUILD_EXAMPLES=ON
-cmake --build build --config Release
-```
-
-For Python usage, installing from PyPI is the recommended path:
-```sh
-pip install --upgrade minionpy
-```
-
-### Using Minion in a C++ Project
-Header include:
-```cpp
-#include <minion/minion.h>
-```
-
-CMake with `find_package` first, then `FetchContent` fallback:
-
-```cmake
-cmake_minimum_required(VERSION 3.18)
-project(my_app LANGUAGES CXX)
-set(CMAKE_CXX_STANDARD 17)
-set(CMAKE_CXX_STANDARD_REQUIRED ON)
-
-include(FetchContent)
-
-find_package(Minion QUIET CONFIG)
-if(NOT Minion_FOUND)
-    FetchContent_Declare(
-        minion
-        GIT_REPOSITORY https://github.com/khoirulmuzakka/Minion.git
-        GIT_TAG main
-        GIT_SHALLOW TRUE
-    )
-    set(MINION_BUILD_CEC ON CACHE BOOL "Build CEC library" FORCE)
[…]
