+++ docs/source/index.rst
@@ -20,6 +20,9 @@ This documentation covers the installation, usage, and details of both the C++ a
    installation
    minion/index
    minionpy/index
+   BFGS
+   Convergence
+   Vectorization
    Algorithms
    contribute
    changelog
