### 2026-04-29 #10145 @khoirulmuzakka
Dear @gkthiruvathukal, @anugrahjo, and @monacofj,

Thank you very much for your time and effort in reviewing my submission and for testing Minion. I greatly appreciate your detailed and constructive feedback. I have carefully reviewed all comments, created corresponding GitHub issues, and believe that the concerns raised have now been addressed. Please find a summary of the revisions below:

1. Installation instructions
   The installation guide has been expanded to include macOS and manual compilation of minionpy from source (for unsupported platforms or Python versions). I have verified these steps on macOS (Apple Silicon), and they work as intended.

2. Contribution guidelines and testing
   The contribution guidelines have been completely rewritten. Unit tests now cover both the C++ backend and Python interface. Instructions for running tests locally have been added. The CI pipeline has also been updated to include these tests.

3. Manuscript updates
   The manuscript has been revised to include all required sections per JOSS guidelines. Citations for the CEC benchmark suite and additional relevant references have been added.

4. Author contributions
   M. Finsterbusch and S. Möller contributed through funding acquisition and general supervision of the project. While they were not directly involved in implementation, their roles have been essential to the development and direction of the software.

5. Multiprocessing issue in Jupyter notebooks
   The issue encountered in `examples/minimizer.ipynb` is related to platform-specific multiprocessing behavior in Jupyter environments (notably on Windows and macOS). This limitation is documented, and guidance has been added: multiprocessing examples should be executed in a standalone Python script within an `if __name__ == "__main__":` block. The documentation has been updated accordingly.

6. MATLAB dependency clarification
   References to MATLAB have been removed from the documentation, as it is no longer required.

7. CI and testing badges
   CI workflows have been consolidated, and testing is now integrated into the CI pipeline to ensure functional verification beyond build checks.

8. Real-world application examples
   A curve-fitting example was already included in the documentation to demonstrate practical usage beyond benchmark problems.

9. Paper title and scope
   While the term “gradient-free” was considered, it was not included in the title to avoid restricting the perceived scope of the library.  Minion currently includes methods such as L-BFGS and L-BFGS-B, which are essentially a gradient-based method,  where the gradient is numerically estimated. The manuscript clarifies that the focus is on single-objective, black-box, unconstrained, and bound-constrained optimization.

10. Documentation structure
    The previously combined sections (pages 4–7) have been split into separate pages to improve clarity and navigation.

Community engagement and impact

At present, Minion has primarily been used within our research group. However, it has already supported applications in nuclear physics (as cited in the manuscript) and optimization research, including the ARRDE and RCMAES algorithms (the latter accepted at the IEEE CEC conference this year). Given the broad relevance of unconstrained and bound-constrained optimization problems across scientific and engineering domains, we believe the software has strong potential for wider adoption.

Thank you again for your valuable feedback. I hope the revisions satisfactorily address your concerns, and I look forward to your further evaluation.
