+++ minion/include/default_options.h
@@ -256,6 +256,8 @@ class DefaultSettings{
                 {"LSRTDE", default_settings_LSRTDE}, 
                 {"ARRDE", default_settings_ARRDE}, 
                 {"jSO", default_settings_jSO}, 
+                {"jSO_1", default_settings_jSO},
+                {"jSO_2", default_settings_jSO},
                 {"IMODE", default_settings_IMODE},
                 {"AGSK", default_settings_AGSK},
                 {"GWO_DE", default_settings_GWO_DE}, 
