+++ minion/include/minimizer.h
@@ -6,6 +6,7 @@
 #include "j2020.h"
 #include "jade.h"
 #include "jso.h"
+#include "jso_ablation.h"
 #include "lsrtde.h"
 #include "nelder_mead.h"
 #include "lshade.h"
@@ -53,7 +54,7 @@ class Minimizer {
          * @param x0 The initial guesses for the solution. Note that Minion assumes multiple initial guesses, thus, x0 is an std::vector<std::vector<double>> object. These guesses will be used for population initialization in the population-population based algorithms, or minion will pick teh best one in L-BFGS or NelderMead.
          * @param data Additional data to pass to the objective function.
          * @param callback A callback function to call after each iteration.
-         * @param algo Algorithm to use : "LSHADE", "AGSK", "DE", "JADE", "jSO", "IMODE", "NelderMead", "LSRTDE", "NLSHADE_RSP", "j2020", "GWO_DE", "PSO", "SPSO2011", "DMSPSO", "LSHADE_cnEpSin"
+         * @param algo Algorithm to use : "LSHADE", "AGSK", "DE", "JADE", "jSO", "jSO_1", "jSO_2", "IMODE", "NelderMead", "LSRTDE", "NLSHADE_RSP", "j2020", "GWO_DE", "PSO", "SPSO2011", "DMSPSO", "LSHADE_cnEpSin"
          * @param tol The relative tolerance for convergence.
          * @param maxevals The maximum number of function evaluations.
          * @param seed global seed
@@ -81,6 +82,8 @@ class Minimizer {
             else if (algoUpper == "NLSHADE_RSP") optimizer = new NLSHADE_RSP(func, bounds, x0, data, callback, tol, maxevals, seed, options);
             else if (algoUpper == "LSRTDE") optimizer = new LSRTDE(func, bounds, x0, data, callback, tol, maxevals, seed, options);
             else if (algoUpper == "JSO") optimizer = new jSO (func, bounds, x0, data, callback, tol, maxevals, seed, options);
+            else if (algoUpper == "JSO_1") optimizer = new jSO_1(func, bounds, x0, data, callback, tol, maxevals, seed, options);
+            else if (algoUpper == "JSO_2") optimizer = new jSO_2(func, bounds, x0, data, callback, tol, maxevals, seed, options);
             else if (algoUpper == "IMODE") optimizer = new IMODE(func, bounds, x0, data, callback, tol, maxevals, seed, options);
             else if (algoUpper == "ARRDE") optimizer = new ARRDE (func, bounds, x0, data, callback, tol, maxevals, seed, options);
             else if (algoUpper == "GWO_DE") optimizer = new GWO_DE(func, bounds, x0, data, callback, tol, maxevals, seed, options);
@@ -107,7 +110,7 @@ class Minimizer {
          * @param x0 The initial guess for the solution.
          * @param data Additional data to pass to the objective function.
          * @param callback A callback function to call after each iteration.
-         * @param algo Algorithm to use : "LSHADE", "DE", "JADE", "jSO", "DE", "NelderMead", "LSRTDE", "NLSHADE_RSP", "j2020", "GWO_DE", "PSO", "SPSO2011", "DMSPSO", "LSHADE_cnEpSin"
+         * @param algo Algorithm to use : "LSHADE", "DE", "JADE", "jSO", "jSO_1", "jSO_2", "DE", "NelderMead", "LSRTDE", "NLSHADE_RSP", "j2020", "GWO_DE", "PSO", "SPSO2011", "DMSPSO", "LSHADE_cnEpSin"
          * @param tol The relative tolerance for convergence.
          * @param maxevals The maximum number of function evaluations.
          * @param seed global seed
[…]
