+++ minion/src/jso_ablation.cpp
@@ -1,0 +1,630 @@
+#include "jso_ablation.h"
+
+#include "default_options.h"
+
+namespace minion {
+
+void jSOAblationBase::initializeCommon() {
+    auto defaultKey = DefaultSettings().getDefaultSettings("jSO");
+    for (const auto& el : optionMap) {
+        defaultKey[el.first] = el.second;
+    }
+    Options options(defaultKey);
+
+    boundStrategy = options.get<std::string>("bound_strategy", "reflect-random");
+    const std::vector<std::string> all_boundStrategy = {"random", "reflect", "reflect-random", "clip", "periodic", "none"};
+    if (std::find(all_boundStrategy.begin(), all_boundStrategy.end(), boundStrategy) == all_boundStrategy.end()) {
+        std::cerr << "Bound stategy '" + boundStrategy + "' is not recognized. 'Reflect-random' will be used.\n";
+        boundStrategy = "reflect-random";
+    }
+
+    const double dimension = double(bounds.size());
+    populationSize = options.get<int>("population_size", 0);
+    if (populationSize == 0) {
+        populationSize = std::max(size_t(4), size_t(25.0 * log(dimension) * sqrt(dimension)));
+    }
+
+    mutation_strategy = "current_to_pbest_AW_1bin";
+    memorySize = options.get<int>("memory_size", 5);
+    archive_size_ratio = options.get<double>("archive_size_ratio", 1.0);
+    if (archive_size_ratio < 0.0) {
+        archive_size_ratio = 1.0;
+    }
+
+    M_CR = std::vector<double>(memorySize, 0.8);
+    M_F = std::vector<double>(memorySize, 0.3);
+
+    minPopSize = options.get<int>("minimum_population_size", 4);
+    support_tol = false;
+    reltol = 1e-8;
+    restartRelTol = 1e-4;
+    refineRelTol = 1e-6;
+    decrease = 1.0;
+    maxRestart = 1.0;
+    hasInitialized = true;
+}
+
+void jSO_1::initialize() {
+    initializeCommon();
+}
+
+void jSO_2::initialize() {
+    initializeCommon();
+}
+
+void jSO_1::adaptParameters() {
+    adaptWithPopulationSchedule(false);
+}
+
+void jSO_2::adaptParameters() {
+    adaptWithPopulationSchedule(true);
+}
+
+void jSOAblationBase::adaptWithPopulationSchedule(bool nonlinearReduction) {
+    if (nonlinearReduction) {
+        adjustPopulationSizeNonlinear();
+    } else {
+        adjustPopulationSizeLinear();
+    }
+
+    adjustArchiveSize();
+    spread = calcStdDev(fitness) / std::fabs(calcMean(fitness));
+    if (spread <= reltol || do_refine) {
+        processRestartCycle();
+    }
+    updateParameterMemory();
+    resampleControlParameters();
+}
+
+void jSOAblationBase::adjustPopulationSizeLinear() {
+    size_t targetSize = size_t((double(double(minPopSize) - double(populationSize)) * (Nevals / double(maxevals)) + populationSize));
+    if (targetSize < static_cast<size_t>(minPopSize)) {
+        targetSize = static_cast<size_t>(minPopSize);
+    }
+
+    newPopulationSize = targetSize;
+    if (population.size() > newPopulationSize) {
+        const std::vector<size_t> sorted_index = argsort(fitness, true);
+        std::vector<std::vector<double>> new_population_subset(newPopulationSize);
+        std::vector<double> new_fitness_subset(newPopulationSize);
+        for (size_t i = 0; i < newPopulationSize; ++i) {
+            new_population_subset[i] = population[sorted_index[i]];
[…]

[…]
