+++ paper/paper.bib
@@ -476,4 +476,14 @@ location = {Montreal, Qu\'{e}bec, Canada},
 series = {GECCO '09}
 }
 
+@misc{muzakka2026rcmaes,
+      title={RCMAES: A Robust CMA-ES Variant for CEC2026 Competition}, 
+      author={Khoirul Faiq Muzakka and Sören Möller and Martin Finsterbusch},
+      year={2026},
+      eprint={2604.27138},
+      archivePrefix={arXiv},
+      primaryClass={cs.NE},
+      url={https://arxiv.org/abs/2604.27138}, 
+}
+
 
