### 2026-05-02 a69c707
Implemented nlshade_lbc
