### 2026-05-30 #10145 @anugrahjo
I have completed my review and checked off all checklist items except for `Collaborative effort`. I could not find clear evidence for this criterion from the repository or paper, such as contributions from multiple developers, issue/PR discussions involving external users, or community-driven changes. I am not fully sure how to evaluate this criterion in this case, so I defer this decision to the editor's judgment @gkthiruvathukal . Overall, I found the software to be of generally high quality, with clear potential value for the target community.

I have a few very minor comments before acceptance @khoirulmuzakka :
1. L188: The link to the documentation website is incorrect.
2. Please address the comment in issue [4](https://github.com/khoirulmuzakka/Minion/issues/4).
3. In the documentation for [Development Setup](https://minion-py.readthedocs.io/en/latest/contribute.html#development-setup), installing from PyPI for normal Python users is mentioned. This appears to be unintended documentation, possibly from AI-generated docs. I would strongly recommend that the author review the documentation carefully and ensure that every instruction is accurate and intended.
4. In my earlier comment 12, I only meant that it would help to mention some real-world problems in the **Summary** section to better motivate the software.
5. In the **Statement of need** section, the first lines in bullet points 1 and 2 are not limitations but rather capabilities. To be consistent, please rephrase them as limitations.
6. L58 says users can compare different algorithms and CEC benchmarks. I assume the inclusion of CEC benchmarks was a typo, as users are interested in comparing algorithms using CEC benchmarks, but they are generally not interested in comparing different benchmark problems.
7. As per JOSS's AI disclosure policy, the AI tools/models used and their versions need to be disclosed.

@gkthiruvathukal With these final minor comments addressed, I'd strongly recommend publishing this submission, subject to the editor’s judgment on the `Collaborative effort` checklist item.

@khoirulmuzakka Great job with the software!
