+++ minion/cec/cec2011.cpp
@@ -77,6 +77,99 @@ std::array<double, N> rk4Integrate(std::array<double, N> state, double t0, doubl
     return state;
 }
 
+template <size_t N, typename Deriv>
+std::vector<std::array<double, N>> rk45Integrate(std::array<double, N> state, double t0,
+                                                 double t1, Deriv &&deriv,
+                                                 double relTol = 1e-1,
+                                                 double absTol = 1e-6) {
+    std::vector<std::array<double, N>> trajectory;
+    trajectory.push_back(state);
+    double t = t0;
+    double h = std::max(1e-6, (t1 - t0) / 40.0);
+    const double maxStep = std::max(1e-6, (t1 - t0) / 9.0);
+    constexpr int refine = 4;
+
+    while (t < t1) {
+        h = std::min(h, maxStep);
+        if (t + h > t1) {
+            h = t1 - t;
+        }
+
+        const auto startState = state;
+        std::array<double, N> k1{}, k2{}, k3{}, k4{}, k5{}, k6{}, k7{}, temp{};
+        deriv(t, state, k1);
+
+        for (size_t i = 0; i < N; ++i) {
+            temp[i] = state[i] + h * (1.0 / 5.0) * k1[i];
+        }
+        deriv(t + h * (1.0 / 5.0), temp, k2);
+
+        for (size_t i = 0; i < N; ++i) {
+            temp[i] = state[i] + h * ((3.0 / 40.0) * k1[i] + (9.0 / 40.0) * k2[i]);
+        }
+        deriv(t + h * (3.0 / 10.0), temp, k3);
+
+        for (size_t i = 0; i < N; ++i) {
+            temp[i] = state[i] + h * ((44.0 / 45.0) * k1[i] + (-56.0 / 15.0) * k2[i] + (32.0 / 9.0) * k3[i]);
+        }
+        deriv(t + h * (4.0 / 5.0), temp, k4);
+
+        for (size_t i = 0; i < N; ++i) {
+            temp[i] = state[i] +
+                      h * ((19372.0 / 6561.0) * k1[i] + (-25360.0 / 2187.0) * k2[i] + (64448.0 / 6561.0) * k3[i] +
+                           (-212.0 / 729.0) * k4[i]);
+        }
+        deriv(t + h * (8.0 / 9.0), temp, k5);
+
+        for (size_t i = 0; i < N; ++i) {
+            temp[i] = state[i] +
+                      h * ((9017.0 / 3168.0) * k1[i] + (-355.0 / 33.0) * k2[i] + (46732.0 / 5247.0) * k3[i] +
+                           (49.0 / 176.0) * k4[i] + (-5103.0 / 18656.0) * k5[i]);
+        }
+        deriv(t + h, temp, k6);
+
+        std::array<double, N> y5{};
+        for (size_t i = 0; i < N; ++i) {
+            y5[i] = state[i] +
+                    h * ((35.0 / 384.0) * k1[i] + (500.0 / 1113.0) * k3[i] + (125.0 / 192.0) * k4[i] +
+                         (-2187.0 / 6784.0) * k5[i] + (11.0 / 84.0) * k6[i]);
+        }
+        deriv(t + h, y5, k7);
+
+        std::array<double, N> y4{};
+        for (size_t i = 0; i < N; ++i) {
+            y4[i] = state[i] +
+                    h * ((5179.0 / 57600.0) * k1[i] + (7571.0 / 16695.0) * k3[i] + (393.0 / 640.0) * k4[i] +
+                         (-92097.0 / 339200.0) * k5[i] + (187.0 / 2100.0) * k6[i] + (1.0 / 40.0) * k7[i]);
+        }
+
+        double errorNorm = 0.0;
+        for (size_t i = 0; i < N; ++i) {
[…]
