+++ minionpy/cec.py
@@ -154,7 +154,7 @@ CEC2011_METADATA.update(
                 "-{pi} -{pi} -{pi} -{pi} -{pi}".format(pi=np.pi)
             ),
             _np_from_str(
-                "2300 4.05 1 1 500 500 500 500 500 600 0.99 0.99 0.99 0.99 0.99 0.99 6 6 6 6 6 "
+                "2200 5 1 1 500 500 500 500 500 600 0.99 0.99 0.99 0.99 0.99 0.99 6 6 6 6 6 "
                 "{pi} {pi} {pi} {pi} {pi}".format(pi=np.pi)
             ),
         ),
