+++ comparison_scripts/verify_cec2011_octave.py
@@ -1,0 +1,364 @@
+#!/usr/bin/env python3
+from __future__ import annotations
+
+import argparse
+import math
+import os
+import random
+import subprocess
+import sys
+import tempfile
+from typing import Iterable
+
+
+REPO_ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
+MINIONPY_LIB = os.path.join(REPO_ROOT, "minionpy", "lib")
+
+if REPO_ROOT not in sys.path:
+    sys.path.insert(0, REPO_ROOT)
+if MINIONPY_LIB not in sys.path:
+    sys.path.insert(0, MINIONPY_LIB)
+
+
+def _cec2011_problem02_bounds(ndim: int):
+    lb = [0.0] * ndim
+    for idx in range(3, ndim):
+        lb[idx] = -4.0 - 0.25 * int((idx - 4) / 3)
+
+    ub = [0.0] * ndim
+    ub[0] = ub[1] = 4.0
+    ub[2] = math.pi
+    for idx in range(3, ndim):
+        ub[idx] = 4.0 + 0.25 * int((idx - 4) / 3)
+    return lb, ub
+
+
+def _cec2011_problem05_bounds(ndim: int):
+    lb = [-1.0] * ndim
+    lb[:3] = [0.0, 0.0, 0.0]
+    ub = [0.0] * ndim
+    ub[0] = ub[1] = 4.0
+    ub[2] = math.pi
+    for idx in range(3, ndim):
+        ub[idx] = 4.0 + 0.25 * int((idx - 4) / 3)
+    return lb, ub
+
+
+def _cec2011_problem06_bounds(ndim: int):
+    lb = [-1.0] * ndim
+    lb[:3] = [0.0, 0.0, 0.0]
+    ub = [0.0] * ndim
+    ub[0] = ub[1] = 4.0
+    ub[2] = math.pi
+    for idx in range(3, ndim, 3):
+        ub[idx] = 4.0 + 0.25 * int((1 - 4) / 3)
+        if idx + 1 < ndim:
+            ub[idx + 1] = 4.0 + 0.25 * int((2 - 4) / 3)
+        if idx + 2 < ndim:
+            ub[idx + 2] = 4.0 + 0.25 * int((3 - 4) / 3)
+    return lb, ub
+
+
+def _from_str(data: str):
+    return [float(token) for token in data.split()]
+
+
+CEC2011_METADATA = {
+    1: (6, [-6.4] * 6, [6.35] * 6),
+    2: (30, *_cec2011_problem02_bounds(30)),
+    3: (1, [-0.6], [0.9]),
+    4: (1, [0.0], [5.0]),
+    5: (30, *_cec2011_problem05_bounds(30)),
+    6: (30, *_cec2011_problem06_bounds(30)),
+    7: (20, [0.0] * 20, [2.0 * math.pi] * 20),
+    8: (7, [0.0] * 7, [15.0] * 7),
+    9: (
+        126,
+        [0.0] * 126,
+        _from_str(
+            """
+            0.217 0.024 0.076 0.892 0.128 0.25 0.058 0.112 0.062 0.082 0.035 0.09 0.032 0.095 0.022 0.175 0.032
+            0.087 0.035 0.024 0.106 0.217 0.024 0.026 0.491 0.228 0.3 0.058 0.112 0.062 0.082 0.035 0.09 0.032
+            0.095 0.022 0.175 0.032 0.087 0.035 0.024 0.106 0.216 0.024 0.076 0.216 0.216 0.216 0.058 0.112
[…]
