### 2026-06-01 4f2cd09
F4 now use interpolating table
