+++ comparison_scripts/verify_cec2011_octave.py
@@ -2,9 +2,12 @@
 from __future__ import annotations
 
 import argparse
+import concurrent.futures
+import glob
 import math
 import os
 import random
+import shutil
 import subprocess
 import sys
 import tempfile
@@ -116,16 +119,8 @@ CEC2011_METADATA = {
 }
 
 
-DEFAULT_TOLERANCES = {
-    3: (1e-2, 1e-2),
-    4: (1e-2, 1e-2),
-    21: (1e-4, 1e-4),
-    22: (1e-4, 1e-4),
-}
-
-
 def get_tolerance(problem: int):
-    return DEFAULT_TOLERANCES.get(problem, (1e-8, 1e-7))
+    return (0.0, 1e-9)
 
 
 def _build_cpp_evaluator(problem: int):
@@ -146,6 +141,65 @@ def _octave_env():
         if key.startswith("SNAP"):
             env.pop(key, None)
     return env
+
+
+def _candidate_octave_paths():
+    candidates = []
+
+    for env_key in ("OCTAVE_CLI", "OCTAVE_BIN", "OCTAVE_EXECUTABLE"):
+        value = os.environ.get(env_key)
+        if value:
+            candidates.append(value)
+
+    for name in ("octave-cli", "octave", "octave-cli.exe", "octave.exe"):
+        resolved = shutil.which(name)
+        if resolved:
+            candidates.append(resolved)
+
+    if os.name == "nt":
+        roots = []
+        for env_key in ("ProgramFiles", "ProgramFiles(x86)", "LocalAppData"):
+            root = os.environ.get(env_key)
+            if root:
+                roots.append(root)
+
+        patterns = [
+            os.path.join("GNU Octave", "Octave-*", "mingw64", "bin", "octave-cli.exe"),
+            os.path.join("GNU Octave", "Octave-*", "mingw64", "bin", "octave.exe"),
+            os.path.join("Octave", "Octave-*", "mingw64", "bin", "octave-cli.exe"),
+            os.path.join("Octave", "Octave-*", "mingw64", "bin", "octave.exe"),
+        ]
+        for root in roots:
+            for pattern in patterns:
+                candidates.extend(glob.glob(os.path.join(root, pattern)))
+
+    deduped = []
+    seen = set()
+    for candidate in candidates:
+        normalized = os.path.normcase(os.path.abspath(candidate))
+        if normalized not in seen and os.path.exists(candidate):
+            deduped.append(candidate)
+            seen.add(normalized)
+    return deduped
+
+
+def resolve_octave_binary(explicit_path: str | None = None):
+    candidates = []
+    if explicit_path:
+        candidates.append(explicit_path)
+    candidates.extend(_candidate_octave_paths())
+
+    for candidate in candidates:
+        resolved = shutil.which(candidate) if not os.path.isabs(candidate) else candidate
+        if resolved and os.path.exists(resolved):
+            return resolved
+
+    searched = ", ".join(candidates) if candidates else "PATH and standard install locations"
+    raise RuntimeError(
+        "Unable to find an Octave executable. "
+        "Install GNU Octave, add it to PATH, or pass --octave-bin / set OCTAVE_CLI. "
+        f"Searched: {searched}"
+    )
 
 
 def _octave_expression(problem: int, x_expr: str):
@@ -209,10 +263,11 @@ def _format_matrix_for_octave(matrix: list[list[float]]):
     return "[ " + " ; ".join(rows) + " ]"
 
 
-def evaluate_with_octave(problem: int, samples: list[list[float]]):
+def evaluate_with_octave(problem: int, samples: list[list[float]], octave_bin: str | None = None):
     matrix_literal = _format_matrix_for_octave(samples)
[…]
