### 2026-06-05 605da41
API changes. Now tol is part of options, not as a constructor argument of the Minimizer
