+++ minion/include/arrde.h
@@ -126,7 +126,6 @@ class ARRDE : public Differential_Evolution {
          * @param x0 The initial guesses for the solution. Note that Minion assumes multiple initial guesses, thus, x0 is an std::vector<std::vector<double>> object. These guesses will be used for population initialization.
          * @param data Additional data for the objective function.
          * @param callback Callback function for intermediate results.
-         * @param tol The tolerance for stopping criteria.
          * @param maxevals The maximum number of evaluations.
          * @param seed The seed for random number generation.
          * @param options Option map that specifies further configurational settings for the algorithm.
@@ -137,12 +136,11 @@ class ARRDE : public Differential_Evolution {
             const std::vector<std::vector<double>>& x0 = {},
             void* data = nullptr, 
             std::function<void(MinionResult*)> callback = nullptr,
-            double tol = 0.0001, 
             size_t maxevals = 100000, 
             int seed=-1, 
             std::map<std::string, ConfigValue> options = std::map<std::string, ConfigValue>()
         ) :  
-            Differential_Evolution(func, bounds, x0, data, callback, tol, maxevals, seed, options){support_tol=false;};
+            Differential_Evolution(func, bounds, x0, data, callback, maxevals, seed, options){support_tol=false;};
 
         /**
          * @brief Adapts parameters of the algorithm.
