+++ docs/source/Algorithms.rst
@@ -19,7 +19,7 @@ In C++, you can retrieve the default settings for an algorithm and adjust the pa
     options["population_size"] = 50; 
 
     // Initialize the minimizer with custom settings
-    auto min = minion::Minimizer(rosenbrock_vect, bounds, {}, nullptr, nullptr, "LSHADE", 0.0, max_evals, -1, options);
+    auto min = minion::Minimizer(rosenbrock_vect, bounds, {}, nullptr, nullptr, "LSHADE", max_evals, -1, options);
 
     // Perform optimization
     auto min_result = min.optimize();
@@ -40,14 +40,13 @@ In Python, you can pass the parameters via a dictionary when creating the `Minim
     }
 
     # Use ARRDE algorithm for optimization
-    min = mpy.Minimizer(func=objective_function, 
-                        x0=x0, 
-                        bounds=[(-10, 10)] * dimension, 
-                        algo="ARRDE", 
-                        relTol=0.0, 
-                        maxevals=10000, 
-                        callback=None, 
-                        seed=None, 
+    min = mpy.Minimizer(func=objective_function,
+                        x0=x0,
+                        bounds=[(-10, 10)] * dimension,
+                        algo="ARRDE",
+                        maxevals=10000,
+                        callback=None,
+                        seed=None,
                         options=options)
 
     # Run optimization
@@ -75,6 +74,32 @@ Differential Evolution (DE) is a population-based stochastic optimization algori
 *Reference : Storn, R and Price, K, Differential Evolution - a Simple and Efficient Heuristic for Global Optimization over Continuous Spaces, Journal of Global Optimization, 1997, 11, 341 - 359.*
 
 Algorithm name : ``"DE"``
+
+Parameters :
+
+- ``population_size``: 0
+
+  .. note:: Population size. ``0`` lets the solver choose its heuristic default from problem dimension and budget.
+
+- ``mutation_rate``: 0.5
+
+  .. note:: Differential weight :math:`F` used to scale mutation steps.
+
+- ``crossover_rate``: 0.8
+
+  .. note:: Binomial or exponential crossover probability :math:`CR`.
+
+- ``mutation_strategy``: ``best1bin``
+
+  .. note:: Mutation / crossover strategy. Available strategies include ``"best1bin"``, ``"best1exp"``, ``"rand1bin"``, ``"rand1exp"``, ``"current_to_best1bin"``, ``"current_to_best1exp"``, ``"current_to_pbest1bin"``, ``"current_to_pbest1exp"``, ``"current_to_pbest_A_1bin"``, ``"current_to_pbest_A_1exp"``, ``"current_to_pbest_AW_1bin"``, and ``"current_to_pbest_AW_1exp"``.
+
+- ``convergence_tol``: 1e-4
+
+  .. note:: Diversity-based convergence tolerance. Optimization stops when the population fitness spread falls below this threshold.
+
+- ``bound_strategy``: ``reflect-random``
+
+  .. note:: Boundary handling policy. Available strategies: ``"random"``, ``"reflect-random"``, ``"clip"``.
 
 
 Particle Swarm Optimization (PSO)
@@ -106,6 +131,10 @@ Parameters :
 - ``velocity_clamp``: 0.2  
 
   .. note:: Fraction of the search range used as a velocity limit. ``0`` disables clamping.
+
+- ``convergence_tol``: 1e-4
+
+  .. note:: Diversity-based convergence tolerance on swarm fitness values. Smaller values require the swarm to contract more before stopping.
 
 - ``bound_strategy``: ``reflect-random``  
 
[…]
