+++ docs/source/Convergence.rst
@@ -6,22 +6,13 @@ As a result, the computational budget is primarily limited by the maximum number
 This differs from other optimization libraries, where an algorithm stops either when the function no longer improves 
 or when a predefined maximum number of iterations is reached.
 
-In Minion, we do use tolerance-based convergence criteria, but some algorithms, especially population-based ones, 
-do not support this. The primary reason for this is that the population is not designed to converge completely; 
-instead, it is maintained to ensure continued exploration of the search space. 
-Additionally, we do not use iteration-based stopping criteria, as the number of function calls per iteration can vary, 
-making it less intuitive to map to maxevals.
+In Minion, tolerance-based convergence is configured through the algorithm options map or dictionary, using the
+``"convergence_tol"`` key for algorithms that support it. We do not use a global constructor-level tolerance anymore.
+Additionally, we do not use iteration-based stopping criteria, as the number of function calls per iteration can vary,
+making it less intuitive to map to ``maxevals``.
 
-Below is a list of algorithms that **ignore** the ``relTol`` parameter. Even if it is set, it will have no effect:
-
-- ARRDE  
-- AGSK  
-- j2020  
-- Dual Annealing  
-- RCMAES
-- BIPOP-aCMAES
-
-For these, the ``relTol`` (in Python) or ``tol`` (in C++) parameters specify the maximum allowed value for the standard deviation of the 
-function values divided by the average of the function values before the algorithm stops.
+For supported population-based algorithms, ``"convergence_tol"`` typically controls a diversity-based stopping rule
+derived from the spread of population fitness values. For ``NelderMead`` and ``DA``, the stopping rule is algorithm-specific
+rather than population-diversity-based.
 
 Note that L-BFGS-B has its own stopping criteria, which is specified in the algorithm options (``g_epsilon``, ``g_epsilon_rel``, ``f_reltol``).
