+++ README.md
@@ -67,6 +67,184 @@ Minion also serves as a **research platform** for developing and testing new opt
 
 Detailed installation, native package, C++ build, and usage instructions are available in the official documentation.
 
+## Algorithm Usage
+
+Across both C++ and Python, Minion uses the same core model:
+- The objective must be **vectorized**: it receives a batch of candidate points and returns one objective value per point.
+- `x0` contains **multiple initial guesses**, where each inner vector/list is one candidate start.
+- `x0` is not a full population specification. In population-based algorithms, Minion first initializes the population using the algorithm's normal rule, then replaces some initialized individuals with the supplied guesses. In non-population-based algorithms, Minion evaluates the guesses and starts from the best one.
+- After constructing the `Minimizer`, call `optimize()` to run the algorithm and get a result object.
+
+### C++ Code
+
+Minion expects a **vectorized** objective function:
+
+```cpp
+#include <minion.h>
+#include <iostream>
+#include <vector>
+
+double rosenbrock_scalar(const std::vector<double>& x) {
+    double sum = 0.0;
+    for (size_t i = 0; i + 1 < x.size(); ++i) {
+        const double a = x[i + 1] - x[i] * x[i];
+        const double b = 1.0 - x[i];
+        sum += 100.0 * a * a + b * b;
+    }
+    return sum;
+}
+
+std::vector<double> rosenbrock_batch(const std::vector<std::vector<double>>& X, void*) {
+    std::vector<double> out;
+    out.reserve(X.size());
+    for (const auto& x : X) {
+        out.push_back(rosenbrock_scalar(x));
+    }
+    return out;
+}
+
+int main() {
+    const size_t dim = 10;
+    std::vector<std::pair<double, double>> bounds(dim, {-5.0, 5.0});
+
+    // x0 is a vector of guesses: each inner vector is one candidate start.
+    std::vector<std::vector<double>> x0 = {
+        std::vector<double>(dim, 0.0),
+        std::vector<double>(dim, 1.0)
+    };
+
+    auto options = minion::DefaultSettings().getDefaultSettings("DE");
+    options["population_size"] = 50;
+    options["convergence_tol"] = 1e-6;
+
+    minion::Minimizer optimizer(
+        rosenbrock_batch, bounds, x0, nullptr, nullptr, "DE", 20000, 42, options
+    );
+
+    minion::MinionResult result = optimizer.optimize();
+    std::cout << "best f = " << result.fun << "\n";
+}
+```
+
+In C++, `MinionFunction` is the vectorized entry point: it takes `std::vector<std::vector<double>>` and returns one value per candidate point.
+
+### Python Code
+
+MinionPy uses the same idea: the objective function should be **vectorized** and accept a batch of points.
+
+```python
+import minionpy as mpy
+
+def rosenbrock_batch(X):
+    values = []
+    for x in X:
+        total = 0.0
+        for i in range(len(x) - 1):
+            a = x[i + 1] - x[i] * x[i]
+            b = 1.0 - x[i]
+            total += 100.0 * a * a + b * b
+        values.append(total)
+    return values
+
+dimension = 10
+bounds = [(-5.0, 5.0)] * dimension
+
+# x0 is a list of guesses: each inner list is one candidate start.
+x0 = [
+    [0.0] * dimension,
+    [1.0] * dimension,
+]
+
+options = {
+    "population_size": 50,
+    "convergence_tol": 1e-6,
+}
+
+optimizer = mpy.Minimizer(
+    func=rosenbrock_batch,
+    x0=x0,
+    bounds=bounds,
+    algo="DE",
+    maxevals=20000,
[…]
