### 2026-06-06 40ab0f8
Streamlined python cec api, update docs
