+++ README.md
@@ -43,10 +43,10 @@ Minion also serves as a **research platform** for developing and testing new opt
     - Generalized Simulated Annealing (Dual Annealing) 
     - L-BFGS-B (vectorized & noise-robust) 
     - L-BFGS (vectorized & noise-robust) 
-- **Highly parallelized**
-  - Designed for **vectorized function evaluations**, supporting **multithreading and multiprocessing** to speed up optimization.
-- **Optimized C++ backend with Python API**
-  - Enjoy the performance of C++ with the simplicity of Python.
+- **Native Batch Evaluation**
+  - Algorithms work with **vectorized / batched objective evaluations** by default, making it straightforward to use multithreading or multiprocessing inside the objective when needed.
+- **C++ Library with Python Bindings**
+  - The core implementation is in C++, with matching Python access through `minionpy`.
 - **CEC Benchmark Suite**  
   - Includes `CEC2011`, `CEC2014`, `CEC2017`, `CEC2019`, `CEC2020`, and `CEC2022` benchmark problems for rigorous algorithm testing.  
   - The benchmark problems are directly adapted from the original C++ implementations.  
@@ -179,17 +179,19 @@ In Python, `func(X)` receives `X` as `list[list[float]]` and returns one value p
 
 ## CEC Benchmark Usage
 
-The CEC wrappers follow the same optimizer workflow as ordinary objectives. The main difference is that the benchmark evaluators already implement batch evaluation logic.
+Minion exposes the CEC suites as callable benchmark objects. The usage pattern is the same in C++ and Python:
+- Construct a benchmark evaluator such as `CEC2017Functions(function_number, dimension)`.
+- Pass that evaluator to `Minimizer` as the objective.
+- Provide `bounds` separately to `Minimizer`.
+- The benchmark evaluators are already vectorized, so in Python no wrapper is needed. In C++, a thin adapter is usually used to forward the batch to the benchmark object.
 
-### C++ Code
+Most suites use the constructor form `(function_number, dimension)`, for example `CEC2014`, `CEC2017`, `CEC2020`, and `CEC2022`. `CEC2019` and `CEC2011` also accept that form for API consistency, although their effective dimensions are suite-defined. For `CEC2011`, MinionPy additionally exposes `get_bounds()`.
 
-Minion provides CEC benchmark wrappers in C++. These already match Minion's vectorized interface, so you can pass them directly through a thin adapter:
+Example using `CEC2017` in C++:
 
 ```cpp
 #include <minion.h>
 #include <minion_cec.h>
-#include <iostream>
-#include <vector>
 
 std::vector<double> cec2017_batch(const std::vector<std::vector<double>>& X, void* data) {
     auto* cec = static_cast<minion::CECBase*>(data);
@@ -197,34 +199,34 @@ std::vector<double> cec2017_batch(const std::vector<std::vector<double>>& X, voi
 }
 
 int main() {
-    const int dim = 30;
-    minion::CEC2017Functions cec_f1(1, dim);
-    std::vector<std::pair<double, double>> bounds(dim, {-100.0, 100.0});
-    std::vector<std::vector<double>> x0 = {
-        std::vector<double>(dim, 0.0)
-    };
+    const int function_number = 1;
+    const int dimension = 30;
+    const size_t maxevals = 30000;
+    const int seed = 20250306;
 
-    auto options = minion::DefaultSettings().getDefaultSettings("ARRDE");
+    minion::CEC2017Functions cec_f1(function_number, dimension);
[…]
