+++ docs/source/minion/index.rst
@@ -7,5 +7,4 @@ Minion (C++)
    :caption: Contents:
 
    examples
-   cec
    cppapi

[…]
