+++ docs/source/algolist.rst
@@ -1,0 +1,91 @@
+Implemented Algorithms
+======================
+
+This page provides a compact list of the optimization algorithms currently implemented in Minion and MinionPy, together with their canonical literature references. For implementation-specific hybrids, the cited paper is the closest parent or representative method.
+
+For detailed descriptions, parameter documentation, and default options, see :doc:`Algorithms`.
+
+
+Differential Evolution Family
+=============================
+
+``DE``
+    Storn, R and Price, K, *Differential Evolution - a Simple and Efficient Heuristic for Global Optimization over Continuous Spaces*, Journal of Global Optimization, 1997, 11, 341-359.
+
+``JADE``
+    J. Zhang and A. C. Sanderson, *JADE: Adaptive Differential Evolution With Optional External Archive*, IEEE Transactions on Evolutionary Computation, 13(5), 945-958, 2009.
+
+``LSHADE``
+    R. Tanabe and A. S. Fukunaga, *Improving the search performance of SHADE using linear population size reduction*, IEEE CEC, 2014.
+
+``LSHADE_cnEpSin``
+    N. H. Awad, M. Z. Ali and P. N. Suganthan, *Ensemble sinusoidal differential covariance matrix adaptation with Euclidean neighborhood for solving CEC2017 benchmark problems*, IEEE CEC, 2017.
+
+``jSO``
+    J. Brest, M. S. Maučec and B. Bošković, *Single objective real-parameter optimization: Algorithm jSO*, IEEE CEC, 2017.
+
+``j2020``
+    J. Brest, M. S. Maučec and B. Bošković, *Differential Evolution Algorithm for Single Objective Bound-Constrained Optimization: Algorithm j2020*, IEEE CEC, 2020.
+
+``NLSHADE_RSP``
+    V. Stanovov, S. Akhmedova and E. Semenkin, *NL-SHADE-RSP Algorithm with Adaptive Archive and Selective Pressure for CEC 2021 Numerical Optimization*, IEEE CEC, 2021.
+
+``LSRTDE``
+    V. Stanovov and E. Semenkin, *Success Rate-based Adaptive Differential Evolution L-SRTDE for CEC 2024 Competition*, IEEE CEC, 2024.
+
+``ARRDE``
+    Khoirul Faiq Muzakka, Ahsani Hafizhu Shali, Haris Suhendar, Sören Möller, Martin Finsterbusch, *Robust Differential Evolution via Nonlinear Population Size Reduction and Adaptive Restart: The ARRDE Algorithm*, arXiv, 2025. https://arxiv.org/abs/2511.18429
+
+``IMODE``
+    Karam M. Sallam, Saber M. Elsayed, Ripon K. Chakrabortty, and Michael J. Ryan, *Improved Multi-operator Differential Evolution Algorithm for Solving Unconstrained Problems*, IEEE CEC, 2020.
+
+``AGSK``
+    A. W. Mohamed, A. A. Hadi, A. K. Mohamed and N. H. Awad, *Evaluating the Performance of Adaptive Gaining-Sharing Knowledge Based Algorithm on CEC 2020 Benchmark Problems*, IEEE CEC, 2020.
+
+``GWO_DE``
+    S. Mirjalili, S. M. Mirjalili, and A. Lewis, *Grey Wolf Optimizer*, Advances in Engineering Software, 69, 46-61, 2014. Minion implements a Grey Wolf / Differential Evolution hybrid based on this family of methods.
+
+
+Particle Swarm and Swarm-Based Methods
+======================================
+
+``PSO``
[…]
