+++ docs/source/minionpy/index.rst
@@ -1,6 +1,6 @@
 MinionPy (python)
 =================
-Welcome to the MinionPy documentation! MinionPy is a Python wrapper for the Minion C++ library, providing an easy-to-use interface to state-of-the-art optimization algorithms.
+MinionPy is a Python wrapper for the Minion C++ library, providing an easy-to-use interface to state-of-the-art optimization algorithms.
 
 .. toctree::
    :maxdepth: 2
