### 2026-06-06 c769755
Made more robust algo names
