### 2026-06-06 949c312
Version bump (1.6.0)
