### 2026-06-06 b3512a0
Fixed wheel generation on intel silicon mac
