+++ minion/cec/cec.cpp
@@ -1,5 +1,4 @@
 #include "cec.h" 
-#include <filesystem>
 
 #if defined(_MSC_VER) // Check if compiling with MSVC
 #pragma warning(push)
@@ -16,6 +15,7 @@
 
 
 #include <string>
+#include <vector>
 #if defined(_WIN32)
     #ifndef NOMINMAX
         #define NOMINMAX
@@ -23,9 +23,13 @@
     #ifndef WIN32_LEAN_AND_MEAN
         #define WIN32_LEAN_AND_MEAN
     #endif
+    #include <direct.h>
     #include <windows.h>
 #else
     #include <dlfcn.h>
+    #include <limits.h>
+    #include <sys/stat.h>
+    #include <unistd.h>
 #endif
 
 
@@ -62,30 +66,158 @@ std::string getLibraryDirectory() {
 
 namespace {
 
-std::string normalize_path(const std::filesystem::path& path) {
-    std::error_code ec;
-    const std::filesystem::path normalized = std::filesystem::weakly_canonical(path, ec);
-    if (!ec) {
-        return normalized.string();
+char path_separator() {
+#if defined(_WIN32)
+    return '\\';
+#else
+    return '/';
+#endif
+}
+
+bool is_path_separator(char c) {
+    return c == '/' || c == '\\';
+}
+
+bool has_drive_prefix(const std::string& path) {
+    return path.size() > 1 &&
+           ((path[0] >= 'A' && path[0] <= 'Z') || (path[0] >= 'a' && path[0] <= 'z')) &&
+           path[1] == ':';
+}
+
+bool is_absolute_path(const std::string& path) {
+    return !path.empty() && (is_path_separator(path.front()) || has_drive_prefix(path));
+}
+
+std::string join_path(const std::string& base, const std::string& child) {
+    if (base.empty()) {
+        return child;
     }
-    return path.lexically_normal().string();
+    if (child.empty()) {
+        return base;
+    }
+    if (is_absolute_path(child)) {
+        return child;
+    }
+    if (is_path_separator(base.back())) {
+        return base + child;
+    }
+    return base + path_separator() + child;
+}
+
+std::vector<std::string> split_path(const std::string& path) {
+    std::vector<std::string> parts;
+    std::string current;
+    for (char c : path) {
+        if (is_path_separator(c)) {
+            if (!current.empty()) {
+                parts.push_back(current);
+                current.clear();
+            }
+        } else {
+            current.push_back(c);
+        }
+    }
+    if (!current.empty()) {
+        parts.push_back(current);
+    }
+    return parts;
+}
+
+std::string normalize_lexical_path(const std::string& path) {
+    if (path.empty()) {
+        return ".";
+    }
+
+    const bool absolute = is_absolute_path(path);
+    const std::string root = has_drive_prefix(path)
+        ? path.substr(0, 2) + path_separator()
+        : (absolute ? std::string(1, path_separator()) : "");
+    std::vector<std::string> normalized_parts;
+    const std::string path_without_root = has_drive_prefix(path) ? path.substr(2) : path;
+    for (const std::string& part : split_path(path_without_root)) {
+        if (part == ".") {
+            continue;
+        }
+        if (part == "..") {
+            if (!normalized_parts.empty() && normalized_parts.back() != "..") {
+                normalized_parts.pop_back();
+            } else if (!absolute) {
+                normalized_parts.push_back(part);
+            }
+            continue;
+        }
+        normalized_parts.push_back(part);
+    }
+
+    std::string normalized = root;
+    for (size_t i = 0; i < normalized_parts.size(); ++i) {
[…]
