### 2026-06-10 5a0b1a7
Clarify constructor form usage in README
