### 2026-06-10 c5c1b89
Added manual trigger for github workflows
