+++ .github/workflows/wheel.yaml
@@ -1,6 +1,7 @@
 name: Wheel Build
 
 on:
+  workflow_dispatch:
   push:
     tags:
       - "v*"
