+++ .github/workflows/ci.yml
@@ -28,9 +28,9 @@ jobs:
             python-version: "3.9"
           - os: macos-15-intel
             python-version: "3.14"
-          - os: windows-latest
+          - os: windows-2022
             python-version: "3.9"
-          - os: windows-latest
+          - os: windows-2022
             python-version: "3.14"
 
     steps:
