+++ paper/paper.md
@@ -28,7 +28,7 @@ bibliography: paper.bib
 
 # Summary
 
-Minion is a derivative-free optimization library for unconstrained and bound-constrained single-objective blackbox problems, where users provide objective-function evaluations rather than analytic gradients. It provides a C++ backend with a Python interface (MinionPy), supporting applications in engineering, machine learning, and scientific computing.
+Optimization problems arise throughout science, engineering, and machine learning, including tasks such as simulation calibration, experimental data fitting, hyperparameter tuning, and design optimization. A common setting is single-objective optimization, where the goal is to minimize or maximize one performance measure. In many practical cases, the objective function is a black box: it can be evaluated at chosen points, but no closed-form expression, gradient, or other derivative information is available. To support optimization problems of this kind, we introduce Minion, a derivative-free optimization library for unconstrained and bound-constrained single-objective blackbox problems. It is implemented in C++, with a Python interface available through MinionPy.
 
 The library offers a centralized implementation of advanced derivative-free optimization algorithms, including state-of-the-art Differential Evolution (DE) methods that have performed strongly in IEEE Congress on Evolutionary Computation (CEC) competitions, Covariance Matrix Adaptation Evolution Strategy (CMA-ES) variants, and several Particle Swarm Optimization (PSO) variants. Alongside these research-grade solvers, Minion ships widely used optimizers such as Nelder–Mead, Dual Annealing (generalized simulated annealing), L-BFGS, and L-BFGS-B, allowing practitioners to combine established baselines with advanced heuristics within a single API. Many existing optimization libraries include only elementary variants of these methods and lack standardized benchmark problems. Minion addresses this by integrating multiple CEC benchmark suites (2011, 2014, 2017, 2019, 2020, and 2022) to facilitate algorithm evaluation and comparison.
 
@@ -46,15 +46,15 @@ DEAP [@DEAP_JMLR2012] provides a highly extensible Python framework for construc
 
 Minion aims to complement these ecosystems by concentrating on single-objective optimization with built-in support for batch evaluation, finite-difference quasi-Newton solvers, modern DE, CMA-ES, and PSO variants, and integrated CEC benchmark suites. The goal is not to replace these libraries, but to offer an option tailored to scenarios where such capabilities are central requirements.
 
-The decision to build Minion rather than contribute a single extension to an existing package follows from this scope. Adding one or two solvers to SciPy, NLopt, DEAP, or pagmo2 would not by itself provide a coherent benchmark platform, a shared batched-objective interface across both global and local solvers, and C++/Python access to a curated collection of recent CEC-oriented algorithms. Minion's scholarly contribution is therefore the integration of these components into a reproducible research tool: users can compare modern DE variants, CMA-ES variants, swarm methods, classical local search, and CEC benchmark suites without changing objective-function conventions or stitching together several libraries.
+The decision to build Minion rather than contribute a single extension to an existing package follows from this scope. Adding one or two solvers to SciPy, NLopt, DEAP, or pagmo2 would not by itself provide a coherent benchmark platform, a shared batched-objective interface across both global and local solvers, and C++/Python access to a curated collection of recent CEC-oriented algorithms. Minion's scholarly contribution is therefore the integration of these components into a reproducible research tool: users can compare modern DE variants, CMA-ES variants, swarm-based methods, and classical local search algorithms, and optionally evaluate them on standardized CEC benchmark suites without changing objective-function conventions or stitching together multiple libraries
 
 # Statement of need
 
 Minion was created to address several limitations in existing optimization libraries:
 
[…]
