+++ minionpy/__init__.py
@@ -5,7 +5,7 @@ from .parallel import *
 from .cec import CEC2011Functions as CEC2011
 
 # Define the package's version
-__version__ = "1.6.0"
+__version__ = "1.6.1"
 
 # Optionally, define the package name
 __name__ = "minionpy"
