### 2026-06-14 816c4c0
Added CI for pypi package
