+++ .github/workflows/pypi_test.yml
@@ -1,0 +1,68 @@
+name: PyPI Test
+
+on:
+  workflow_dispatch:
+
+jobs:
+  pypi-test:
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        include:
+          - os: ubuntu-latest
+            python-version: "3.9"
+          - os: ubuntu-latest
+            python-version: "3.14"
+          - os: macos-latest
+            python-version: "3.9"
+          - os: macos-latest
+            python-version: "3.14"
+          - os: macos-15-intel
+            python-version: "3.9"
+          - os: macos-15-intel
+            python-version: "3.14"
+          - os: windows-2022
+            python-version: "3.9"
+          - os: windows-2022
+            python-version: "3.14"
+
+    steps:
+      - name: Setup Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
+
+      - name: Install minionpy from PyPI
+        run: |
+          python -m pip install --upgrade pip
+          python -m pip install --upgrade minionpy
+
+      - name: Download test_minionpy.py from current commit
+        uses: actions/github-script@v7
+        with:
+          script: |
+            const fs = require("fs");
+            const path = require("path");
+            const { owner, repo } = context.repo;
+            const response = await github.rest.repos.getContent({
+              owner,
+              repo,
+              path: "tests/test_minionpy.py",
+              ref: context.sha,
+            });
+
+            if (!("content" in response.data)) {
+              core.setFailed("Unable to fetch tests/test_minionpy.py");
+              return;
+            }
+
+            const content = Buffer.from(response.data.content, "base64").toString("utf8");
+            const outputDir = path.join(process.env.RUNNER_TEMP, "pypi-test");
+            fs.mkdirSync(outputDir, { recursive: true });
+            fs.writeFileSync(path.join(outputDir, "test_minionpy.py"), content, "utf8");
+
+      - name: Run Python tests against installed package
+        run: python test_minionpy.py
+        working-directory: ${{ runner.temp }}/pypi-test
+        env:
+          MINIONPY_IMPORT_MODE: installed
