### 2026-06-14 #10145 @khoirulmuzakka
@anugrahjo I have released version v1.6.1, which has been successfully tested on Linux, Windows, and macOS (Apple Silicon). Additionally, I updated the wheel-building CI to install the generated wheel and run tests automatically. I have also created a new pypi_test CI workflow to verify the uploaded wheel across all three platforms.
