+++ minion/src/bindings.cpp
@@ -99,6 +99,21 @@ PYBIND11_MODULE(minionpycpp, m) {
 
         .def("optimize", &LSRTDE::optimize, py::call_guard<py::gil_scoped_release>());
 
+    py::class_<RDEX, MinimizerBase>(m, "RDEX")
+        .def(py::init<MinionFunction, const std::vector<std::pair<double, double>>&,
+                      const std::vector<std::vector<double>>&, void*, std::function<void(MinionResult*)>,
+                      size_t, int, std::map<std::string, ConfigValue> >(),
+            py::arg("func"),
+            py::arg("bounds"),
+            py::arg("x0") = std::vector<std::vector<double>>(),
+            py::arg("data") = nullptr,
+            py::arg("callback") = nullptr,
+            py::arg("maxevals") = 100000,
+            py::arg("seed") = -1,
+            py::arg("options") = std::map<std::string, ConfigValue>())
+
+        .def("optimize", &RDEX::optimize, py::call_guard<py::gil_scoped_release>());
+
     py::class_<LSHADE, Differential_Evolution>(m, "LSHADE")
         .def(py::init<MinionFunction, const std::vector<std::pair<double, double>>&,
                       const std::vector<std::vector<double>>&, void*, std::function<void(MinionResult*)>,
