+++ minion/src/rdex.cpp
@@ -1,0 +1,508 @@
+#include "rdex.h"
+
+#include <algorithm>
+#include <array>
+
+#include "default_options.h"
+
+namespace minion {
+
+void RDEX::initialize() {
+    auto defaultKey = DefaultSettings().getDefaultSettings("RDEX");
+    for (const auto& el : optionMap) {
+        defaultKey[el.first] = el.second;
+    }
+    Options options(defaultKey);
+
+    boundStrategy = options.get<std::string>("bound_strategy", "random");
+    const std::vector<std::string> allBoundStrategy = {"random", "reflect", "reflect-random", "clip", "periodic", "none"};
+    if (std::find(allBoundStrategy.begin(), allBoundStrategy.end(), boundStrategy) == allBoundStrategy.end()) {
+        std::cerr << "Bound stategy '" << boundStrategy << "' is not recognized. 'random' will be used.\n";
+        boundStrategy = "random";
+    }
+
+    populSize = options.get<int>("population_size", 0);
+    if (populSize == 0) {
+        populSize = int(18 * bounds.size());
+    }
+
+    memorySize = options.get<int>("memory_size", 5);
+    successRate = options.get<double>("success_rate", 0.5);
+    ebHybridRateInit = options.get<double>("eb_hybrid_rate_init", 0.7);
+    ebHybridRate = ebHybridRateInit;
+    perturbationRate = options.get<double>("perturbation_rate", 0.4);
+    maxFEval = int(maxevals);
+
+    initialize_population(populSize, int(bounds.size()));
+    hasInitialized = true;
+}
+
+void RDEX::qSort2int(double* mass, int* mass2, int low, int high) {
+    int i = low;
+    int j = high;
+    const double x = mass[(low + high) >> 1];
+    do {
+        while (mass[i] < x) {
+            ++i;
+        }
+        while (mass[j] > x) {
+            --j;
+        }
+        if (i <= j) {
+            std::swap(mass[i], mass[j]);
+            std::swap(mass2[i], mass2[j]);
+            ++i;
+            --j;
+        }
+    } while (i <= j);
+    if (low < j) {
+        qSort2int(mass, mass2, low, j);
+    }
+    if (i < high) {
+        qSort2int(mass, mass2, i, high);
+    }
+}
+
+void RDEX::initialize_population(int newNInds, int newNVars) {
+    nVars = newNVars;
+    nIndsCurrent = newNInds;
+    nIndsFront = newNInds;
+    nIndsFrontMax = newNInds;
+    populSize = newNInds * 2;
+    generation = 0;
+    chosenOne = 0;
+    memoryIter = 0;
+    successFilled = 0;
+    globalbestinit = false;
+    nfeval = 0;
+    pfIndex = 0;
+
+    popul.assign(populSize, std::vector<double>(nVars, 0.0));
+    populFront.assign(nIndsFront, std::vector<double>(nVars, 0.0));
+    populTemp.assign(populSize, std::vector<double>(nVars, 0.0));
+    fitArr.assign(populSize, 0.0);
+    fitArrCopy.assign(populSize, 0.0);
+    fitArrFront.assign(nIndsFront, 0.0);
+    weights.assign(populSize, 0.0);
+    tempSuccessCr.assign(populSize, 0.0);
+    tempSuccessF.assign(populSize, 0.0);
+    fitDelta.assign(populSize, 0.0);
+    memoryCr.assign(memorySize, 1.0);
+    memoryF.assign(memorySize, 1.0);
+    trial.assign(nVars, 0.0);
+    indices.assign(populSize, 0);
+    indices2.assign(populSize, 0);
+    fitMass.assign(nIndsFrontMax, 0.0);
+
[…]
