### 2026-06-27 a734337
FIxed memory leaks the CEC suites
