+++ minion/cec/cec2017.cpp
@@ -20,7 +20,7 @@ CEC2017Functions::CEC2017Functions(int function_number, int dimension)
 	}
 
 namespace CEC2017{
-	void cec17_test_func(double *x, double *f, int nx, int mx,int func_num)
+    void cec17_test_func(double *x, double *f, int nx, int mx,int func_num)
     {
         int cf_num=10,i,j;
         if (ini_flag==1)
@@ -33,144 +33,154 @@ namespace CEC2017{
 
         if (ini_flag==0)
         {
-            FILE *fpt;
+            FILE *fpt = nullptr;
             char FileName[256];
-            free(M);
-            free(OShift);
-            free(y);
-            free(z);
-            free(x_bound);
+            resetThreadLocalCECState();
 
-            OShift=nullptr;
-            M=nullptr;
-            
-            y=(double *)malloc(sizeof(double)  *  nx);
-            z=(double *)malloc(sizeof(double)  *  nx);
-            x_bound=(double *)malloc(sizeof(double)  *  nx);
-            for (i=0; i<nx; i++)
-                x_bound[i]=100.0;
-
-            if (!(nx==2||nx==10||nx==20||nx==30||nx==50||nx==100))
+            try
             {
-                throw std::runtime_error("\nError: Test functions are only defined for D=2,10,20,30,50,100.\n");
-            }
-            if (nx==2&&((func_num>=17&&func_num<=22)||(func_num>=29&&func_num<=30)))
-            {
-                throw std::runtime_error("\nError: hf01,hf02,hf03,hf04,hf05,hf06,cf07&cf08 are NOT defined for D=2.\n");
-            }
-
-            /* Load Matrix M*/
-            snprintf(FileName, sizeof(FileName), "%s/input_data_2017/M_%d_D%d.txt", dirPath.c_str(), func_num,nx);
-            fpt = fopen(FileName,"r");
-            if (fpt==NULL)
-            {
-                throw std::runtime_error("\n Error: Cannot open input file for reading \n");
-                std::cout << FileName << "\n";
-            }
-            if (func_num<20)
-            {
-                M=(double*)malloc(nx*nx*sizeof(double));
-                if (M==NULL)
+                y=(double *)malloc(sizeof(double)  *  nx);
+                z=(double *)malloc(sizeof(double)  *  nx);
+                x_bound=(double *)malloc(sizeof(double)  *  nx);
+                if (y==NULL || z==NULL || x_bound==NULL)
                     throw std::runtime_error("\nError: there is insufficient memory available!\n");
-                for (i=0; i<nx*nx; i++)
+
+                for (i=0; i<nx; i++)
+                    x_bound[i]=100.0;
+
+                if (!(nx==2||nx==10||nx==20||nx==30||nx==50||nx==100))
                 {
-                    fscanf(fpt,"%lf",&M[i]);
+                    throw std::runtime_error("\nError: Test functions are only defined for D=2,10,20,30,50,100.\n");
                 }
-            }
-            else
-            {
-                M=(double*)malloc(cf_num*nx*nx*sizeof(double));
-                if (M==NULL)
+                if (nx==2&&((func_num>=17&&func_num<=22)||(func_num>=29&&func_num<=30)))
+                {
+                    throw std::runtime_error("\nError: hf01,hf02,hf03,hf04,hf05,hf06,cf07&cf08 are NOT defined for D=2.\n");
+                }
+
+                /* Load Matrix M*/
+                snprintf(FileName, sizeof(FileName), "%s/input_data_2017/M_%d_D%d.txt", dirPath.c_str(), func_num,nx);
+                fpt = fopen(FileName,"r");
+                if (fpt==NULL)
+                {
[…]
