+++ minion/cec/cec2022.cpp
@@ -39,127 +39,136 @@ namespace CEC2022 {
 
         if (ini_flag==0)
         {
-            FILE *fpt;
+            FILE *fpt = nullptr;
             char FileName[256];
-            free(M);
-            free(OShift);
-            free(y);
-            free(z);
-            free(x_bound);
+            resetThreadLocalCECState();
 
-            OShift=nullptr;
-            M=nullptr;
-            
-            y=(double *)malloc(sizeof(double)  *  nx);
-            z=(double *)malloc(sizeof(double)  *  nx);
-            x_bound=(double *)malloc(sizeof(double)  *  nx);
-            for (i=0; i<nx; i++)
-                x_bound[i]=100.0;
-
-            if (!(nx==2||nx==10||nx==20))
+            try
             {
-                throw std::runtime_error("\nError: Test functions are only defined for D=2,10,20.\n");
-            }
-            if (nx==2&&(func_num==6||func_num==7||func_num==8))
-            {
-                throw std::runtime_error("\nError:  NOT defined for D=2.\n");
-            }
-
-            /* Load Matrix M*/
-            snprintf(FileName, sizeof(FileName), "%s/input_data_2022/M_%d_D%d.txt", dirPath.c_str(), func_num,nx);
-            fpt = fopen(FileName,"r");
-            if (fpt==NULL)
-            {
-                throw std::runtime_error("\n Error: Cannot open M_%d_D%d.txt for reading \n");
-            }
-            if (func_num<9)
-            {
-                M=(double*)malloc(nx*nx*sizeof(double));
-                if (M==NULL)
+                y=(double *)malloc(sizeof(double)  *  nx);
+                z=(double *)malloc(sizeof(double)  *  nx);
+                x_bound=(double *)malloc(sizeof(double)  *  nx);
+                if (y==NULL || z==NULL || x_bound==NULL)
                     throw std::runtime_error("\nError: there is insufficient memory available!\n");
-                for (i=0; i<nx*nx; i++)
-                {
-                    fscanf(fpt,"%lf",&M[i]);
-                }
-            }
-            else
-            {
-                M=(double*)malloc(cf_num*nx*nx*sizeof(double));
-                if (M==NULL)
-                    throw std::runtime_error("\nError: there is insufficient memory available!\n");
-                for (i=0; i<cf_num*nx*nx; i++)
-                {
-                    fscanf(fpt,"%lf",&M[i]);
-                }
-            }
-            fclose(fpt);
+                for (i=0; i<nx; i++)
+                    x_bound[i]=100.0;
 
-            /* Load shift_data */
-            snprintf(FileName, sizeof(FileName), "%s/input_data_2022/shift_data_%d.txt",  dirPath.c_str(), func_num);
-            fpt = fopen(FileName,"r");
-            if (fpt==NULL)
-            {
-                throw std::runtime_error("\n Error: Cannot open shift_data_%d.txt for reading \n");
-            }
-
-            if (func_num<9)
-            {
-                OShift=(double *)malloc(nx*sizeof(double));
-                if (OShift==NULL)
-                throw std::runtime_error("\nError: there is insufficient memory available!\n");
-                for(i=0;i<nx;i++)
+                if (!(nx==2||nx==10||nx==20))
                 {
-                    fscanf(fpt,"%lf",&OShift[i]);
+                    throw std::runtime_error("\nError: Test functions are only defined for D=2,10,20.\n");
                 }
-            }
-            else
-            {
-                //OShift=(double *)malloc(nx*sizeof(double));
-                OShift=(double *)malloc(nx*cf_num*sizeof(double));
-                if (OShift==NULL)
-                throw std::runtime_error("\nError: there is insufficient memory available!\n");
-                for(i=0;i<cf_num-1;i++)
[…]
