### 2026-06-28 bc7c43c
Improved robustness of CMAES family due to eigen decomposition failed
