+++ minion/src/rcmaes.cpp
@@ -185,6 +185,7 @@ void RCMAES::Parameter::reserve(size_t n_offsprings_reserve_, size_t n_parents_r
     p_s = Eigen::VectorXd::Zero(static_cast<Eigen::Index>(n_params));
     eigvals_C = Eigen::VectorXd::Ones(static_cast<Eigen::Index>(n_params));
     C = Eigen::MatrixXd::Identity(static_cast<Eigen::Index>(n_params), static_cast<Eigen::Index>(n_params));
+    C_pure = Eigen::MatrixXd::Identity(static_cast<Eigen::Index>(n_params), static_cast<Eigen::Index>(n_params));
     C_invsqrt = Eigen::MatrixXd::Identity(static_cast<Eigen::Index>(n_params), static_cast<Eigen::Index>(n_params));
     B = Eigen::MatrixXd::Identity(static_cast<Eigen::Index>(n_params), static_cast<Eigen::Index>(n_params));
     D = Eigen::MatrixXd::Identity(static_cast<Eigen::Index>(n_params), static_cast<Eigen::Index>(n_params));
@@ -262,6 +263,7 @@ void RCMAES::Parameter::reinit(
 
     sigma = sigma_;
     C.setIdentity();
+    C_pure.setIdentity();
     C_invsqrt.setIdentity();
     B.setIdentity();
     D.setIdentity();
@@ -557,12 +559,23 @@ void RCMAES::updateWeights() {
 void RCMAES::updateCovarianceMatrix() {
     double h1 = (1.0 - era.h_sig) * era.c_c * (2.0 - era.c_c);
     double w_sum = era.w.head(static_cast<Eigen::Index>(era.n_offsprings)).sum();
+    double w_sum_pos = era.w.head(static_cast<Eigen::Index>(era.n_parents)).sum();
+    Eigen::MatrixXd previousC = era.C;
+
+    Eigen::MatrixXd yWeightedPure = era.y_offsprings_ranked.block(0, 0, static_cast<Eigen::Index>(era.n_params), static_cast<Eigen::Index>(era.n_parents)) *
+                                    era.w.head(static_cast<Eigen::Index>(era.n_parents)).asDiagonal() *
+                                    era.y_offsprings_ranked.block(0, 0, static_cast<Eigen::Index>(era.n_params), static_cast<Eigen::Index>(era.n_parents)).transpose();
 
     Eigen::MatrixXd yWeighted = era.y_offsprings_ranked.block(0, 0, static_cast<Eigen::Index>(era.n_params), static_cast<Eigen::Index>(era.n_offsprings)) *
                                 era.w_var.head(static_cast<Eigen::Index>(era.n_offsprings)).asDiagonal() *
                                 era.y_offsprings_ranked.block(0, 0, static_cast<Eigen::Index>(era.n_params), static_cast<Eigen::Index>(era.n_offsprings)).transpose();
 
-    era.C = (1.0 + era.c_1 * h1 - era.c_1 - era.c_mu * w_sum) * era.C +
+    era.C_pure = (1.0 + era.c_1 * h1 - era.c_1 - era.c_mu * w_sum_pos) * previousC +
+                 era.c_1 * (era.p_c * era.p_c.transpose()) +
+                 era.c_mu * yWeightedPure;
+    era.C_pure = 0.5 * (era.C_pure + era.C_pure.transpose());
+
+    era.C = (1.0 + era.c_1 * h1 - era.c_1 - era.c_mu * w_sum) * previousC +
              era.c_1 * (era.p_c * era.p_c.transpose()) +
              era.c_mu * yWeighted;
[…]
