### 2026-06-29 42ee7d4
FIrst comparison
