+++ examples/main_cec_multi_libcmaes.cpp
@@ -1,0 +1,264 @@
+#define minimize_cec_functions minimize_cec_functions_minion
+#define main main_cec_multi_minion
+#include "main_cec_multi.cpp"
+#undef main
+#undef minimize_cec_functions
+
+#ifdef CMAES_H
+#undef CMAES_H
+#endif
+#include <libcmaes/cmaes.h>
+
+namespace {
+
+libcmaes::CMASolutions run_libcmaes_optimizer(
+    minion::CECBase* cecfunc,
+    int effective_dimension,
+    const std::vector<std::pair<double, double>>& bounds,
+    int population_size,
+    int max_evals,
+    const std::string& algo,
+    int seed,
+    double sigma0,
+    MinEvLogger* logger) {
+    std::vector<double> x0;
+    x0.reserve(bounds.size());
+    for (const auto& bound : bounds) {
+        x0.push_back(0.5 * (bound.first + bound.second));
+    }
+
+    std::vector<double> lbounds;
+    std::vector<double> ubounds;
+    lbounds.reserve(bounds.size());
+    ubounds.reserve(bounds.size());
+    for (const auto& bound : bounds) {
+        lbounds.push_back(bound.first);
+        ubounds.push_back(bound.second);
+    }
+
+    using GenoPhenoT = libcmaes::GenoPheno<libcmaes::pwqBoundStrategy>;
+    GenoPhenoT gp(lbounds.data(), ubounds.data(), effective_dimension);
+    libcmaes::CMAParameters<GenoPhenoT> params(
+        x0,
+        sigma0 > 0.0 ? sigma0 : 0.3,
+        population_size > 0 ? population_size : -1,
+        static_cast<std::uint64_t>(seed),
+        gp);
+    params.set_str_algo(algo);
+    params.set_max_fevals(max_evals);
+    params.set_quiet(true);
+
+    libcmaes::FitFunc fitfunc = [cecfunc, logger](const double* x, const int& n) -> double {
+        std::vector<std::vector<double>> candidate(1, std::vector<double>(x, x + n));
+        auto values = cecfunc->operator()(candidate);
+        if (logger) {
+            logger->update(values);
+        }
+        return values.empty() ? std::numeric_limits<double>::infinity() : values.front();
+    };
+
+    return libcmaes::cmaes(fitfunc, params);
+}
+
+} // namespace
+
+double minimize_cec_functions(int function_number,
+                              int dimension,
+                              int population_size,
+                              int max_evals,
+                              int year = 2022,
+                              std::string algo = "acmaes",
+                              int seed = -1,
+                              MinEvLogger* logger = nullptr) {
+    minion::CECBase* cecfunc;
+    std::vector<std::pair<double, double>> bounds;
+    int effective_dimension = get_effective_dimension(function_number, dimension, year);
+    if (year == 2019) {
+        for (int i = 0; i < effective_dimension; i++) {
+            if (function_number == 1) bounds.push_back(std::make_pair(-8192, 8192));
+            else if (function_number == 2) bounds.push_back(std::make_pair(-16384, 16384));
+            else if (function_number == 3) bounds.push_back(std::make_pair(-4, 4));
+            else bounds.push_back(std::make_pair(-100, 100));
+        }
+    } else if (year == 2011) {
+        const auto& problem = get_cec2011_problem(function_number);
+        effective_dimension = problem.dimension;
+        bounds = problem.bounds;
+    } else {
+        bounds = std::vector<std::pair<double, double>>(effective_dimension, std::make_pair(-100.0, 100.0));
+    }
+
+    if (year == 2020) cecfunc = new minion::CEC2020Functions(function_number, effective_dimension);
[…]

[…]
