+++ CMakeLists.txt
@@ -28,10 +28,8 @@ option(MINION_BUILD_CEC "Build compiled CEC library" OFF)
 option(MINION_BUILD_PYTHON "Build Python extension module" ${_default_python})
 option(MINION_BUILD_EXAMPLES "Build C++ examples" OFF)
 option(MINION_BUILD_TESTS "Build repository test targets" OFF)
-option(MINION_BUILD_LIBCMAES "Build libcmaes comparison example" OFF)
-
-set(MINION_LIBCMAES_DIR "${CMAKE_CURRENT_SOURCE_DIR}/external/libcmaes/build" CACHE PATH
-    "Path to a built libcmaes package with libcmaes-config.cmake")
+option(MINION_BUILD_LIBCMAES "Build libcmaes comparison example" ON)
+option(MINION_FETCH_LIBCMAES "Fetch libcmaes when building the libcmaes example" ON)
 
 file(GLOB MINION_ALGO_IMPL_SOURCES minion/src/*.cpp)
 list(REMOVE_ITEM MINION_ALGO_IMPL_SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/minion/src/bindings.cpp)
@@ -44,6 +42,23 @@ option(MINION_FETCH_EIGEN "Fetch Eigen headers when Eigen3 package is not found"
 # | Eigen (find or download)   |
 # + -------------------------- +
 find_package(Eigen3 QUIET CONFIG)
+set(_MINION_EIGEN3_PACKAGE_USABLE FALSE)
+if(Eigen3_FOUND AND TARGET Eigen3::Eigen)
+    get_target_property(_MINION_EIGEN3_INCLUDE_DIRS Eigen3::Eigen INTERFACE_INCLUDE_DIRECTORIES)
+    if(_MINION_EIGEN3_INCLUDE_DIRS)
+        foreach(_MINION_EIGEN3_INCLUDE_DIR IN LISTS _MINION_EIGEN3_INCLUDE_DIRS)
+            if(EXISTS "${_MINION_EIGEN3_INCLUDE_DIR}")
+                set(_MINION_EIGEN3_PACKAGE_USABLE TRUE)
+                break()
+            endif()
+        endforeach()
+    endif()
+endif()
+
+if(NOT _MINION_EIGEN3_PACKAGE_USABLE)
+    set(Eigen3_FOUND FALSE)
+endif()
+
 if(NOT Eigen3_FOUND)
     if(NOT MINION_FETCH_EIGEN)
         message(FATAL_ERROR "Eigen3 not found. Install Eigen3 or set MINION_FETCH_EIGEN=ON.")
@@ -62,9 +77,56 @@ if(NOT Eigen3_FOUND)
 
     if(NOT TARGET Eigen3::Eigen)
         add_library(Eigen3::Eigen INTERFACE IMPORTED)
-        set_target_properties(Eigen3::Eigen PROPERTIES
-            INTERFACE_INCLUDE_DIRECTORIES "${eigen_SOURCE_DIR}"
+    endif()
+    set_target_properties(Eigen3::Eigen PROPERTIES
+        INTERFACE_INCLUDE_DIRECTORIES "${eigen_SOURCE_DIR}"
+    )
+endif()
+
+if(MINION_BUILD_LIBCMAES AND NOT Eigen3_FOUND)
+    # libcmaes expects a real Eigen3 package, so provide a tiny config shim
+    # when we only have a fetched source tree.
+    set(MINION_EIGEN3_PACKAGE_DIR "${CMAKE_CURRENT_BINARY_DIR}/_minion_eigen3_package")
[…]
