### 2026-06-29 6ef838a
Refactored cmaes
