+++ minion/src/cmaes.cpp
@@ -1,5 +1,4 @@
 #include "cmaes.h"
-#include "default_options.h"
 #include "utility.h"
 
 #include <algorithm>
@@ -17,126 +16,10 @@ CMAES::CMAES(
     size_t maxevals,
     int seed,
     std::map<std::string, ConfigValue> options)
-    : MinimizerBase(func, bounds, x0, data, callback, maxevals, seed, options) {}
+    : CMAESBase(func, bounds, x0, data, callback, maxevals, seed, std::move(options)) {}
 
 void CMAES::initialize() {
-    auto defaults = DefaultSettings().getDefaultSettings("CMAES");
-    for (const auto& item : optionMap) {
-        defaults[item.first] = item.second;
-    }
-    Options options(defaults);
-
-    dimension = bounds.size();
-    if (dimension == 0) {
-        throw std::runtime_error("CMAES requires bounded variables");
-    }
-
-    useBounds = !bounds.empty();
-
-    lambda = static_cast<size_t>(options.get<int>("population_size", 0));
-    if (lambda == 0) {
-        lambda = 4 + static_cast<size_t>(std::floor(3.0 * std::log(static_cast<double>(dimension)))) ;
-    }
-    lambda = std::max<size_t>(lambda, 5);
-
-    mu = static_cast<size_t>(options.get<int>("mu", 0));
-    if (mu == 0) {
-        mu = lambda / 2;
-    }
-    mu = std::max<size_t>(mu, 1);
-    mu = std::min(mu, lambda);
-
-    weights.resize(mu);
-    for (size_t i = 0; i < mu; ++i) {
-        weights[i] = std::log(static_cast<double>(mu) + 0.5) - std::log(static_cast<double>(i) + 1.0);
-    }
-    double weightSum = std::accumulate(weights.begin(), weights.end(), 0.0);
-    for (double& w : weights) {
-        w /= weightSum;
-    }
-    muEff = 0.0;
-    for (double w : weights) {
-        muEff += w * w;
-    }
-    muEff = 1.0 / muEff;
-
-    sigma = options.get<double>("initial_step", 0.3);
-    stoppingTol = getConvergenceTolerance(options, 1e-4);
-    if (sigma <= 0.0) {
-        sigma = 0.3;
-    }
-
-    double avgRange = 0.0;
-    for (const auto& b : bounds) {
-        avgRange += (b.second - b.first);
-    }
-    avgRange = (dimension > 0) ? avgRange / static_cast<double>(dimension) : 1.0;
-
-    sigma *= avgRange;
-
-    double ccDefault = (4.0 + muEff / static_cast<double>(dimension)) / (dimension + 4.0 + 2.0 * muEff / static_cast<double>(dimension));
-    double csDefault = (muEff + 2.0) / (static_cast<double>(dimension) + muEff + 5.0);
-    double c1Default = 2.0 / ((dimension + 1.3) * (dimension + 1.3) + muEff);
-    double cmuDefault = std::min(1.0 - c1Default, 2.0 * (muEff - 2.0 + 1.0 / muEff) / ((dimension + 2.0) * (dimension + 2.0) + muEff));
-
-    cc = options.get<double>("cc", ccDefault);
-    if (cc <= 0.0) cc = ccDefault;
-    cs = options.get<double>("cs", csDefault);
-    if (cs <= 0.0) cs = csDefault;
-    c1 = options.get<double>("c1", c1Default);
-    if (c1 <= 0.0) c1 = c1Default;
-    cmu = options.get<double>("cmu", cmuDefault);
-    if (cmu <= 0.0) cmu = cmuDefault;
[…]
