### 2026-06-29 468bd1b
Refactor bipop and rcmaes
