+++ minion/src/bipop_acmaes.cpp
@@ -7,6 +7,7 @@
 #include <cmath>
 #include <limits>
 #include <numeric>
+#include <stdexcept>
 
 namespace minion {
 
@@ -19,106 +20,7 @@ BIPOP_aCMAES::BIPOP_aCMAES(
     size_t maxevals,
     int seed,
     std::map<std::string, ConfigValue> options)
-    : MinimizerBase(func, bounds, x0, data, callback, maxevals, seed, options) {}
-
-void BIPOP_aCMAES::Parameter::reserve(size_t n_offsprings_reserve_, size_t n_parents_reserve_, size_t n_params_) {
-    n_params = n_params_;
-    n_offsprings_reserve = n_offsprings_reserve_;
-    n_parents_reserve = n_parents_reserve_;
-
-    x_offsprings.resize(static_cast<Eigen::Index>(n_params), static_cast<Eigen::Index>(n_offsprings_reserve));
-    x_parents_ranked.resize(static_cast<Eigen::Index>(n_params), static_cast<Eigen::Index>(n_parents_reserve));
-    z_offsprings.resize(static_cast<Eigen::Index>(n_params), static_cast<Eigen::Index>(n_offsprings_reserve));
-    y_offsprings.resize(static_cast<Eigen::Index>(n_params), static_cast<Eigen::Index>(n_offsprings_reserve));
-    y_offsprings_ranked.resize(static_cast<Eigen::Index>(n_params), static_cast<Eigen::Index>(n_offsprings_reserve));
-    f_offsprings = Eigen::VectorXd::Constant(static_cast<Eigen::Index>(n_offsprings_reserve), std::numeric_limits<double>::infinity());
-    w = Eigen::VectorXd::Zero(static_cast<Eigen::Index>(n_offsprings_reserve));
-    w_var = Eigen::VectorXd::Zero(static_cast<Eigen::Index>(n_offsprings_reserve));
-    y_mean = Eigen::VectorXd::Zero(static_cast<Eigen::Index>(n_params));
-    x_mean = Eigen::VectorXd::Zero(static_cast<Eigen::Index>(n_params));
-    x_mean_old = Eigen::VectorXd::Zero(static_cast<Eigen::Index>(n_params));
-    p_c = Eigen::VectorXd::Zero(static_cast<Eigen::Index>(n_params));
-    p_s = Eigen::VectorXd::Zero(static_cast<Eigen::Index>(n_params));
-    eigvals_C = Eigen::VectorXd::Ones(static_cast<Eigen::Index>(n_params));
-    C = Eigen::MatrixXd::Identity(static_cast<Eigen::Index>(n_params), static_cast<Eigen::Index>(n_params));
-    C_pure = Eigen::MatrixXd::Identity(static_cast<Eigen::Index>(n_params), static_cast<Eigen::Index>(n_params));
-    C_invsqrt = Eigen::MatrixXd::Identity(static_cast<Eigen::Index>(n_params), static_cast<Eigen::Index>(n_params));
-    B = Eigen::MatrixXd::Identity(static_cast<Eigen::Index>(n_params), static_cast<Eigen::Index>(n_params));
-    D = Eigen::MatrixXd::Identity(static_cast<Eigen::Index>(n_params), static_cast<Eigen::Index>(n_params));
-    keys_offsprings.resize(n_offsprings_reserve);
-}
-
-void BIPOP_aCMAES::Parameter::reinit(size_t n_offsprings_, size_t n_parents_, size_t n_params_, const Eigen::VectorXd& x_mean_, double sigma_) {
-    n_params = n_params_;
-    n_offsprings = std::min(n_offsprings_, n_offsprings_reserve);
-    n_parents = std::min(n_parents_, n_parents_reserve);
-    i_iteration = 0;
-    i_func_eval = 0;
-
[…]

[…]
