+++ minion/src/rcmaes.cpp
@@ -105,6 +105,49 @@ std::vector<double> RCMAES::denormalizePoint(const std::vector<double>& candidat
     return candidate;
 }
 
+Eigen::VectorXd RCMAES::buildCustomActiveWeights(size_t lambdaValue, size_t muValue) const {
+    Eigen::VectorXd signedWeights = Eigen::VectorXd::Zero(static_cast<Eigen::Index>(lambdaValue));
+
+    double wNegSum = 0.0;
+    double wPosSum = 0.0;
+    for (size_t i = 0; i < lambdaValue; ++i) {
+        const double wi =
+            std::log((static_cast<double>(lambdaValue) + 1.0) / 2.0) - std::log(static_cast<double>(i) + 1.0);
+        signedWeights(static_cast<Eigen::Index>(i)) = wi;
+        if (wi >= 0.0) {
+            wPosSum += wi;
+        } else {
+            wNegSum += wi;
+        }
+    }
+
+    double wSumParent = 0.0;
+    double wSqSumParent = 0.0;
+    for (size_t i = 0; i < muValue; ++i) {
+        const double wi = signedWeights(static_cast<Eigen::Index>(i));
+        wSumParent += wi;
+        wSqSumParent += wi * wi;
+    }
+    const double muEffSigned = (wSqSumParent > 0.0) ? ((wSumParent * wSumParent) / wSqSumParent) : 1.0;
+
+    const double dim = static_cast<double>(dimension);
+    const double aMu = 1.0 + c1 / std::max(1e-12, cmu);
+    const double aMueff = 1.0 + 2.0 * muEffSigned;
+    const double aPosdef = (1.0 - c1 - cmu) / (dim * std::max(1e-12, cmu));
+    const double aMin = std::min({aMu, aMueff, aPosdef});
+
+    for (size_t i = 0; i < lambdaValue; ++i) {
+        const double wi = signedWeights(static_cast<Eigen::Index>(i));
+        if (wi >= 0.0) {
+            signedWeights(static_cast<Eigen::Index>(i)) = wi / std::max(1e-12, wPosSum);
+        } else {
+            signedWeights(static_cast<Eigen::Index>(i)) = aMin * wi / std::max(1e-12, std::abs(wNegSum));
+        }
+    }
+
+    return signedWeights;
+}
+
 void RCMAES::configurePopulationParameters(size_t lambdaValue) {
     lambda = std::min<size_t>(2000, std::max<size_t>(lambdaValue, 4));
     mu = std::max<size_t>(static_cast<size_t>(std::ceil(mu_ratio * static_cast<double>(lambda))), 1);
@@ -175,6 +218,7 @@ void RCMAES::initialize() {
     if (sigma0 <= 0.0) {
         sigma0 = 0.3;
     }
+    useCustomActive = options.getSilent<bool>("useCustomActive", true);
 
     cov_scale.clear();
     cov_scale.reserve(dimension);
@@ -402,16 +446,62 @@ MinionResult RCMAES::optimize() {
             pc = (1.0 - cc) * pc + hsig * std::sqrt(cc * (2.0 - cc) * muEff) * y_w;
 
             const Eigen::MatrixXd CinvSqrt = B * D.cwiseInverse().asDiagonal() * B.transpose();
-            const ActiveCMAUpdate activeUpdate =
-                buildActiveCMAUpdate(population, order, meanOld, sigma, weights, CinvSqrt, cmu, muEff);
-
-            const double alphaminusold = 0.5;
             const Eigen::MatrixXd previousC = C;
             const Eigen::MatrixXd spc = pc * pc.transpose();
-            C = (1.0 - c1 - cmu + activeUpdate.cminus * alphaminusold) * previousC +
-                c1 * spc +
[…]
