### 2026-06-30 ba982cc
Refactor cmaes algo family
