+++ minion/include/bipop_acmaes.h
@@ -40,7 +40,6 @@ private:
     std::vector<double> applyBounds(const std::vector<double>& candidate) const;
     std::vector<double> eigenToStd(const Eigen::VectorXd& vec) const;
     void configureRegime(const Eigen::Ref<const Eigen::VectorXd>& startMean, double startSigma, size_t lambdaValue);
-    void applyCovarianceScale();
     void checkStoppingCriteria(bool& shouldStopRun) const;
     size_t runRegime(const Eigen::Ref<const Eigen::VectorXd>& startMean, double startSigma, size_t lambdaValue);
 
@@ -48,11 +47,8 @@ private:
     size_t maxIterations = 0;
 
     double sigma0 = 0.0;
-    double avg_range = 1.0;
 
     Eigen::VectorXd initialMean;
-
-    std::vector<double> cov_scale;
 
     size_t globalGeneration = 0;
 };
