### 2026-06-30 439b60f
Use proper bound handling
