### 2026-06-30 d4c738c
Refactored
