+++ minion/src/acmaes.cpp
@@ -40,39 +40,7 @@ MinionResult ACMAES::optimize() {
             ++generation;
 
             for (size_t k = 0; k < lambda; ++k) {
-                bool valid = false;
-                std::vector<double> candidate(dimension, 0.0);
-                Eigen::VectorXd z = Eigen::VectorXd::Zero(dimension);
-                int retries = 0;
-                while (!valid && retries < 20) {
-                    for (size_t d = 0; d < dimension; ++d) {
-                        z(static_cast<Eigen::Index>(d)) = rand_norm(0.0, 1.0);
-                    }
-                    Eigen::VectorXd step = B * (D.asDiagonal() * z);
-                    Eigen::VectorXd x = mean + sigma * step;
-                    for (size_t d = 0; d < dimension; ++d) {
-                        candidate[d] = x(static_cast<Eigen::Index>(d));
-                    }
-                    if (useBounds) {
-                        bool inside = true;
-                        for (size_t d = 0; d < dimension; ++d) {
-                            if (candidate[d] < bounds[d].first || candidate[d] > bounds[d].second) {
-                                inside = false;
-                                break;
-                            }
-                        }
-                        if (!inside) {
-                            ++retries;
-                            continue;
-                        }
-                    }
-                    valid = true;
-                }
-                if (!valid) {
-                    candidate = applyBounds(candidate);
-                }
-                population[k] = candidate;
-                zs[k] = z;
+                population[k] = sampleCandidate(mean, B, D, sigma);
             }
 
             auto fitVals = func(population, data);
