### 2026-07-01 2a64219
Use center point as initialization
