+++ examples/main_cec_multi_allmaxevals.cpp
@@ -14,8 +14,6 @@
 #include <filesystem>
 #include <iomanip>
 #include <memory>
-#include <atomic>
-#include <mutex>
 
 namespace {
 
@@ -402,7 +400,6 @@ double minimize_cec_functions(int function_number,
 
     auto settings = minion::DefaultSettings().getDefaultSettings(algo);
     settings["population_size"] = popsize;
-    settings["convergence_tol"] = 0.0;
     std::vector<std::vector<double>> x0={};
     if (algo == "NelderMead" || algo == "L_BFGS_B" || algo == "DA"){
         std::vector<double> x00;
@@ -417,6 +414,7 @@ double minimize_cec_functions(int function_number,
                                 logger ? static_cast<void*>(&ctx) : static_cast<void*>(cecfunc),
                                 callBack,
                                 algo,
+                                0.0,
                                 max_evals,
                                 seed,
                                 settings);
@@ -486,7 +484,6 @@ int main(int argc, char* argv[]) {
     std::string algo="ARRDE";
     int popsize=0;
     int year = 2017;
-    int Nmaxevals = int(1e+4*dimension);
     int Nthreads = 1;
     int log_min_ev = 0;
     if (argc > 1) {
@@ -507,13 +504,10 @@ int main(int argc, char* argv[]) {
     }
 
     if (argc > 6) {
-        Nmaxevals = std::atoi(argv[6]); // Use third argument for algo, no conversion needed
+        Nthreads = std::atoi(argv[6]); // Number of worker threads for parallel runs
     }
     if (argc > 7) {
-        Nthreads = std::atoi(argv[7]); // Number of worker threads for parallel runs
-    }
-    if (argc > 8) {
-        log_min_ev = std::atoi(argv[8]);
+        log_min_ev = std::atoi(argv[7]);
     }
 
     Nthreads = std::max(1, Nthreads);
@@ -529,96 +523,94 @@ int main(int argc, char* argv[]) {
     else if (year==2011) funcnums = {1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22};
     else throw std::runtime_error("Year invalid.");
 
-    std::vector<std::vector<double>> results(numRuns);
-    std::unordered_map<int, std::vector<std::vector<double>>> min_ev_logs;
-    if (log_min_ev) {
-        for (int func : funcnums) {
-            int eff_dim = get_effective_dimension(func, dimension, year);
-            size_t step = std::max<size_t>(1, 10 * static_cast<size_t>(eff_dim));
-            size_t count = static_cast<size_t>(Nmaxevals) / step + 1;
-            min_ev_logs.emplace(func, std::vector<std::vector<double>>(count, std::vector<double>(numRuns, std::numeric_limits<double>::infinity())));
+    const std::vector<int> max_eval_multipliers = {
+        500, 1000, 2000, 3000, 4000, 5000, 6000, 7000, 8000, 9000, 10000
+    };
+
+    for (int multiplier : max_eval_multipliers) {
+        const int Nmaxevals = multiplier * dimension;
+        std::vector<std::vector<double>> results(numRuns);
+        std::unordered_map<int, std::vector<std::vector<double>>> min_ev_logs;
+        if (log_min_ev) {
+            for (int func : funcnums) {
+                int eff_dim = get_effective_dimension(func, dimension, year);
+                size_t step = std::max<size_t>(1, 10 * static_cast<size_t>(eff_dim));
+                size_t count = static_cast<size_t>(Nmaxevals) / step + 1;
+                min_ev_logs.emplace(func, std::vector<std::vector<double>>(count, std::vector<double>(numRuns, std::numeric_limits<double>::infinity())));
+            }
[…]
