+++ examples/main_bbob_multi.cpp
@@ -1,0 +1,248 @@
+#include "minion.h"
+#include "bbob2009.h"
+
+#include <algorithm>
+#include <chrono>
+#include <cmath>
+#include <csignal>
+#include <cstdlib>
+#include <atomic>
+#include <iomanip>
+#include <iostream>
+#include <limits>
+#include <map>
+#include <mutex>
+#include <thread>
+#include <string>
+#include <vector>
+
+namespace {
+
+volatile std::sig_atomic_t g_stop_requested = 0;
+
+void handle_sigint(int) {
+    g_stop_requested = 1;
+}
+
+bool is_supported_bbob_dimension(int dimension) {
+    return dimension == 2 || dimension == 3 || dimension == 5 || dimension == 10 || dimension == 20 || dimension == 40;
+}
+
+std::vector<std::string> parse_algorithms(const std::string& csv) {
+    if (csv.empty()) {
+        return {"CMAES", "ACMAES", "RCMAES", "BIPOP_aCMAES"};
+    }
+
+    std::vector<std::string> algos;
+    size_t start = 0;
+    while (start < csv.size()) {
+        const size_t end = csv.find(',', start);
+        const std::string item = csv.substr(start, end == std::string::npos ? std::string::npos : end - start);
+        if (!item.empty()) {
+            algos.push_back(item);
+        }
+        if (end == std::string::npos) {
+            break;
+        }
+        start = end + 1;
+    }
+    return algos;
+}
+
+std::vector<int> parse_functions(int max_function) {
+    std::vector<int> functions;
+    for (int i = 1; i <= max_function; ++i) {
+        functions.push_back(i);
+    }
+    return functions;
+}
+
+minion::MinionResult run_algorithm(
+const minion::BBOB2009Problem& problem,
+    const std::string& algo,
+    int maxevals,
+    int seed,
+    double sigma0,
+    int population_size) {
+    const std::vector<std::pair<double, double>>& bounds = problem.bounds();
+    const std::vector<std::vector<double>> x0 = {problem.initialSolution()};
+
+    minion::MinionFunction objective = [&problem](const std::vector<std::vector<double>>& candidates, void*) {
+        return problem.evaluateBatch(candidates);
+    };
+
+    std::map<std::string, minion::ConfigValue> options;
+    options["population_size"] = population_size;
+    options["rel_initial_step"] = sigma0;
+    options["bound_strategy"] = std::string("reflect-random");
+
+    minion::Minimizer optimizer(
+        objective,
+        bounds,
+        x0,
+        nullptr,
+        nullptr,
+        algo,
+        static_cast<size_t>(maxevals),
+        seed,
+        options);
+    return optimizer.optimize();
+}
+
+}  // namespace
+
+int main(int argc, char* argv[]) {
+    int numRuns = 1;
+    int dimension = 10;
+    std::string algo_csv = "CMAES,ACMAES,RCMAES,BIPOP_aCMAES";
+    int popsize = 0;
+    int year = 2018;
+    int maxevals = static_cast<int>(1e4 * dimension);
+    int Nthreads = 1;
+    int log_min_ev = 0;
+
+    if (argc > 1) {
+        numRuns = std::max(1, std::atoi(argv[1]));
+    }
+    if (argc > 2) {
+        dimension = std::max(1, std::atoi(argv[2]));
+    }
+    if (argc > 3) {
+        algo_csv = argv[3];
+    }
+    if (argc > 4) {
+        popsize = std::atoi(argv[4]);
+    }
+    if (argc > 5) {
+        year = std::atoi(argv[5]);
+    }
+    if (argc > 6) {
[…]
