+++ external/bbob2009/bbob2009.h
@@ -1,0 +1,48 @@
+#ifndef BBOB2009_H
+#define BBOB2009_H
+
+#include <coco.h>
+
+#include <cstddef>
+#include <string>
+#include <utility>
+#include <vector>
+
+namespace minion {
+
+class BBOB2009Problem {
+public:
+    BBOB2009Problem(int function_number, int dimension, int year = 2018);
+    ~BBOB2009Problem();
+
+    BBOB2009Problem(const BBOB2009Problem&) = delete;
+    BBOB2009Problem& operator=(const BBOB2009Problem&) = delete;
+    BBOB2009Problem(BBOB2009Problem&& other) noexcept;
+    BBOB2009Problem& operator=(BBOB2009Problem&& other) noexcept;
+
+    size_t dimension() const;
+    const std::vector<std::pair<double, double>>& bounds() const;
+    const std::vector<double>& initialSolution() const;
+    double bestValue() const;
+    const std::string& id() const;
+    const std::string& name() const;
+
+    double evaluate(const std::vector<double>& x) const;
+    std::vector<double> evaluateBatch(const std::vector<std::vector<double>>& candidates) const;
+    void recommendSolution(const std::vector<double>& x) const;
+
+private:
+    void reset();
+
+    coco_suite_t* suite_ = nullptr;
+    coco_problem_t* problem_ = nullptr;
+    std::vector<std::pair<double, double>> bounds_;
+    std::vector<double> initial_solution_;
+    std::string id_;
+    std::string name_;
+    double best_value_ = 0.0;
+};
+
+}  // namespace minion
+
+#endif

[…]
