+++ examples/main_bbob_multi.cpp
@@ -9,9 +9,11 @@
 #include <atomic>
 #include <iomanip>
 #include <iostream>
+#include <fstream>
 #include <limits>
 #include <map>
 #include <mutex>
+#include <filesystem>
 #include <thread>
 #include <string>
 #include <vector>
@@ -28,33 +30,98 @@ bool is_supported_bbob_dimension(int dimension) {
     return dimension == 2 || dimension == 3 || dimension == 5 || dimension == 10 || dimension == 20 || dimension == 40;
 }
 
-std::vector<std::string> parse_algorithms(const std::string& csv) {
-    if (csv.empty()) {
-        return {"CMAES", "ACMAES", "RCMAES", "BIPOP_aCMAES"};
-    }
-
-    std::vector<std::string> algos;
-    size_t start = 0;
-    while (start < csv.size()) {
-        const size_t end = csv.find(',', start);
-        const std::string item = csv.substr(start, end == std::string::npos ? std::string::npos : end - start);
-        if (!item.empty()) {
-            algos.push_back(item);
-        }
-        if (end == std::string::npos) {
-            break;
-        }
-        start = end + 1;
-    }
-    return algos;
-}
-
 std::vector<int> parse_functions(int max_function) {
     std::vector<int> functions;
     for (int i = 1; i <= max_function; ++i) {
         functions.push_back(i);
     }
     return functions;
+}
+
+struct MinEvLogger {
+    size_t step = 0;
+    size_t evals = 0;
+    size_t next_index = 0;
+    double best = std::numeric_limits<double>::infinity();
+    double optimum = 0.0;
+    std::vector<double> samples;
+
+    MinEvLogger(size_t dimension, size_t max_evals, double global_optimum) {
+        step = std::max<size_t>(1, 10 * dimension);
+        optimum = global_optimum;
+        size_t count = max_evals / step + 1;
+        samples.assign(count, std::numeric_limits<double>::infinity());
+    }
+
+    void update(const std::vector<double>& values) {
+        for (double v : values) {
+            if (v < best) {
+                best = v;
+            }
+            ++evals;
+            while (next_index < samples.size() && evals >= next_index * step) {
+                samples[next_index] = best - optimum;
+                ++next_index;
+            }
+        }
+    }
+
+    void finalize() {
+        while (next_index < samples.size()) {
+            samples[next_index] = best - optimum;
+            ++next_index;
+        }
+    }
+};
+
+struct LogContext {
+    minion::BBOB2009Problem* problem = nullptr;
+    MinEvLogger* logger = nullptr;
+};
+
+std::vector<double> objective_function_logged(const std::vector<std::vector<double>>& candidates, void* data) {
+    auto* ctx = static_cast<LogContext*>(data);
+    auto values = ctx->problem->evaluateBatch(candidates);
+    if (ctx->logger) {
+        ctx->logger->update(values);
+    }
+    return values;
+}
+
+void dumpResultsToFile(const std::vector<std::vector<double>>& results, const std::string& filename) {
+    std::ofstream file(filename);
+    if (!file.is_open()) {
+        std::cerr << "Error: Unable to open file " << filename << std::endl;
+        return;
+    }
+
+    for (const auto& row : results) {
+        for (size_t i = 0; i < row.size(); ++i) {
+            file << row[i];
+            if (i + 1 < row.size()) {
+                file << "\t";
+            }
+        }
+        file << "\n";
+    }
+}
+
+void dumpMatrixToFile(const std::vector<std::vector<double>>& matrix, const std::string& filename) {
+    std::ofstream file(filename);
+    if (!file.is_open()) {
+        std::cerr << "Error: Unable to open file " << filename << std::endl;
+        return;
[…]
