### 2026-07-02 440fce1
Fixed restart bug
