+++ examples/main_bbob_multi_libcmaes.cpp
@@ -1,0 +1,332 @@
+#include <cstdint>
+#include <memory>
+
+#define main main_bbob_multi_minion
+#include "main_bbob_multi.cpp"
+#undef main
+
+#ifdef CMAES_H
+#undef CMAES_H
+#endif
+#include <libcmaes/cmaes.h>
+
+namespace {
+
+std::vector<std::pair<double, double>> normalized_bounds_for(size_t dimension) {
+    return std::vector<std::pair<double, double>>(dimension, std::make_pair(-1.0, 1.0));
+}
+
+std::vector<double> normalize_point(
+    const std::vector<double>& point,
+    const std::vector<std::pair<double, double>>& bounds) {
+    if (bounds.empty()) {
+        return point;
+    }
+
+    std::vector<double> normalized(point.size(), 0.0);
+    for (size_t i = 0; i < point.size(); ++i) {
+        const double low = bounds[i].first;
+        const double high = bounds[i].second;
+        const double range = high - low;
+        normalized[i] = range > 0.0 ? (2.0 * (point[i] - low) / range - 1.0) : 0.0;
+    }
+    return normalized;
+}
+
+std::vector<double> denormalize_point(
+    const std::vector<double>& point,
+    const std::vector<std::pair<double, double>>& bounds) {
+    if (bounds.empty()) {
+        return point;
+    }
+
+    std::vector<double> actual(point.size(), 0.0);
+    for (size_t i = 0; i < point.size(); ++i) {
+        const double low = bounds[i].first;
+        const double high = bounds[i].second;
+        const double range = high - low;
+        actual[i] = range > 0.0 ? (low + 0.5 * (point[i] + 1.0) * range) : low;
+    }
+    return actual;
+}
+
+libcmaes::CMASolutions run_libcmaes_optimizer(
+    const minion::BBOB2009Problem& problem,
+    int population_size,
+    int max_evals,
+    const std::string& algo,
+    int seed,
+    MinEvLogger* logger) {
+    const std::vector<std::pair<double, double>>& bounds = problem.bounds();
+    const std::vector<std::pair<double, double>> normalized_bounds = normalized_bounds_for(bounds.size());
+    const std::vector<double> x0 = normalize_point(problem.initialSolution(), bounds);
+
+    using GenoPhenoT = libcmaes::GenoPheno<libcmaes::pwqBoundStrategy>;
+    std::vector<double> lbounds;
+    std::vector<double> ubounds;
+    lbounds.reserve(normalized_bounds.size());
+    ubounds.reserve(normalized_bounds.size());
+    for (const auto& bound : normalized_bounds) {
+        lbounds.push_back(bound.first);
+        ubounds.push_back(bound.second);
+    }
+
+    GenoPhenoT gp(lbounds.data(), ubounds.data(), static_cast<int>(bounds.size()));
+    libcmaes::CMAParameters<GenoPhenoT> params(
+        x0,
+        0.3,
+        population_size > 0 ? population_size : -1,
+        static_cast<std::uint64_t>(seed),
+        gp);
+    params.set_str_algo(algo);
+    params.set_max_fevals(max_evals);
+    params.set_quiet(true);
+
+    libcmaes::FitFunc fitfunc = [&problem, logger, bounds](const double* x, const int& n) -> double {
+        std::vector<double> normalized_candidate(x, x + n);
+        const std::vector<double> candidate = denormalize_point(normalized_candidate, bounds);
+        const double value = problem.evaluate(candidate);
+        if (logger) {
+            logger->update(std::vector<double>{value});
+        }
+        return value;
+    };
+
+    return libcmaes::cmaes(fitfunc, params);
+}
+
+} // namespace
+
+double minimize_bbob_functions(int function_number,
+                               const minion::BBOB2009Problem& problem,
[…]
