+++ examples/main_cec_multi.cpp
@@ -16,6 +16,7 @@
 #include <memory>
 #include <atomic>
 #include <mutex>
+#include <sstream>
 
 namespace {
 
@@ -319,6 +320,59 @@ struct LogContext {
     MinEvLogger* logger = nullptr;
 };
 
+struct Job {
+    int run_index = 0;
+    size_t function_index = 0;
+    int function_number = 0;
+};
+
+std::string format_eta(std::chrono::seconds seconds) {
+    const auto total_seconds = seconds.count();
+    const auto hours = total_seconds / 3600;
+    const auto minutes = (total_seconds % 3600) / 60;
+    const auto secs = total_seconds % 60;
+
+    std::ostringstream oss;
+    oss << std::setw(2) << std::setfill('0') << hours << ':'
+        << std::setw(2) << std::setfill('0') << minutes << ':'
+        << std::setw(2) << std::setfill('0') << secs;
+    return oss.str();
+}
+
+void render_progress_bar(
+    size_t completed,
+    size_t total,
+    const std::chrono::steady_clock::time_point& start_time) {
+    constexpr size_t kBarWidth = 40;
+    if (total == 0) {
+        std::cout << "\rProgress: [----------------------------------------] 0/0 (0%) ETA --:--:--" << std::flush;
+        return;
+    }
+
+    const double ratio = static_cast<double>(completed) / static_cast<double>(total);
+    const size_t filled = std::min(kBarWidth, static_cast<size_t>(ratio * static_cast<double>(kBarWidth)));
+    std::ostringstream percent;
+    percent << std::fixed << std::setprecision(1) << (ratio * 100.0);
+
+    std::cout << "\rProgress: [";
+    for (size_t i = 0; i < kBarWidth; ++i) {
+        std::cout << (i < filled ? '#' : '-');
+    }
+
+    std::string eta_text = "--:--:--";
+    if (completed > 0 && completed < total) {
+        const auto elapsed = std::chrono::steady_clock::now() - start_time;
+        const double elapsed_sec = std::chrono::duration<double>(elapsed).count();
+        const double sec_per_job = elapsed_sec / static_cast<double>(completed);
+        const double remaining_sec = sec_per_job * static_cast<double>(total - completed);
+        eta_text = format_eta(std::chrono::seconds(static_cast<long long>(std::max(0.0, std::ceil(remaining_sec)))));
+    } else if (completed >= total) {
+        eta_text = "00:00:00";
+    }
+
+    std::cout << "] " << completed << "/" << total << " (" << percent.str() << "%) ETA " << eta_text << std::flush;
+}
+
 std::vector<double> objective_function_logged(const std::vector<std::vector<double>>& x, void* data) {
     auto* ctx = static_cast<LogContext*>(data);
     auto values = ctx->func->operator()(x);
@@ -373,7 +427,8 @@ double minimize_cec_functions(int function_number,
                               int year = 2022,
                               std::string algo = "ARRDE",
                               int seed = -1,
-                              MinEvLogger* logger = nullptr) {
+                              MinEvLogger* logger = nullptr,
+                              std::ostream* out = &std::cout) {
     minion::CECBase* cecfunc;
     std::vector<std::pair<double, double>> bounds;
     int effective_dimension = get_effective_dimension(function_number, dimension, year);
@@ -425,9 +480,11 @@ double minimize_cec_functions(int function_number,
     double ret = result.fun;
 
     // Output the results
-    std::cout << "Optimization Results for Function " << function_number << ":\n";
[…]
