+++ examples/libcmaes_test.py
@@ -1,328 +1,0 @@
-from __future__ import annotations
-
-import math
-import os
-import sys
-from pathlib import Path
-
-import matplotlib.pyplot as plt
-import numpy as np
-
-
-os.environ.setdefault("OMP_NUM_THREADS", "1")
-os.environ.setdefault("MKL_NUM_THREADS", "1")
-
-
-def find_repo_root() -> Path:
-    current = Path(__file__).resolve().parent
-    while True:
-        if (current / "minion").exists() and (current / "minionpy").exists():
-            return current
-        if current.parent == current:
-            raise RuntimeError("Could not locate the repository root")
-        current = current.parent
-
-
-REPO_ROOT = find_repo_root()
-sys.path.insert(0, str(REPO_ROOT))
-sys.path.insert(0, str(REPO_ROOT / "minionpy"))
-
-import cocoex
-import minionpy as mpy
-import libcmaes_bridge
-
-
-def recommended_lambda(dimension: int) -> int:
-    return max(5, 4 + int(math.floor(3.0 * math.log(dimension))))
-
-
-class ProgressTracker:
-    def __init__(self, fopt: float, progress_step: int):
-        self.fopt = float(fopt)
-        self.progress_step = max(1, int(progress_step))
-        self.evals = 0
-        self.best = math.inf
-        self.next_progress = self.progress_step
-        self.eval_history: list[int] = []
-        self.err_history: list[float] = []
-
-    def _record_progress(self) -> None:
-        while self.evals >= self.next_progress:
-            self.eval_history.append(self.next_progress)
-            self.err_history.append(max(abs(self.best - self.fopt), 1e-300))
-            self.next_progress += self.progress_step
-
-    def update(self, values) -> None:
-        for value in np.asarray(values, dtype=float).ravel():
-            self.evals += 1
-            value = float(value)
-            if value < self.best:
-                self.best = value
-            self._record_progress()
-
-    def batch_objective(self, scalar_objective):
-        def wrapped(x_batch):
-            x_batch = np.asarray(x_batch, dtype=float)
-            if x_batch.ndim == 1:
-                x_batch = x_batch[None, :]
-            values = [float(scalar_objective(x)) for x in x_batch]
-            self.update(values)
-            return values
-
-        return wrapped
-
-    def scalar_objective(self, scalar_objective):
-        def wrapped(x):
-            value = float(scalar_objective(np.asarray(x, dtype=float)))
-            self.update([value])
-            return value
-
-        return wrapped
-
-    @property
-    def final_error(self) -> float:
-        return max(abs(self.best - self.fopt), 1e-300)
-
-
-MINION_ALGOS = [
-    ("Minion CMAES", "CMAES", 0.3),
-    ("Minion BIPOP_aCMAES", "BIPOP_aCMAES", 0.3),
-    ("Minion RCMAES", "RCMAES", 0.2),
-]
-
-LIBCMAES_ALGOS = [
-    ("libcmaes cmaes", "cmaes", 0.3),
-    ("libcmaes acmaes", "acmaes", 0.3),
-    ("libcmaes bipop", "bipop", 0.3),
-    ("libcmaes abipop", "abipop", 0.3),
-]
-
-
-def make_cec2017_problem(function_number: int, dimension: int):
-    cec = mpy.CEC2017Functions(function_number, dimension)
-    fopt = 100 * function_number
-    bounds = [(-100.0, 100.0)] * dimension
-    x0 = [0.0] * dimension
-
-    def scalar(x):
-        return float(cec([np.asarray(x, dtype=float).tolist()])[0])
-
-    return scalar, fopt, bounds, x0
-
-
-def make_bbob_problem(function_number: int, dimension: int):
[…]
