+++ compile.sh
@@ -65,9 +65,8 @@ if [[ "${build_type}" == "Debug" ]]; then
   cmake_args=(
     -G "Unix Makefiles"
     "-DCMAKE_BUILD_TYPE=${build_type}"
-    -DMINION_BUILD_CEC=ON
+    -DMINION_BUILD_BENCHMARK=ON
     -DMINION_BUILD_EXAMPLES=ON
-    -DMINION_BUILD_LIBCMAES=ON
     -DMINION_BUILD_PYTHON=ON
     "-DCMAKE_CXX_FLAGS_DEBUG=-O0 -g3 -fno-omit-frame-pointer"
     "-DCMAKE_C_FLAGS_DEBUG=-O0 -g3 -fno-omit-frame-pointer"
@@ -77,9 +76,8 @@ else
   cmake_args=(
     -G "Unix Makefiles"
     "-DCMAKE_BUILD_TYPE=${build_type}"
-    -DMINION_BUILD_CEC=ON
+    -DMINION_BUILD_BENCHMARK=ON
     -DMINION_BUILD_EXAMPLES=ON
-    -DMINION_BUILD_LIBCMAES=ON
     -DMINION_BUILD_PYTHON=ON
     ..
   )

[…]
