### 2026-07-06 322bc8c
Refactor, added benchmark functionality
