+++ minion/benchmark/benchmark.cpp
@@ -214,16 +214,6 @@ std::string format_scientific(double value, int acc) {
     return oss.str();
 }
 
-struct CecLogContext {
-    minion::CECBase* func = nullptr;
-    MinEvLogger* logger = nullptr;
-};
-
-struct BbobLogContext {
-    minion::BBOB2009Problem* problem = nullptr;
-    MinEvLogger* logger = nullptr;
-};
-
 struct Job {
     int run_index = 0;
     size_t function_index = 0;
@@ -250,7 +240,7 @@ std::string format_progress_bar(
     constexpr size_t kBarWidth = 20;
     std::ostringstream oss;
     if (total == 0) {
-        oss << "Progress: [----------------------------------------] 0/0 (0%) ETA --:--:--";
+        oss << "[----------------------------------------] 0/0  0.0%  ETA --:--:--";
         return oss.str();
     }
 
@@ -285,28 +275,6 @@ std::string format_progress_bar(
         << "  ETA " << eta_text;
     return oss.str();
 }
-
-std::vector<double> objective_function_logged(const std::vector<std::vector<double>>& x, void* data) {
-    auto* ctx = static_cast<CecLogContext*>(data);
-    auto values = ctx->func->operator()(x);
-    if (ctx->logger) {
-        ctx->logger->update(values);
-    }
-    return values;
-}
-
-std::vector<double> objective_function_logged_bbob(const std::vector<std::vector<double>>& candidates, void* data) {
-    auto* ctx = static_cast<BbobLogContext*>(data);
-    auto values = ctx->problem->evaluateBatch(candidates);
-    if (ctx->logger) {
-        ctx->logger->update(values);
-    }
-    return values;
-}
-
-void callBack(minion::MinionResult* res) {
-    //std::cout << "Best fitness " << res->fun << "\n";
-};
 
 minion::Benchmark::Benchmark(minion::BenchmarkConfig config) : config_(std::move(config)) {}
 
@@ -378,9 +346,8 @@ double minimize_cec_functions(int function_number,
                               int year = 2022,
                               std::string algo = "ARRDE",
                               int seed = -1,
-                              MinEvLogger* logger = nullptr,
-                              std::ostream* out = &std::cout) {
-    minion::CECBase* cecfunc;
+                              MinEvLogger* logger = nullptr) {
+    std::unique_ptr<minion::CECBase> cecfunc;
     std::vector<std::pair<double, double>> bounds;
     int effective_dimension = get_effective_dimension(function_number, dimension, year);
     if (year==2019) { 
@@ -396,12 +363,12 @@ double minimize_cec_functions(int function_number,
         bounds = problem.bounds;
     } else bounds = std::vector<std::pair<double, double>>(effective_dimension, std::make_pair(-100.0, 100.0));
 
-    if (year==2020) cecfunc = new minion::CEC2020Functions(function_number, effective_dimension);
-    else if (year==2022) cecfunc = new minion::CEC2022Functions(function_number, effective_dimension);
-    else if (year==2017) cecfunc = new minion::CEC2017Functions(function_number, effective_dimension);
-    else if (year==2019) cecfunc = new minion::CEC2019Functions(function_number, effective_dimension);
-    else if (year==2014) cecfunc = new minion::CEC2014Functions(function_number, effective_dimension);
-    else if (year==2011) cecfunc = new minion::CEC2011Functions(function_number, effective_dimension);
+    if (year==2020) cecfunc = std::make_unique<minion::CEC2020Functions>(function_number, effective_dimension);
+    else if (year==2022) cecfunc = std::make_unique<minion::CEC2022Functions>(function_number, effective_dimension);
[…]
