+++ cmake/coco_colorama_shim/colorama.py
@@ -1,0 +1,11 @@
+def init(*args, **kwargs):
+    return None
+
+
+class _Ansi:
+    def __getattr__(self, name):
+        return ""
+
+
+Fore = _Ansi()
+Style = _Ansi()
