+++ docs/source/index.rst
@@ -9,7 +9,7 @@ Welcome to Minion's documentation!
 
 Welcome to the **Minion** library documentation. Minion is a high-performance, derivative-free optimization library designed for solving complex optimization problems where gradient information is unavailable or unreliable. The library implements state-of-the-art algorithms, many of which have won top positions in IEEE CEC competitions, making them highly competitive compared to traditional optimization libraries like SciPy, NLopt, OptimLib, pyGMO, and pagmo2.
 
-Minion also provides a research-friendly environment for developing and evaluating new optimization algorithms. It includes benchmark functions from multiple CEC competitions (2011, 2014, 2017, 2019, 2020, and 2022) to facilitate algorithm testing and comparison.
+Minion also provides a research-friendly environment for developing and evaluating new optimization algorithms. It includes benchmark functions from multiple CEC competitions (2011, 2014, 2017, 2019, 2020, and 2022) and the BBOB2009 suite to facilitate algorithm testing and comparison.
 
 This documentation covers the installation, usage, and details of both the C++ and Python interfaces of Minion.
 

[…]
