+++ minionpy/__init__.py
@@ -7,7 +7,7 @@ from .parallel import *
 from .cec import CEC2011Functions as CEC2011
 
 # Define the package's version
-__version__ = "1.6.1"
+__version__ = "1.7.0"
 
 # Optionally, define the package name
 __name__ = "minionpy"
