+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "minionpy"
-version = "1.6.1"
+version = "1.7.0"
 description = "MinionPy is the Python implementation of the Minion (C++) ibrary."
 readme = "minionpy.md"
 license = "MIT"
