+++ LICENSE.txt
@@ -40,3 +40,16 @@ code was distributed prior to the CEC competitions for research and benchmarking
 to the best of our knowledge, is free to use and redistribute for non-commercial, 
 academic, and research applications.  
 
+
+BBOB2009 / COCO Benchmark Support
+----------------------------------
+
+This project also includes BBOB2009 benchmark support through the COCO
+benchmark framework from the upstream repository
+``https://github.com/numbbo/coco.git``. According to the upstream COCO
+license, the code is generally distributed under the 3-clause BSD License,
+apart from a few documented exceptions. Minion uses the ``code-experiments``
+component from that repository together with the local ``bbob2009`` adapter
+code in ``external/bbob2009``. Please consult the COCO repository ``LICENSE``
+and ``AUTHORS`` files, as well as the ``bbob2009`` sources, for the exact
+license details that apply to those benchmark files.
