+++ docs/source/cec.rst
@@ -292,6 +292,211 @@ For ``BBOB2009``, the problem object exposes the same information:
 For a full C++ per-problem bound setup, see the benchmark implementation in ``minion/benchmark/benchmark.cpp`` and the integration test in ``tests/test_minion.cpp``.
 
 
+CEC Benchmark Function Details
+------------------------------
+
+The tables below summarize the basic benchmark families used by each numbered function in the current CEC2014 and CEC2017 implementations. In the published suite design, CEC2020 and CEC2022 are best understood as selected subsets of the CEC2017-style benchmark family. The implementation does not simply reuse the CEC2017 public function numbers one-for-one; instead, it dispatches through the shared benchmark families. The tables below therefore map each public suite function to the shared family used by the implementation.
+
+CEC2014
+^^^^^^^
+
+.. list-table:: CEC2014 function mapping
+   :header-rows: 1
+
+   * - Function number
+     - Basic functions used
+   * - 1
+     - ``ellips_func``
+   * - 2
+     - ``bent_cigar_func``
+   * - 3
+     - ``discus_func``
+   * - 4
+     - ``rosenbrock_func``
+   * - 5
+     - ``ackley_func``
+   * - 6
+     - ``weierstrass_func``
+   * - 7
+     - ``griewank_func``
+   * - 8
+     - ``rastrigin_func``
+   * - 9
+     - ``rastrigin_func``
+   * - 10
+     - ``schwefel_func``
+   * - 11
+     - ``schwefel_func``
+   * - 12
+     - ``katsuura_func``
+   * - 13
+     - ``happycat_func``
+   * - 14
+     - ``hgbat_func``
+   * - 15
+     - ``grie_rosen_func``
+   * - 16
+     - ``escaffer6_func``
+   * - 17
+     - ``hf01`` = ``schwefel_func`` + ``rastrigin_func`` + ``ellips_func``
+   * - 18
+     - ``hf02`` = ``bent_cigar_func`` + ``hgbat_func`` + ``rastrigin_func``
+   * - 19
+     - ``hf03`` = ``griewank_func`` + ``weierstrass_func`` + ``rosenbrock_func`` + ``escaffer6_func``
+   * - 20
+     - ``hf04`` = ``hgbat_func`` + ``discus_func`` + ``grie_rosen_func`` + ``rastrigin_func``
+   * - 21
+     - ``hf05`` = ``escaffer6_func`` + ``hgbat_func`` + ``rosenbrock_func`` + ``schwefel_func`` + ``ellips_func``
+   * - 22
+     - ``hf06`` = ``katsuura_func`` + ``happycat_func`` + ``grie_rosen_func`` + ``schwefel_func`` + ``ackley_func``
+   * - 23
+     - ``cf01`` = ``rosenbrock_func`` + ``ellips_func`` + ``bent_cigar_func`` + ``discus_func``
+   * - 24
+     - ``cf02`` = ``schwefel_func`` + ``rastrigin_func`` + ``hgbat_func``
+   * - 25
+     - ``cf03`` = ``schwefel_func`` + ``rastrigin_func`` + ``ellips_func``
+   * - 26
+     - ``cf04`` = ``schwefel_func`` + ``happycat_func`` + ``ellips_func`` + ``weierstrass_func`` + ``griewank_func``
+   * - 27
[…]
