+++ minion/algorithms/minion_cec.h
@@ -4,6 +4,7 @@
 #include "cec.h"
 #include "cec2011.h"
 #include "cec2014.h"
+#include "cec_combine.h"
 #include "cec2017.h"
 #include "cec2019.h"
 #include "cec2020.h"
