+++ minionpy/__init__.py
@@ -5,6 +5,7 @@ from .benchmark import *
 from .bbob import *
 from .parallel import *
 from .cec import CEC2011Functions as CEC2011
+from .cec import CEC20142017Functions as CEC20142017
 
 # Define the package's version
 __version__ = "1.7.0"
