+++ minion/benchmark/compositionf.cpp
@@ -1,0 +1,384 @@
+#include "compositionf.h"
+
+#include <algorithm>
+#include <cmath>
+#include <limits>
+#include <numeric>
+#include <sstream>
+#include <stdexcept>
+#include <utility>
+
+namespace minion {
+namespace {
+
+constexpr double kWeightFloor = 1.0e-300;
+
+bool is_common_style(int style) {
+    return style >= 1 && style <= 10;
+}
+
+bool is_shifted_style(int style) {
+    return style >= 11 && style <= 20;
+}
+
+std::vector<double> make_default_deltas(std::size_t count) {
+    std::vector<double> deltas(count, 10.0);
+    for (std::size_t i = 0; i < count; ++i) {
+        deltas[i] = 10.0 * static_cast<double>(i + 1);
+    }
+    return deltas;
+}
+
+std::vector<double> make_default_biases(std::size_t count) {
+    return std::vector<double>(count, 0.0);
+}
+
+std::vector<double> make_default_scales(std::size_t count) {
+    return std::vector<double>(count, 1.0);
+}
+
+std::vector<std::vector<double>> make_default_rotations(std::size_t count) {
+    return std::vector<std::vector<double>>(count);
+}
+
+std::vector<std::vector<double>> make_default_centers(
+    std::size_t count,
+    int dimension,
+    const std::vector<std::shared_ptr<FunctionBase>>& components,
+    int style) {
+    std::vector<std::vector<double>> centers(count, std::vector<double>(static_cast<std::size_t>(dimension), 0.0));
+    if (is_common_style(style)) {
+        if (!components.empty()) {
+            centers[0] = components[0]->x_opt;
+            for (std::size_t i = 1; i < count; ++i) {
+                centers[i] = centers[0];
+            }
+        }
+        return centers;
+    }
+    if (!components.empty()) {
+        centers[0] = components[0]->x_opt;
+    }
+    return centers;
+}
+
+void apply_rotation(
+    const std::vector<double>& input,
+    const std::vector<double>& rotation,
+    int dimension,
+    std::vector<double>& output) {
+    output.assign(static_cast<std::size_t>(dimension), 0.0);
+    for (int row = 0; row < dimension; ++row) {
+        double sum = 0.0;
+        const std::size_t row_offset = static_cast<std::size_t>(row) * static_cast<std::size_t>(dimension);
+        for (int col = 0; col < dimension; ++col) {
+            sum += rotation[row_offset + static_cast<std::size_t>(col)] * input[static_cast<std::size_t>(col)];
+        }
+        output[static_cast<std::size_t>(row)] = sum;
+    }
+}
+
+double squared_distance(const double* x, const std::vector<double>& center, int dimension) {
+    double sum = 0.0;
+    for (int i = 0; i < dimension; ++i) {
+        const double diff = x[i] - center[static_cast<std::size_t>(i)];
+        sum += diff * diff;
+    }
+    return sum;
+}
+
+std::string component_list_string(const std::vector<std::shared_ptr<FunctionBase>>& components) {
+    std::ostringstream oss;
+    for (std::size_t i = 0; i < components.size(); ++i) {
+        if (i != 0) {
+            oss << ", ";
+        }
+        oss << components[i]->name;
+    }
+    return oss.str();
+}
+
+bool same_optimum(
+    const std::vector<double>& lhs,
+    const std::vector<double>& rhs) {
+    if (lhs.size() != rhs.size()) {
+        return false;
+    }
+    for (std::size_t i = 0; i < lhs.size(); ++i) {
+        if (std::abs(lhs[i] - rhs[i]) > 1.0e-12) {
+            return false;
+        }
+    }
+    return true;
+}
+
[…]
