+++ docs/benchmark/minion_benchmark.tex
@@ -5,15 +5,27 @@
 \usepackage{lmodern}
 \usepackage{amsmath,amssymb,amsfonts,amsthm}
 \usepackage{geometry}
+\usepackage{graphicx}
 \usepackage{hyperref}
 \usepackage{enumitem}
+
 
 \geometry{margin=1in}
 \setlist[itemize]{leftmargin=1.5em}
 
-\newtheorem{proposition}{Proposition}
+% --- Add these lines here ---
+\theoremstyle{remark}
+\newtheorem{remark}{Remark}
+% ----------------------------
 
-\title{Minion Benchmark Functions}
+
+% --- Theorems and Propositions (Italicized text style) ---
+\theoremstyle{plain}
+\newtheorem{theorem}{Theorem}
+\newtheorem{proposition}{Proposition}
+\newtheorem{corollary}{Corollary}
+
+\title{General Framework for Scalable Construction of Optimization Benchmark Function}
 \author{Minion Project}
 \date{\today}
 
@@ -23,171 +35,865 @@
 
 \tableofcontents
 
-\section{Construction of Benchmark Functions}
-\begin{proposition}[Common minimizer under weighted composition]
-	Let
-	\[
-	f(x) = \sum_{i=1}^m w_i(x) g_i(x),
-	\]
-	and suppose that $x_g$ is a common global minimizer of all component functions
-	$g_i$. Define
-	\[
-	\Delta g_i(x) = g_i(x) - g_i(x_g),
-	\qquad
-	\Delta w_i(x) = w_i(x) - w_i(x_g).
-	\]
-	Then $x_g$ is a global minimizer of $f$ if and only if
-	\[
-	\sum_i w_i(x)\,\Delta g_i(x)
-	+
-	\sum_i \Delta w_i(x)\,g_i(x_g)
-	\ge 0
-	\qquad
-	\forall x.
-	\]
-\end{proposition}
 
-\begin{proof}
-	Since $x_g$ is a global minimizer of $f$ if and only if $f(x)-f(x_g)\ge 0$ for
-	all $x$, we expand
-	\[
-	f(x)-f(x_g)
-	=
-	\sum_i w_i(x)g_i(x) - \sum_i w_i(x_g)g_i(x_g).
-	\]
-	Add and subtract $\sum_i w_i(x)g_i(x_g)$:
-	\[
-	f(x)-f(x_g)
-	=
-	\sum_i w_i(x)\bigl[g_i(x)-g_i(x_g)\bigr]
-	+
-	\sum_i \bigl[w_i(x)-w_i(x_g)\bigr] g_i(x_g),
-	\]
-	which is exactly
-	\[
-	\sum_i w_i(x)\,\Delta g_i(x)
-	+
-	\sum_i \Delta w_i(x)\,g_i(x_g).
-	\]
-	Hence $x_g$ is a global minimizer if and only if the displayed inequality holds
-	for all $x$.
-\end{proof}
+\section{Introduction}
+\label{sec:introduction}
 
-\begin{proposition}[Exact-dominance sufficient condition]
-Consider
-\[
-f(x)=\sum_{i=1}^m w_i(x)\bigl(b_i+g_i(x-x_i^\ast)\bigr),
-\]
-where each component satisfies
-\[
-g_i(z)\ge 0
-\qquad \text{for all } z,
-\qquad
-g_i(0)=0.
-\]
-Assume further that
-\[
-w_i(x)\ge 0
-\qquad \forall i,x,
-\qquad
-\sum_{i=1}^m w_i(x)=1
-\qquad \forall x,
-\]
-that component $1$ has the smallest bias,
-\[
-b_1 \le b_i
[…]
