### 2026-07-14 d4b37e8
Updated paper
