+++ docs/benchmark/scripts/plot_deceptive_weighted_composition_3d.py
@@ -1,0 +1,143 @@
+import os
+
+import matplotlib.pyplot as plt
+import numpy as np
+
+
+def rastrigin(x, y):
+    return 20.0 + x**2 + y**2 - 10.0 * (np.cos(2.0 * np.pi * x) + np.cos(2.0 * np.pi * y))
+
+
+def ackley(x, y):
+    r = np.sqrt(0.5 * (x**2 + y**2))
+    c = 0.5 * (np.cos(2.0 * np.pi * x) + np.cos(2.0 * np.pi * y))
+    return -20.0 * np.exp(-0.2 * r) - np.exp(c) + np.e + 20.0
+
+
+def griewank(x, y):
+    return 1.0 + (x**2 + y**2) / 4000.0 - np.cos(x) * np.cos(y / np.sqrt(2.0))
+
+
+def steep_ellipsoid(x, y):
+    return x**2 + 70.0 * y**2
+
+
+def rosenbrock_at_zero(x, y):
+    u = x + 1.0
+    v = y + 1.0
+    return 100.0 * (v - u**2) ** 2 + (1.0 - u) ** 2
+
+
+def sphere(x, y):
+    return x**2 + y**2
+
+
+def normalize(values):
+    return values / np.percentile(values, 95.0)
+
+
+def deceptive_composition(xx, yy, global_fun, deceptive_fun, global_scale, deceptive_scale, offset):
+    x_global = np.array([-45.0, -35.0])
+    x_deceptive = np.array([45.0, 35.0])
+
+    gx = (xx - x_global[0]) / global_scale
+    gy = (yy - x_global[1]) / global_scale
+    dx = (xx - x_deceptive[0]) / deceptive_scale
+    dy = (yy - x_deceptive[1]) / deceptive_scale
+
+    q_global = normalize(global_fun(gx, gy))
+    q_deceptive = normalize(deceptive_fun(dx, dy)) + offset
+
+    d_global = (xx - x_global[0]) ** 2 + (yy - x_global[1]) ** 2
+    d_deceptive = (xx - x_deceptive[0]) ** 2 + (yy - x_deceptive[1]) ** 2
+    sigma2 = 42.0**2
+
+    raw_global = np.exp(-d_global / (2.0 * sigma2))
+    raw_deceptive = np.exp(-d_deceptive / (2.0 * sigma2))
+    w_global = raw_global / (raw_global + raw_deceptive)
+    w_deceptive = raw_deceptive / (raw_global + raw_deceptive)
+
+    local_radius2 = 24.0**2
+    global_core = d_global <= local_radius2
+    deceptive_core = d_deceptive <= local_radius2
+    w_global[global_core] = 1.0
+    w_deceptive[global_core] = 0.0
+    w_global[deceptive_core] = 0.0
+    w_deceptive[deceptive_core] = 1.0
+
+    f = w_global * q_global + w_deceptive * q_deceptive
+    return f, x_global, x_deceptive, offset
+
+
+def main():
+    x = np.linspace(-100.0, 100.0, 161)
+    y = np.linspace(-100.0, 100.0, 161)
+    xx, yy = np.meshgrid(x, y)
+
+    examples = [
+        ("Rastrigin + steep Ellipsoid", rastrigin, steep_ellipsoid, 12.0, 18.0, 0.16),
+        ("Ackley + Rosenbrock", ackley, rosenbrock_at_zero, 13.0, 5.8, 0.14),
+        ("Griewank + Sphere", griewank, sphere, 7.0, 20.0, 0.12),
+    ]
+
+    panels = []
+    for title, global_fun, deceptive_fun, global_scale, deceptive_scale, offset in examples:
+        f, x_global, x_deceptive, deceptive_offset = deceptive_composition(
+            xx,
[…]
