+++ docs/benchmark/minion_benchmark.tex
@@ -28,6 +28,9 @@
 \title{General Framework for Scalable Construction of Optimization Benchmark Function}
 \author{
 	Khoirul Faiq Muzakka\(^{\ast\dagger}\),
+	Others*, 
+	Others*,
+	Others*,
 	Martin Finsterbusch\(^{\ast}\)\\
 	\small \(^{\ast}\)Institute of Energy Materials and Devices (IMD-2)\\
 	\small Forschungszentrum J\"ulich GmbH, Germany\\
@@ -40,26 +43,15 @@
 \maketitle
 
 \begin{abstract}
-Benchmark suites such as BBOB/COCO and the CEC collections provide standard
-test problems for comparing optimization algorithms, but their fixed size and
-fixed landscape designs limit their use for large-scale robustness testing. This
-paper proposes a general framework for constructing scalable families of
+This paper proposes a general framework for constructing scalable families of
 continuous optimization benchmark functions from known base functions. The
-construction combines coordinate transformations, value transformations, and
-composition operators while preserving analytically known global optima under
-explicit sufficient conditions. The framework separates the roles of the main
-operators: coordinate transformations control the embedding and geometry of
-component landscapes, value transformations control objective-value response and
-modality, and composition operators determine how component features interact in
-the final objective. We provide prescriptions for common-optimum constructions,
-multiple-optimum constructions, deceptive constructions, and nested
-compositions, together with a discussion of how operator choices affect local
-minima, curvature, flatness, and multimodality. The framework is intended as a
-programmable extension of existing benchmark suites for generating many
-reproducible problem instances with controlled landscape properties and known
-reference optima. A C++ and Python implementation is provided to support
-practical benchmark generation and future large-scale algorithm robustness
-studies.
+framework gives explicit construction prescriptions together with sufficient
+conditions that preserve analytically known global optima. It is intended as a
+programmable extension of existing benchmark suites, enabling many reproducible
+problem instances with controlled landscape properties and known reference
+optima. A C++ and Python implementation, FuncCraft, is provided to support
+practical benchmark generation and future large-scale studies of algorithm
+robustness.
 \end{abstract}
 
 \setcounter{tocdepth}{1}
@@ -98,13 +90,14 @@ conditioning, separability, variable interaction, basin geometry, and deception.
 In this work, we propose a general framework for the scalable construction of
 optimization benchmark functions. The central idea is to build new functions by
 combining transformed base functions through coordinate transformations, value
-transformations, and composition operators. The generated functions take the
+transformations, and composition functions. The generated functions take the
 general form
 \[
 f(x)
 =
 \psi
 \left(
+x,
 \phi_1\!\left(g_1(T_1(x))\right),
 \ldots,
 \phi_m\!\left(g_m(T_m(x))\right)
@@ -112,11 +105,12 @@ f(x)
 \]
 where \(g_i\) are corrected base functions with known optima, \(T_i\) are
 coordinate transformations, \(\phi_i\) are value transformations, and \(\psi\)
[…]
