+++ docs/benchmark/minion_benchmark.tex
@@ -25,7 +25,7 @@
 \newtheorem{proposition}{Proposition}
 \newtheorem{corollary}{Corollary}
 
-\title{General Framework for Scalable Construction of Optimization Benchmark Function}
+\title{General Framework for Scalable Construction of Optimization Benchmark Functions}
 \author{
 	Khoirul Faiq Muzakka\(^{\ast\dagger}\),
 	Others*, 
@@ -49,7 +49,7 @@ framework gives explicit construction prescriptions together with sufficient
 conditions that preserve analytically known global optima. It is intended as a
 programmable extension of existing benchmark suites, enabling many reproducible
 problem instances with controlled landscape properties and known reference
-optima. A C++ and Python implementation, FuncCraft, is provided to support
+optima. A C++ and Python implementation, \textsc{FuncCraft}, is provided to support
 practical benchmark generation and future large-scale studies of algorithm
 robustness.
 \end{abstract}
@@ -62,86 +62,110 @@ robustness.
 \label{sec:introduction}
 
 Benchmark functions play a central role in the development and evaluation of
-optimization algorithms. They provide controlled test problems for studying
-convergence behavior, robustness, scalability, sensitivity to landscape
-properties, and comparative performance across algorithms. Widely used benchmark
-suites, such as the CEC benchmark sets and the BBOB testbed, have become
-standard tools for evaluating numerical optimization methods.
+optimization algorithms. They provide controlled test problems for analyzing
+convergence behavior, robustness, scalability, exploration--exploitation
+trade-offs, and algorithmic performance under different landscape properties.
+Consequently, benchmark suites such as the CEC competitions and the BBOB
+testbed have become standard evaluation platforms for numerical optimization.
 
-Despite their importance, existing benchmark suites are typically limited in
-size. Many commonly used collections contain only a few dozen functions, often
-with a fixed set of transformations, dimensions, and landscape structures. This
-is useful for standardized comparison, but it also limits the diversity of
-problems on which an algorithm can be tested. In particular, robustness studies
-require algorithms to be evaluated against many distinct problem instances, not
-only against a small fixed collection. An algorithm that performs well on a
-limited benchmark suite may still fail on related landscapes that are not
-represented in the test set.
+Despite their widespread use, existing benchmark suites share a common
+limitation: they contain only a relatively small number of manually designed
+benchmark functions. Typical benchmark collections consist of only a few dozen
+problems with fixed landscape characteristics. While such benchmark suites are
+indispensable for standardized comparison, they inevitably sample only a tiny
+fraction of the enormous space of possible optimization landscapes. As a
+result, an optimization algorithm may perform well on a particular benchmark
+suite while exhibiting substantially different behavior on related, but
+unrepresented, landscape structures.
 
-This motivates the need for a systematic framework for generating large numbers
-of benchmark functions in a controlled and reproducible way. Instead of defining
-each benchmark function manually, one would like to construct new functions
-automatically from known base functions while preserving important analytical
-information, especially the location and value of the global optimum. Such a
-framework should allow the user to generate hundreds or thousands of distinct
-optimization problems, while still controlling properties such as modality,
-conditioning, separability, variable interaction, basin geometry, and deception.
+This limitation becomes increasingly important as optimization algorithms become
+more sophisticated. Modern optimization methods are often designed to be robust
+across a wide variety of problem classes rather than being tailored to a small
[…]
