+++ minion/algorithms/l_bfgs_b.h
@@ -60,7 +60,7 @@ public:
      * initialization. If omitted, a random point sampled from the bounds is
      * used.
      * @param data Additional data for the objective function.
-     * @param callback Callback function for intermediate results.
+     * @param callback Callback invoked with intermediate results. Return true to stop optimization; return false to continue.
      * @param maxevals The maximum number of evaluations.
      * @param seed The seed for random number generation.
      * @param options Option map that specifies further configurational settings for the algorithm.
@@ -70,7 +70,7 @@ public:
         const std::vector<std::pair<double, double>>& bounds_,
         const std::vector<std::vector<double>>& x0 = {},
         void* data = nullptr,
-        std::function<void(MinionResult*)> callback = nullptr,
+        std::function<bool(MinionResult*)> callback = nullptr,
         size_t maxevals = 100000,
         int seed = -1,
         std::map<std::string, ConfigValue> options = std::map<std::string, ConfigValue>()
