+++ minion/algorithms/spso2011.h
@@ -20,7 +20,7 @@ public:
      * @param bounds Search-space bounds.
      * @param x0 Optional initial swarm positions.
      * @param data User payload for the objective.
-     * @param callback Progress callback invoked with the best-so-far state.
+     * @param callback Callback invoked with intermediate results. Return true to stop optimization; return false to continue.
      * @param maxevals Maximum number of objective evaluations.
      * @param seed RNG seed (negative -> random seed).
      * @param options Configuration map (phi values, neighbourhood size, etc.).
@@ -30,7 +30,7 @@ public:
         const std::vector<std::pair<double, double>>& bounds,
         const std::vector<std::vector<double>>& x0 = {},
         void* data = nullptr,
-        std::function<void(MinionResult*)> callback = nullptr,
+        std::function<bool(MinionResult*)> callback = nullptr,
         size_t maxevals = 100000,
         int seed = -1,
         std::map<std::string, ConfigValue> options = std::map<std::string, ConfigValue>()
