+++ docs/source/minionpy/examples.rst
@@ -6,4 +6,4 @@ Below is an example of a Jupyter notebook showcasing the use of `minionpy`:
 .. toctree::
    :maxdepth: 2
 
-   minimizer
+   minimizer_notebook
