+++ docs/source/l_bfgs_b_notebook.rst
@@ -1,0 +1,6 @@
+L-BFGS-B Comparison Notebook
+============================
+
+The comparison notebook is available in the documentation source tree:
+
+:download:`l_bfgs_b.ipynb <l_bfgs_b.ipynb>`
