+++ docs/source/minionpy/minimizer_notebook.rst
@@ -1,0 +1,6 @@
+MinionPy Minimizer Notebook
+===========================
+
+The minimizer example notebook is available in the documentation source tree:
+
+:download:`minimizer.ipynb <minimizer.ipynb>`
