### 2026-07-28 0ed9d63
Removed GWO_DE
