+++ minion/algorithms/gwo_de.cpp
@@ -1,247 +1,0 @@
-#include "gwo_de.h"
-#include <algorithm>
-#include <numeric>
-#include <random>
-#include <cmath>
-#include "default_options.h"
-
-namespace minion {
-
-void GWO_DE::initialize  (){
-    auto defaultKey = DefaultSettings().getDefaultSettings("GWO_DE");
-    for (auto el : optionMap) defaultKey[el.first] = el.second;
-    Options options(defaultKey);
-
-    boundStrategy = options.get<std::string> ("bound_strategy", "reflect-random");
-    std::vector<std::string> all_boundStrategy = {"random", "reflect", "reflect-random", "clip", "periodic", "none"};
-    if (std::find(all_boundStrategy.begin(), all_boundStrategy.end(), boundStrategy)== all_boundStrategy.end()) {
-        std::cerr << "Bound stategy '"+ boundStrategy+"' is not recognized. 'Reflect-random' will be used.\n";
-        boundStrategy = "reflect-random";
-    }
-
-    size_t population_size = options.get<int> ("population_size", 0); 
-    if (population_size == 0) {
-        population_size = std::max<size_t>(5 * bounds.size(), 20);
-    }
-    if (population_size < 4) {
-        population_size = 4;
-    }
-
-    dimension = bounds.size();
-    alpha_score = std::numeric_limits<double>::infinity();
-    beta_score = std::numeric_limits<double>::infinity();
-    delta_score = std::numeric_limits<double>::infinity();
-    eval_count = 0;
-    alpha_pos = std::vector<double> (dimension, 0.0);
-    beta_pos  = std::vector<double> (dimension, 0.0);
-    delta_pos = std::vector<double> (dimension, 0.0);
-    population = std::vector<std::vector<double>>(population_size, std::vector<double>(dimension, 0.0));
-    fitness = std::vector<double> (population_size, std::numeric_limits<double>::infinity());
-    CR = options.get<double>("crossover_rate", 0.7); 
-    F= options.get<double>("mutation_rate", 0.5); 
-    elimination_prob = options.get<double>("elimination_prob", 0.7); 
-    initialize_population();
-    evaluate_population();
-    update_leaders();
-    hasInitialized=true;
-}
-
-MinionResult GWO_DE::optimize() {
-    try {
-        resetBestSoFar();
-        if (!hasInitialized) initialize();
-        size_t iter = 0;
-
-        while (eval_count < maxevals) {
-            double progress = maxevals > 0 ? static_cast<double>(eval_count) / static_cast<double>(maxevals) : 0.0;
-            progress = std::clamp(progress, 0.0, 1.0);
-            double a = 2.0 * (1.0 - progress);
-
-            if (eval_count + population.size() > maxevals) {
-                break;
-            }
-            auto gwo_candidates = generate_gwo_candidates(a);
-            enforce_bounds(gwo_candidates, bounds, boundStrategy);
-            auto gwo_fitness = func(gwo_candidates, data);
-            eval_count += gwo_candidates.size();
-
-            if (eval_count >= maxevals) {
-                break;
-            }
-            auto de_candidates = differential_evolution();
-            enforce_bounds(de_candidates, bounds, boundStrategy);
-            auto de_fitness = func(de_candidates, data);
-            eval_count += de_candidates.size();
-
-            auto totalCandidates = population.size() + gwo_candidates.size() + de_candidates.size();
-            std::vector<std::vector<double>> combined;
-            std::vector<double> combinedFitness;
-            combined.reserve(totalCandidates);
-            combinedFitness.reserve(totalCandidates);
-
[…]
