+++ minionpy/__init__.py
@@ -8,7 +8,7 @@ from .cec import CEC2011Functions as CEC2011
 from .cec import CEC20142017Functions as CEC20142017
 
 # Define the package's version
-__version__ = "1.7.0"
+__version__ = "1.8.0"
 
 # Optionally, define the package name
 __name__ = "minionpy"
