+++ tests/test_minion.cpp
@@ -1,9 +1,11 @@
+#include <cmath>
 #include <iomanip>
 #include <iostream>
-#include <cmath>
 #include <limits>
 #include <map>
+#include <sstream>
 #include <string>
+#include <utility>
 #include <vector>
 
 #include <minion.h>
@@ -11,12 +13,29 @@
 
 namespace {
 
+struct CheckRunner {
+    using Check = std::pair<std::string, bool (*)()>;
+
+    int run(const std::vector<Check>& checks) const {
+        const size_t total = checks.size();
+        for (size_t i = 0; i < total; ++i) {
+            const auto& [name, check] = checks[i];
+            if (!check()) {
+                std::cerr << "Test " << (i + 1) << "/" << total << ": " << name << ": failed\n";
+                return 1;
+            }
+            std::cout << "Test " << (i + 1) << "/" << total << ": " << name << ": passed\n";
+        }
+        std::cout << total << "/" << total << " Passed.\n";
+        return 0;
+    }
+};
+
 std::vector<double> sphere_batch(const std::vector<std::vector<double>>& X, void*) {
     std::vector<double> out(X.size(), 0.0);
     for (size_t i = 0; i < X.size(); ++i) {
-        const auto& x = X[i];
         double f = 0.0;
-        for (double value : x) {
+        for (double value : X[i]) {
             f += value * value;
         }
         out[i] = f;
@@ -27,11 +46,10 @@ std::vector<double> sphere_batch(const std::vector<std::vector<double>>& X, void
 std::vector<double> rosenbrock_batch(const std::vector<std::vector<double>>& X, void*) {
     std::vector<double> out(X.size(), 0.0);
     for (size_t i = 0; i < X.size(); ++i) {
-        const auto& x = X[i];
         double f = 0.0;
-        for (size_t j = 0; j + 1 < x.size(); ++j) {
-            const double a = x[j + 1] - x[j] * x[j];
-            const double b = 1.0 - x[j];
+        for (size_t j = 0; j + 1 < X[i].size(); ++j) {
+            const double a = X[i][j + 1] - X[i][j] * X[i][j];
+            const double b = 1.0 - X[i][j];
             f += 100.0 * a * a + b * b;
         }
         out[i] = f;
@@ -44,164 +62,190 @@ std::vector<double> cec2017_batch(const std::vector<std::vector<double>>& X, voi
     return (*cec)(X);
 }
 
+const std::vector<std::string>& algorithms() {
+    static const std::vector<std::string> values = {
+        "DE",
+        "LSHADE",
+        "AGSK",
+        "JADE",
+        "j2020",
+        "NLSHADE_RSP",
+        "NLSHADE_LBC",
+        "LSRTDE",
+        "RDEX",
+        "jSO",
+        "IMODE",
+        "ARRDE",
+        "NelderMead",
+        "ABC",
+        "PSO",
+        "SPSO2011",
+        "DMSPSO",
+        "LSHADE_cnEpSin",
+        "CMAES",
+        "ACMAES",
+        "RCMAES",
+        "BIPOP_aCMAES",
+        "DA",
+        "L_BFGS_B",
+        "L_BFGS",
+    };
+    return values;
+}
+
+std::map<std::string, minion::ConfigValue> settings_for(const std::string& algo) {
+    auto settings = minion::DefaultSettings().getDefaultSettings(algo);
+    settings["convergence_tol"] = 1e-8;
+    if (algo == "ARRDE") {
+        settings["minimum_population_size"] = 4;
+    }
+    return settings;
+}
+
+bool is_valid_result(const minion::MinionResult& result, size_t maxevals, size_t slack) {
+    return std::isfinite(result.fun) &&
+           result.nfev <= maxevals + slack &&
[…]
