### 2026-08-02 677feae
Implemented more stopping mechanism based on x_tol, f_tol, and maxiters
