+++ minion/algorithms/lshade.cpp
@@ -39,8 +39,9 @@ void LSHADE::initialize  (){
         reduction_strategy="linear";
     }
 
-        minPopSize = options.get<int>("minimum_population_size", 4);
-    if (populationSize == minPopSize) popreduce = false; 
+    minPopSize = options.get<int>("minimum_population_size", 4);
+    configureConvergenceTolerances(options);
+    if (populationSize == minPopSize) popreduce = false;
     else popreduce = true;
     hasInitialized=true;
 }
