### 2026-08-02 3d71eaa
Removed MinionResults succeeded() and Updated docs
