+++ examples/main_run_custom.cpp
@@ -137,8 +137,7 @@ int main(int argc, char* argv[]) {
                       << std::setw(18) << benchmark->function()->name << "  "
                       << "f=" << std::setw(14) << std::setprecision(8) << std::scientific << result.fun << "  "
                       << "nfev=" << std::setw(8) << result.nfev << "  "
-                      << "status=" << result.status << "  "
-                      << "ok=" << (result.succeeded() ? "1" : "0") << std::endl;
+                      << "status=" << result.status << std::endl;
         }
 
         const double mean_fun = total_fun / static_cast<double>(suite.size());
