+++ examples/L_bfgs.py
@@ -218,7 +218,8 @@ def test_optimization_noise(test_func, bounds, dimension, func_name, maxevals, s
             "population_size": 0,
             "N_points_derivative": n_points,
             "func_noise_ratio": noise_ratio,
-            "convergence_tol" : 0.0
+            "x_tol" : 0.0,
+            "f_tol" : 0.0
         }
         if extra_options:
             options.update(extra_options)
