### 2026-08-02 #10145 @anugrahjo
> Hi [@muzakka](https://github.com/muzakka) [@monacofj](https://github.com/monacofj) [@anugrahjo](https://github.com/anugrahjo) Going ahead, I will be helping edit this submission.
>
> [@anugrahjo](https://github.com/anugrahjo) [@muzakka](https://github.com/muzakka) My 2 cents - Ideally I would love to see PyPi wheel packages working everywhere including Intel-based macOS systems, however given Intel-based macOS systems are discontinued since 2020 and Apple Silicon is primary architecture now, I think build from source should be acceptable for Intel-based macOS systems. I see this as tradeoff for research software builders - maintaining binaries/wheels/etc for all hardware and OS combination is hard and we should be practical about it.

@abhishektiwari I agree, provided this limitation is clearly documented. My main concern was that wheels were being uploaded without any testing. Since this issue has now been addressed, I've verified the changes and checked this item off my review checklist.
