### 2026-08-20 #10145 @monacofj
Hello

I must apologize once more (personal issues messing with my agenda, but
that can't count as an excuse).
Today I'm taking the time to complete the missing bits of my review.
I'm really sorry for the delay.

[]s monaco

On Thu, Aug 6, 2026 at 8:58 PM Abhishek Tiwari ***@***.***>
wrote:

> *abhishektiwari* left a comment (openjournals/joss-reviews#10145)
> <https://github.com/openjournals/joss-reviews/issues/10145#issuecomment-5210110155>
>
> @monacofj <https://github.com/monacofj> - Hope all is well. How are you
> progressing with review of this submission? Let me know if any questions or
> concerns.
>
