### 2026-08-20 #10145 @monacofj
Hello @abhishektiwari and everyone else,

I have completed my review, and I think the submission has improved substantially.

In my initial review, I left some checklist items unmarked because they had not yet been clearly addressed. I see that the authors have since worked on those items, and I consider all of them well addressed now.

The issue I opened in the software repository  concerning the installation process (khoirulmuzakka/Minion/issues/2) was successfully resolved by the authors. I was able to clone the repository and go through the installation process again, and everything worked like a charm.

I only had a couple of minor observations during this final pass.

One concerns the documentation. The README says that "Detailed installation, native package, C++ build, and usage instructions are available in the official documentation." Those instructions are indeed available under `docs/source/installation.rst` and in the published documentation, although a direct link from that sentence in the README might make them easier to find. This is a very minor point, though.

Second, regarding **Collaborative effort**, the Git history appears to show a single code contributor. However, the project has also received contributions from external users through issue reports and discussions, some of which have led to fixes or improvements. Thus, although community engagement is still relatively limited, it does exist. Given the niche nature of this kind of research software, I do not consider this an atypical situation, and I marked this item as complete.

Finally, regarding **Contribution and authorship**, the paper has three authors, while the Git history appears to show only one code contributor. However, authorship may legitimately reflect intellectual contributions beyond code commits—such as conceptual development, scientific oversight, validation, or manuscript preparation—which are not necessarily visible in the repository history. I found no reason to consider the author list inappropriate or incomplete, so I marked this item as complete.

In summary, from my perspective, all checklist items are now satisfied. Review completed.

Great job, @khoirulmuzakka !
