### 2026-08-24 86a09a7
Implemented Gradient descent
