+++ minion/algorithms/l_bfgs_b.cpp
@@ -787,67 +787,26 @@ double L_BFGS_B::fun_and_grad(const VectorXd& x, VectorXd& grad){
     if (Nevals > maxevals) throw MaxevalExceedError("Maxevals has been exceeded.");
     if (!state) throw std::runtime_error("L_BFGS_B internal state is not initialized.");
 
-    const int m = std::min(std::max(static_cast<int>(std::ceil((static_cast<double>(N_points) - 1.0) / 2.0)), 1), 8);
     std::vector<double> x_vec(x.data(), x.data() + x.size());
-    std::vector<std::vector<double>> X;
-    X.push_back(x_vec);
+    BlackBoxGradientOptions gradient_options;
+    gradient_options.N_points = N_points;
+    gradient_options.func_noise_ratio = func_noise_ratio;
+    gradient_options.last_f = last_f;
+    gradient_options.fd_epsilon = fd_epsilon;
+    gradient_options.finite_diff_rel_step = fin_diff_rel_step;
+    gradient_options.estimator = "coordinate_fd";
+    gradient_options.bounded = true;
+    gradient_options.bounds = actual_bounds;
+    gradient_options.curvature_diag = state->core.bfgs.compute_hessian_diagonal();
 
-    std::vector<double> hvec;
-    std::vector<double> sec_der = state->core.bfgs.compute_hessian_diagonal();
-    const double ferr = std::max(std::fabs(last_f), 1.0) * func_noise_ratio;
+    const BlackBoxGradientResult result = estimateBlackBoxGradient(func, data, x_vec, gradient_options);
+    Nevals += result.evaluations;
+    last_f = result.value;
 
-    for (int i = 0; i < x.size(); i++) {
-        double d_low = x[i] - actual_bounds[i].first;
-        double d_high = actual_bounds[i].second - x[i];
-        double h_low_corrected = std::max(0.0, d_low / m - 1e-16);
-        double h_high_corrected = std::max(0.0, d_high / m - 1e-16);
-        double h_max_from_bound = std::min(h_low_corrected, h_high_corrected);
-        double h_min = std::pow(epsilon, 0.5) * std::max(1.0, std::fabs(x[i]));
-        double h_max = 0.01 * std::max(1.0, std::fabs(x[i]));
-        double curvature = std::max(std::fabs(sec_der[static_cast<size_t>(i)]), 1e-16);
-        double h_est = 2.0 * std::sqrt(std::max(ferr, 1e-32) / curvature);
-        double h = std::min(h_max, std::max(h_min, h_est));
-        h = std::min(h, h_max_from_bound);
-        if (h <= 0.0) {
-            h = std::min(h_max, std::max(h_min, fin_diff_rel_step * std::max(1.0, std::fabs(x[i]))));
-        }
-        hvec.push_back(h);
-    }
-
-    if (N_points == 1) {
-        for (int i = 0; i < x.size(); i++) {
-            std::vector<double> xp = x_vec;
-            double h = hvec[static_cast<size_t>(i)];
-            if (xp[i] + h > bounds[i].second) h = -h;
-            xp[i] += h;
-            X.push_back(xp);
-        }
-    } else {
-        for (int i = 0; i < x.size(); i++) {
-            for (int j = 1; j <= m; j++) {
-                double h = hvec[static_cast<size_t>(i)];
-                std::vector<double> xpp = x_vec;
-                std::vector<double> xpm = x_vec;
-                xpp[i] += j * h;
-                xpm[i] -= j * h;
-                X.push_back(xpp);
-                X.push_back(xpm);
-            }
-        }
-    }
-
-    auto F = func(X, data);
-    Nevals += F.size();
[…]
