### 2026-08-27 0109f59
Add Gradient-descent to optimization algorithms list
