### 2026-08-27 #10145 @khoirulmuzakka
Many thanks @monacofj !
@abhishektiwari I have finished the checklist. Please let me know if this is good enough. Thanks a lot!

- [ +] Double check Paper authors and affiliations (including ORCIDs)
- [ +] Make a tagged release of your software on Github, and list the version tag of the archived version here.
- [ +] Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
- [ +] Check the archival deposit (e.g., in Zenodo) has the correct title (should match the paper title).
- [ +] Check the archival deposit (e.g., in Zenodo) has the correct authors and affiliations (should match the paper authors including order).
- [ +] Check the archival deposit (e.g., in Zenodo) has the correct version (should match the above Github tag you created).
- [ +] Check the archival deposit (e.g., in Zenodo) has the correct ORCIDs (should match the paper authors).
- [ +] Make sure that the license listed for the archive is the same as the software license.
- [ +] Please list the DOI of the archived version here.

The released tag version is v1.9.1.
Zenodo link : https://zenodo.org/records/22134792
DOI : https://doi.org/10.5281/zenodo.22134792
