### 2026-09-03 #10145 @abhishektiwari
@muzakka Apologies for delay. Please see following corrections in paper.

## Main Body (Edits in paper.md):

- Line 65: Add a period after "libraries"
- Line 10, 12, 31, 169: please check "black box" vs "blackbox for consistency
- Lines 96–97: "exposes both optimize() and call-operator execution" breaks parallelism — "both" joins a function name to a noun phrase.should it be: "exposes execution through both optimize() and the call operator."
- Line 72, 136: "L-SHADE" vs "LSHADE"
- Line 141: "LSRTDE" should be "L-SRTDE"
- JOSS convention is sentence case throughout. Check heading some of them title case ("Software Design" and "Research Impact Statemen) others sentence case (State of the field", "Statement of need", "Availability and documentation", and "AI usage disclosure").

## References (Edits in paper.bib): Please use {} explicitly to capitalise/uncapitalize as by default JOSS formats reference titles as sentence case.

- Line 276 ( Stanovov & Semenkin (2024),): "l-SRTDE" should be  "L-SRTDE".
- Line 269 (Qiu (2013)): "l-BFGS and l-BFGS-b" should be "L-BFGS and L-BFGS-B"
- Lines 201–202: "with euclidean neighborhood" should be "Euclidean neighborhood"
- Lines 292–293: "application to the thomson model" should be "Thomson model"
- Lines 287–288: "scientific computing in python" should be "Python".
- Line 231 (Awad et al., 2016 / CEC 2017): malformed author names. Please check if it should be "Awad, N. H., Ali, M. Z., Suganthan, P. N., Liang, J. J., & Qu, B. Y."
- Line 235 (Liang et al., 2013 / CEC 2014): malformed author names. Please check if it should be "Liang, J. J., Qu, B. Y., & Suganthan, P. N."
- Line 289 (V. Price (2018)): malformed author names. Please check if it should be "V.Price, K., H.Awad, N., Z.Ali, M., & N.Suganthan, P."
- Line 223 (Hansen, Nikolaus (2009)): Should be "Hansen, N."
